### 2017-08-04 #139 @raymondben
Bowerbird
A package for maintaining a local collection of data sets from a range of data providers. Bowerbird can mirror an entire remote collection of files, using `wget`'s recursive download functionality. Bowerbird also provides some functions around data provenance and versioning (it doesn't fundamentally solve these issues, but goes some way towards solutions).

```
Package: bowerbird
Title: Keep a Collection of Sparkly Data Resources
Version: 0.3.4
Authors@R: c(person("Ben","Raymond",email="ben.raymond@aad.gov.au",
  	   role=c("aut","cre")),
	   person("Michael","Sumner",role="aut"))
Description: Tools to get and maintain a data repository from third-party data
    providers.
    assertthat,
    openssl,
    R.utils,
    rvest,
    xml2
LazyLoad: yes
    archive,
    covr
Remotes: jimhester/archive

```

https://github.com/AustralianAntarcticDivision/bowerbird

Data retrieval, but also venturing into reproducibility. Primarily it is intended as a mechanism for maintaining a local data collection (from remote data providers), but could also be used as a wrapper to allow others to reproduce your work (e.g. "you'll need these 100GB of files installed locally; here's the bowerbird script to do so").  Bowerbird also has a few functions to help with data provenance, see `vignette("data_provenance")`

Research scientists/technicians/data managers who want to maintain a local library of data files (either for their own use, or perhaps a single shared library on behalf of a number of local users, as we do). Researchers who want to share work that relies on local copies of data. Also potentially package developers who need some sort of data retrieval that isn't easily accomplished by existing tools (e.g. recursive download of a whole collection of data files from a satellite data provider).

Nothing to our knowledge that really does the same thing. Some similarity to https://github.com/ropensci/rdataretriever, though rdataretriever seems to be angled towards biodiversity data sets in particular and creating sensible local database structures for them. Bowerbird is focused on mirroring remote data to a local file system, and providing some functions around data provenance. Passing overlap with http packages (httr, crul) but these are generally intended for single-transaction sort of usage. Jeroen's `curl` package (not an ropensci one?) is also similar to bowerbird in that it wraps a underlying http client: bowerbird typically uses `wget` under the hood to accomplish its web traffic, whereas curl binds to libcurl. AFAIK curl doesn't support mirroring of external sites (which `wget` does, and which bowerbird relies heavily on).

- [x] includes documentation with examples for all functions. **More needed, but the important ones are covered for now. The more-sparsely-documented functions at this stage are the ones that the average user is unlikely to need to interact with directly**.
- [x] has continuous integration, including reporting of test coverage, using services such as Travis CI, Coeveralls and/or CodeCov. **More test coverage still to be added - codecov is currently reporting around 80%, goodpractice around 70%, not sure which is right!**

- [no] Do you wish to automatically submit to the [Journal of Open Source Software](http://joss.theoj.org/)? If so:

- [more or less] Does the package conform to [rOpenSci packaging guidelines](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md)? Please describe any exceptions:
    - no NEWS file yet
    - some function docs need expanding
    - we use `cat()` for printing progress, despite the packaging guide suggestions to use `message` instead. This is because (a) progress information doesn't really strike me as a "condition", which `message` is intended for, (b) all `cat`-issues messages can be turned off by specifying `bb_sync(...,verbose=FALSE)`, (c) an anticipated common use for bowerbird is for unattended (cron-job) updates to a local data library, in which case it's likely the user will want to `sink()` all output to a log file. Using `cat()` means that a simple `sink()` will catch everything, including output from `wget` calls (if they are made). I think this becomes less reliable if `message` is used (you'd have to `sink(...type="message")` but then I'm not sure it'd catch `wget` output, but admittedly haven't tried this)

## Additional notes re: our presubmission enquiry

General note: the package is not in a final polished state, but we think far enough advanced (and stable enough) to be a good point for onboarding consideration.

> It would be helpful to actually separate out the core mechanism and additional sources. This could go as far as having separate packages (which we could handle together).

Fair suggestion, and one that we've considered - and maybe that split is still reasonable to consider down the track. But for now, at least, we think it's better to keep core-functionality and the data-source-definitions bundled together.

> Have you considered using rappdirs for default data directories?

It's up to the user where they want to put their data. We do make a suggestion in the README and vignette for users to consider rappdirs.
