### 2017-08-04 #139 @raymondben
Minor comment on the goodpractice output, which I meant to include in the initial submission: `setwd` is necessary, but all calls are protected by `on.exit()`, which also takes care of other global changes where necessary (e.g. to `http_proxy` env settings).
