### 2017-08-16 #139 @MilesMcBain
With the Editor's permission, I am posting some feedback ahead of a full review in the interests of discussing potential structural changes as early as possible.

The issue of separating the core functionality from data sources was raised by the editors and the more I poke around under the hood, the more I am convinced this will also be my recommendation.

Part of my motivation is selfish. The number of sources `bowerbird` touches means a comprehensive review of this package, as is, will be an onerous task. Some other reasons I can think of why this might be a good idea:

* Discoverability: For many sources there is significant source specific domain knowledge contained within the handler functions or choice of `method_flags`. Will those in niche areas who could most benefit from this be able to find it in a general package about synchronisation?

* Package Maintenance: It seems like ensuring calls to these different sources and APIs are working as advertised could become a maintenance issue that leads to frequent package updates. Have remote server changes generated much work to date?

* Nicer Data Interfaces: Assuming data sources get moved to new packages that depend on `bowebird`, there may then be scope to tweak the APIs of those packages so they more closely suit their user's needs. By this I mean implement API specific function arguments that mean users do not need to understand the workings of `wget` arguments. An example would be setting a date range as function arguments, instead of directly as `--reject-regex` arguments in the case of the README download size example.
