### 2017-08-18 f1e17d1
remove most data sources, provide a few examples in bb_example_sources
