+++ man/bb_source.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/sources.R
+% Please edit documentation in R/source.R
 \name{bb_source}
 \alias{bb_source}
-\title{Define an external data source}
+\title{Define a data source}
 \usage{
 bb_source(id, name, description = NA_character_, reference, source_url,
   citation, license, comment = NA_character_, method = bb_wget,
@@ -13,23 +13,23 @@ bb_source(id, name, description = NA_character_, reference, source_url,
   warn_empty_auth = TRUE)
 }
 \arguments{
-\item{id}{string: a unique identifier of the data source. If the data source has a DOI, use that. Otherwise, if the original data provider has an identifier for this dataset, that is probably a good choice here (include the data version number if there is one). The ID should be something that changes when the data set changes (is updated). A DOI is ideal for this}
+\item{id}{string: (required) a unique identifier of the data source. If the data source has a DOI, use that. Otherwise, if the original data provider has an identifier for this dataset, that is probably a good choice here (include the data version number if there is one). The ID should be something that changes when the data set changes (is updated). A DOI is ideal for this}
 
-\item{name}{string: a unique name for the data source. This should be a human-readable but still concise name}
+\item{name}{string: (required) a unique name for the data source. This should be a human-readable but still concise name}
 
 \item{description}{string: a description of the data source}
 
-\item{reference}{string: URL to the metadata record or home page of the data source}
+\item{reference}{string: (required) URL to the metadata record or home page of the data source}
 
-\item{source_url}{character vector: one or more source URLs}
+\item{source_url}{character vector: one or more source URLs. Generally required, although some \code{method} functions might not require one}
 
-\item{citation}{string: details of the citation for the data source}
+\item{citation}{string: (required) details of the citation for the data source}
 
-\item{license}{string: description of the license. For standard licenses (e.g. creative commons) include the license descriptor ("CC-BY", etc)}
+\item{license}{string: (required) description of the license. For standard licenses (e.g. creative commons) include the license descriptor ("CC-BY", etc)}
 
 \item{comment}{string: comments about the data source. If only part of the original data collection is mirrored, mention that here}
 
-\item{method}{function, call, or symbol: the function that handles the synchronisation process for this data source}
+\item{method}{function, call, or symbol: (required) the function that handles the synchronisation process for this data source}
 
 \item{method_flags}{string: flags to pass to the method. If method is \code{bb_wget}, these are wget flags. Run \code{wget("--help")} to get help on these flags}
 
@@ -50,10 +50,10 @@ bb_source(id, name, description = NA_character_, reference, source_url,
 \item{warn_empty_auth}{logical: if TRUE, issue a warning if the data source requires authentication (authentication_note is not NA) but user and password have not been provided}
 }
 \value{
-tibble
+data.frame
 }
 \description{
-Define an external data source
+Define a data source
 }
 \examples{
 
@@ -70,7 +70,7 @@ my_source <- bb_source(
    method=quote(bb_wget),
    method_flags="--recursive --level=1 --accept=\\"*bin*.zip,README.TXT\\"",
    postprocess=quote(pp_unzip),
-   collection_size=620)
+   collection_size=0.6)
 
 cf <- bb_config("/my/repo/root")
 cf <- add(cf,my_source)
