+++ NAMESPACE
@@ -8,9 +8,9 @@ export(add)
 export(amps_get)
 export(bb_attributes)
 export(bb_config)
+export(bb_example_sources)
 export(bb_fingerprint)
 export(bb_source)
-export(bb_sources)
 export(bb_summary)
 export(bb_sync)
 export(bb_validate)
