### 2017-08-18 0c5007f
remove specialist data source handlers
