+++ R/bowerbird.R
@@ -9,8 +9,6 @@
 #' @importFrom stringr regex str_detect str_match str_split str_trim
 #' @importFrom R.utils bunzip2 gunzip
 #' @importFrom rmarkdown render
-#' @importFrom rvest html_session jump_to html_attr html_nodes
 #' @importFrom stats na.omit
 #' @importFrom utils download.file read.table str unzip
-#' @importFrom xml2 url_absolute
 NULL
