+++ man/rapid_response_get.Rd
@@ -1,24 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rapid_reponse_handler.R
-\name{rapid_response_get}
-\alias{rapid_response_get}
-\title{Handler for NASA MODIS Rapid Response files}
-\usage{
-rapid_response_get(data_source, verbose = FALSE, local_dir_only = FALSE)
-}
-\arguments{
-\item{data_source}{data.frame: single-row data.frame defining a data source, e.g. as returned by \code{bb_source}}
-
-\item{verbose}{logical: if TRUE, provide additional progress output}
-
-\item{local_dir_only}{logical: if TRUE, just return the local directory into which files from this data source would be saved}
-}
-\value{
-the directory if local_dir_only is TRUE, otherwise TRUE on success
-}
-\description{
-Handler for NASA MODIS Rapid Response files
-}
-\references{
-https://earthdata.nasa.gov/earth-observation-data/near-real-time/rapid-response
-}
