### 2017-08-18 26b6734
adjust tests for stripped-down package
