+++ tests/testthat/test_sources.R
@@ -1,16 +1,11 @@
 context("data sources")
 
 test_that("predefined sources work", {
-    src <- bb_sources(c("NSIDC passive microwave supporting files"))
+    src <- bb_example_sources()
+    expect_gt(nrow(src),0)
+    src <- src[1,]
     expect_s3_class(src,"data.frame")
     expect_equal(nrow(src),1)
-
-    src_all <- bb_sources()
-    expect_gt(nrow(src_all),0)
-    src_si <- bb_sources(data_group="Sea ice")
-    expect_gt(nrow(src_si),0)
-    expect_lt(nrow(src_si),nrow(src_all))
-    expect_true(all(src_si$data_group=="Sea ice"))
 })
 
 test_that("bb_source works with multiple postprocess actions", {
@@ -46,7 +41,7 @@ test_that("bb_source works with multiple postprocess actions", {
 })
 
 test_that("sources with authentication have an authentication_note entry", {
-    src <- bb_sources()
+    src <- bb_example_sources()
     idx <- (!is.na(src$user) | !is.na(src$password)) & na_or_empty(src$authentication_note)
     expect_false(any(idx),sprintf("%d data sources with non-NA authentication but no authentication_note entry",sum(idx)))
 })
@@ -114,83 +109,5 @@ test_that("authentication checks work",{
         data_group="blah")
 })
 
-##test_that("an earthdata source works",{
-##    ## earthdata test
-##    temp_root <- tempdir()
-##    cf <- bb_config(local_file_root=temp_root) %>%
-##        add(bb_sources(name="NSIDC SMMR-SSM/I Nasateam sea ice concentration") %>%
-##            mutate(user="benraymond",
-##                   password="H(~7FkC:(U9GU9ffjC.Y$@hZo",
-##                   source_url="https://daacdata.apps.nsidc.org/pub/DATASETS/nsidc0051_gsfc_nasateam_seaice/",##final-gsfc/south/daily/1978/",
-##                   method=list(quote(earthdata_get)),
-##                   method_flags="--recursive --level=inf --no-parent"))
-##    bb_sync(cf)
-##})
-
-test_that("source nsidc0051 still works under ftp (due to be moved to https)",{
-    skip_on_cran()
-    temp_root <- tempdir()
-    cf <- bb_config(local_file_root=temp_root) %>%
-        add(bb_sources(name="NSIDC SMMR-SSM/I Nasateam sea ice concentration") %>%
-            mutate(source_url="ftp://sidads.colorado.edu/pub/DATASETS/nsidc0051_gsfc_nasateam_seaice/final-gsfc/south/daily/1978/nt_19781231_n07_v1.1_s.bin"))
-    bb_sync(cf)
-
-    fnm <- file.path(temp_root,"sidads.colorado.edu/pub/DATASETS/nsidc0051_gsfc_nasateam_seaice/final-gsfc/south/daily/1978/nt_19781231_n07_v1.1_s.bin")
-    expect_true(file.exists(fnm))
-    fi <- file.info(fnm)
-    expect_gt(fi$size,50e3)
-})
-
-test_that("source nsidc0081 still works under ftp (due to be moved to https)",{
-    skip_on_cran()
-    temp_root <- tempdir()
-    target_file <- format(Sys.Date()-10,"nt_%Y%m%d_f18_nrt_s.bin")
-    cf <- bb_config(local_file_root=temp_root) %>%
-        add(bb_sources(name="NSIDC SMMR-SSM/I Nasateam near-real-time sea ice concentration") %>%
[…]
