+++ tests/testthat/test_utils.R
@@ -1,11 +1,11 @@
 context("test bowerbird utilities")
 
 test_that("check_method_is works",{
-    expect_true(check_method_is("aadc_eds_get",aadc_eds_get)) ## function name as string
-    expect_true(check_method_is(aadc_eds_get,aadc_eds_get)) ## function
-    expect_true(check_method_is(enquote(aadc_eds_get),aadc_eds_get))
-    expect_true(check_method_is(quote(aadc_eds_get()),aadc_eds_get))
-    expect_true(check_method_is(quote(aadc_eds_get),aadc_eds_get)) ## symbol
+    expect_true(check_method_is("bb_wget",bb_wget)) ## function name as string
+    expect_true(check_method_is(bb_wget,bb_wget)) ## function
+    expect_true(check_method_is(enquote(bb_wget),bb_wget))
+    expect_true(check_method_is(quote(bb_wget()),bb_wget))
+    expect_true(check_method_is(quote(bb_wget),bb_wget)) ## symbol
 })
 
 test_that("method dispatch code works",{
@@ -18,7 +18,7 @@ test_that("method dispatch code works",{
     expect_output(do.call(mth,list(data_source=myarg)),myarg)
     expect_output(do.call(mth,list(myarg)),myarg)
     expect_error(do.call(mth,list(blah=myarg)))
-    
+
     ## as symbol
     mth <- quote(mycat)
     expect_true(is.name(mth))
@@ -38,7 +38,7 @@ test_that("method dispatch code works",{
     ## test arg injection nonsense
     expect_identical(inject_args(quote(cat(z=3)),list(a=0,b=1)),list(a=0,b=1,3))
     expect_identical(inject_args(quote(cat(z=3)),list(a=0,b=1),extras_first=FALSE),list(3,a=0,b=1))
-    
+
     ## call constructed as e.g. quote(fun())
     mth <- quote(mycat())
     expect_true(is.call(mth))
@@ -49,7 +49,7 @@ test_that("method dispatch code works",{
     expect_output(do.call(all.names(mth)[1],list(myarg)),myarg)
     ## inject named arg alongside the ones already provided
     arglist <- inject_args(mth,list(data_source=myarg))
-    expect_output(do.call(all.names(mth)[1],arglist),myarg)    
+    expect_output(do.call(all.names(mth)[1],arglist),myarg)
     ## can have extra args in this form
     mth <- quote(mycat("last_var"))
     expect_true(is.call(mth))
@@ -58,10 +58,10 @@ test_that("method dispatch code works",{
     ## call with no args
     expect_error(expect_output(do.call(all.names(mth)[1],list()),"argument \"data_source\" is missing"))
     ## call with one arg (ignore the already-provided one)
-    expect_output(do.call(all.names(mth)[1],list(data_source=myarg)),myarg)    
+    expect_output(do.call(all.names(mth)[1],list(data_source=myarg)),myarg)
     ## inject named arg alongside the ones already provided
     arglist <- inject_args(mth,list(data_source=myarg))
     expect_output(do.call(all.names(mth)[1],arglist),paste(myarg,"last_var"))
 })
 
-    
+
