### 2017-08-18 5e6498c
first-cut adjustment of vignettes and README for stripped-down package
