+++ README.md
@@ -188,48 +188,6 @@ Approximate size: 310 GB
 
 Reference: <http://cmems-resources.cls.fr/?option=com_csw&view=details&tab=info&product_id=SEALEVEL_GLO_PHY_L4_REP_OBSERVATIONS_008_047>
 
-#### CMEMS global gridded SSH near-real-time
-
-For the Global Ocean - Multimission altimeter satellite gridded sea surface heights and derived variables computed with respect to a twenty-year mean. Previously distributed by Aviso+, no change in the scientific content. All the missions are homogenized with respect to a reference mission which is currently Jason-3. The acquisition of various altimeter data is a few days at most. VARIABLES
-
--   sea\_surface\_height\_above\_sea\_level (SSH)
-
--   surface\_geostrophic\_eastward\_sea\_water\_velocity\_assuming\_sea\_level\_for\_geoid (UVG)
-
--   surface\_geostrophic\_northward\_sea\_water\_velocity\_assuming\_sea\_level\_for\_geoid (UVG)
-
--   sea\_surface\_height\_above\_geoid (SSH)
-
--   surface\_geostrophic\_eastward\_sea\_water\_velocity (UVG)
-
--   surface\_geostrophic\_northward\_sea\_water\_velocity (UVG)
-
-Authentication note: Copernicus Marine login required, see <http://marine.copernicus.eu/services-portfolio/register-now/>
-
-Approximate size: 3 GB
-
-Reference: <http://cmems-resources.cls.fr/?option=com_csw&view=details&tab=info&product_id=SEALEVEL_GLO_PHY_L4_NRT_OBSERVATIONS_008_046>
-
-#### CNES-CLS09 Mean Dynamic Topography
-
-CNES-CLS09 Mean Dynamic Topography (v1.1 release)
-
-Authentication note: AVISO login required, see <https://www.aviso.altimetry.fr/en/data/data-access/endatadata-accessregistration-form.html>
-
-Approximate size: 0.1 GB
-
-Reference: <http://www.aviso.altimetry.fr/en/data/products/auxiliary-products/mdt.html>
-
-### Data group: Meteorological
-
-#### Antarctic Mesoscale Prediction System grib files
-
-The Antarctic Mesoscale Prediction System - AMPS - is an experimental, real-time numerical weather prediction capability that provides support for the United States Antarctic Program, Antarctic science, and international Antarctic efforts.
-
-Approximate size: not specified
-
-Reference: <http://www2.mmm.ucar.edu/rt/amps/>
-
 ### Data group: Ocean colour
 
 #### Oceandata SeaWiFS Level-3 mapped monthly 9km chl-a
@@ -240,165 +198,7 @@ Approximate size: 7.2 GB
 
 Reference: <http://oceancolor.gsfc.nasa.gov/>
 
-#### Oceandata MODIS Aqua Level-3 mapped daily 4km chl-a
-
-Daily remote-sensing chlorophyll-a from the MODIS Aqua satellite at 4km spatial resolution
-
-Approximate size: 40 GB
-
-Reference: <http://oceancolor.gsfc.nasa.gov/>
-
-#### Oceandata MODIS Aqua Level-3 mapped monthly 9km chl-a
-
-Monthly remote-sensing chlorophyll-a from the MODIS Aqua satellite at 9km spatial resolution
-
-Approximate size: 8 GB
-
-Reference: <http://oceancolor.gsfc.nasa.gov/>
-
-#### Oceandata VIIRS Level-3 mapped daily 4km chl-a
-
-Daily remote-sensing chlorophyll-a from the VIIRS satellite at 4km spatial resolution
-
-Approximate size: 1 GB
-
[…]
