+++ R/oceandata.R
@@ -1,14 +1,14 @@
 #' Handler for oceandata data sources
 #'
 #' @references https://oceandata.sci.gsfc.nasa.gov/
-#' @param data_source data.frame: single-row data.frame defining a data source, e.g. as returned by \code{bb_source}
+#' @param cfrow data.frame: a single row from a bowerbird configuration (as returned by \code{bb_config})
 #' @param verbose logical: if TRUE, provide additional progress output
 #' @param local_dir_only logical: if TRUE, just return the local directory into which files from this data source would be saved
 #'
 #' @return the directory if local_dir_only is TRUE, otherwise TRUE on success
 #'
 #' @export
-oceandata_get <- function(data_source,verbose=FALSE,local_dir_only=FALSE) {
+oceandata_get <- function(cfrow,verbose=FALSE,local_dir_only=FALSE) {
     ## oceandata synchronisation handler
 
     ## oceandata provides a file search interface, e.g.:
@@ -18,13 +18,13 @@ oceandata_get <- function(data_source,verbose=FALSE,local_dir_only=FALSE) {
     ## returns list of files and SHA1 checksum for each file
     ## each file can be retrieved from https://oceandata.sci.gsfc.nasa.gov/cgi/getfile/filename
 
-    ## expect that data_source$method_flags will contain the search and dtype components of the post string
+    ## expect that cfrow$method_flags will contain the search and dtype components of the post string
     ##  i.e. "search=...&dtype=..." in "dtype=L3m&addurl=1&results_as_file=1&search=A2002*DAY_CHL_chlor*9km*"
     ##  or just include the data type in the search pattern e.g. "search=A2002*L3m_DAY_CHL_chlor*9km*
 
-    assert_that(is.data.frame(data_source))
-    assert_that(nrow(data_source)==1)
-    assert_that(is.string(data_source$method_flags))
+    assert_that(is.data.frame(cfrow))
+    assert_that(nrow(cfrow)==1)
+    assert_that(is.string(cfrow$method_flags))
     assert_that(is.flag(verbose))
     assert_that(is.flag(local_dir_only))
 
@@ -32,7 +32,7 @@ oceandata_get <- function(data_source,verbose=FALSE,local_dir_only=FALSE) {
         ## highest-level dir
         out <- "oceandata.sci.gsfc.nasa.gov"
         ## refine by platform
-        this_search_spec <- sub("search=","",data_source$method_flags)
+        this_search_spec <- sub("search=","",cfrow$method_flags)
         this_platform <- oceandata_platform_map(substr(this_search_spec,1,1))
         if (nchar(this_platform)>0) out <- file.path(out,this_platform)
         if (grepl("L3m",this_search_spec)) {
@@ -40,18 +40,18 @@ oceandata_get <- function(data_source,verbose=FALSE,local_dir_only=FALSE) {
         } else if (grepl("L3",this_search_spec)) {
             out <- file.path(out,"L3BIN")
         }
-        return(file.path(bb_attributes(data_source)$local_file_root,out))
+        return(file.path(bb_attributes(cfrow)$local_file_root,out))
     }
     tries <- 0
     while (tries<3) {
         ## sometimes this takes a couple of attempts!
-        myfiles <- wget("https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi",paste0("-q --post-data=\"cksum=1&",data_source$method_flags,"\" -O -"),stdout=TRUE)
+        myfiles <- wget("https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi",paste0("-q --post-data=\"cksum=1&",cfrow$method_flags,"\" -O -"),stdout=TRUE)
         if (is.null(attr(myfiles,"status")) || length(myfiles)>0) break
         tries <- tries+1
     }
[…]
