### 2017-08-21 dd9f81c
update readme and vignette
