+++ README.Rmd
@@ -35,80 +35,13 @@ On Windows you can use the `install_wget()` function to install it. Otherwise do
 
 ### Configuration
 
-Bowerbird must be configured to tell it which data sets to synchronise and where to save them on the local file system.
-
 Build up a configuration by first defining global options such as the destination on your local file system:
 
 ```{r eval=FALSE}
 cf <- bb_config(local_file_root="~/your/data/directory")
 ```
 
-Add data sources by choosing from those already defined in Bowerbird, or define your own (see below for guidance). Bowerbird comes with configurations for a range of data sources, mostly environmental (marine and Antarctic) in nature. A summary of the pre-configured sources is given at the end of this document.
-
-```{r eval=FALSE}
-cf <- cf %>% add(bb_sources("CERSAT SSM/I sea ice concentration"))
-```
-
-### Synchronisation
-
-Once the configuration has been defined, run the sync process:
-
-```{r eval=FALSE}
-bb_sync(cf)
-```
-
-Congratulations! You now have your own local copy of your chosen data sets.
-
-The pre-packaged environmental data sources can be read, manipulated, and plotted using a range of other R packages, including [RAADTools](https://github.com/AustralianAntarcticDivision/raadtools) and [raster](https://cran.r-project.org/package=raster).
-
-
-## Nuances
-
-### Choosing a data directory
-
-It's up to you where you want your data collection kept, and to provide that location to bowerbird. A common use case for bowerbird is maintaining a central data collection for multiple users, in which case that location is likely to be some sort of networked file share. However, if you are keeping a collection for your own use, you might like to look at https://github.com/r-lib/rappdirs to help find a suitable directory location.
-
-### Modifying data sources
-
-#### Authentication
-
-Some data providers require users to log in. These are indicated by the `authentication_note` column in the configuration table. For these sources, you will need to provide your user name and password, e.g.:
-
-```{r eval=FALSE}
-src <- bb_sources(name="CMEMS global gridded SSH reprocessed (1993-ongoing)")
-src$user <- "yourusername"
-src$password <- "yourpassword"
-cf <- add(cf,src)
-
-## or, using dplyr
-cf <- cf %>% add(bb_sources(name="CMEMS global gridded SSH reprocessed (1993-ongoing)") %>%
-                 mutate(user="yourusername",password="yourpassword"))
-```
-
-#### Reducing download sizes
-
-Sometimes you might only want part of a pre-configured data source. If the data source uses the `bb_wget` method, you can restrict what is downloaded by modifying the data source's `method_flags`, particularly the `--accept`, `--reject`, `--accept-regex`, and `--reject-regex` options. Be sure to leave the original method flags in place, unless you know what you are doing.
-
-For example, the CERSAT SSM/I sea ice concentration data are arranged in yearly directories, so it is fairly easy to restrict ourselves to, say, only the 2017 data:
-
-```{r eval=FALSE}
-cf <- cf %>% add(bb_sources("CERSAT SSM/I sea ice concentration") %>%
-                 mutate(method_flags=paste(method_flags,"--accept-regex=\"/2017/\"")))
-```
-
-See the notes below for further guidances on the accept/reject flags.
-
-Alternatively, for data sources that are divided into subdirectories, one could replace the whole-data-source `source_url` with one or more that point to specific yearly (or other) subdirectories. For example, the default `source_url` for the CERSAT sea ice data above is "ftp://ftp.ifremer.fr/ifremer/cersat/products/gridded/psi-concentration/data/antarctic/daily/netcdf/*" (with yearly subdirectories). So e.g. for 2016 and 2017 data we could do:
-
[…]
