### 2017-08-22 6af9172
add reference to blueant package [ci skip]
