### 2017-08-23 3100477
suppress wget output if verbose=FALSE
