+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: bowerbird
 Type: Package
 Title: Keep a Collection of Sparkly Data Resources
-Version: 0.3.4
+Version: 0.4.0
 Authors@R: c(person("Ben","Raymond",email="ben.raymond@aad.gov.au",
   	   role=c("aut","cre")),
 	   person("Michael","Sumner",role="aut"))
