### 2017-08-24 #139 @MilesMcBain
Thanks for your response @raymondben.

Firstly, one small point about the maintenance: With the split design you get much more freedom. You may not even put all your datasources on CRAN, you can choose to leave some on GitHub where the expectations about maintenance are much lower. In this way the maintenance burden could be reduced.

Secondly, I did get some of the way through my review before we put this on hold, so I thought it might be helpful to have the higher level feedback now while things are being actively worked on:

## Following bowerbird.Rmd

```{r eval=FALSE}
cf <- bb_config(local_file_root="~/data")
cf <- cf %>% add(bb_sources("CERSAT SSM/I sea ice concentration"))
bb_sync(cf)
```

This kicked off a 2.5Gb download that blocked my R session which I had to eventually kill. Not a great first user experience ;). I know down the page you point out download size for this source can be reduced. But the total size of the source is not mentioned until the source list at the bottom.

I would like to make the following recommendations:
* The example should be a small data source that can be downloaded in a minute or two.
* If the intent is for syncs to be able to be made interactively, there needs to be some indication of progress. A message like "Fetching File n/M (xMb), yMb Total remaining" or a progress bar.
* If this package is intended to be used exclusively from the command line that should really be how the vignette directs users. Consider including this.

A question: is there a way to check sync status in interactive mode without fecthing these large repos?

## Programming model

`bb_config` returns a sync configuration object. This would be in line with my expectations of API package usage. A very similar pattern is in the usage section of the [ropensci/osmdata](https://github.com/ropensci/osmdata) README.

Unlike `osmdata::opq`, `bb_config` seems unusual in that returns its configuration as attriubtes of an empty Tibble which is not mentioned in the documentation. It means to inspect configuration one needs to use `attributes(cf)` instead of say `cf$` and using IDE autcomplete features in inspect object structure. I would argue this is unsual enough that it should feature promenantly in the documentation. Users need to be warned about applying operations to the Tibble that obliterate the attributes - very easy to do in the `tidyverse`.

I would also make the comment that I do not think this is the idiomatic `tidyverse` way of approaching things. I only mention this since it seems you use many `tidyverse` funcitons. I think the `tidyverse` way would be to use a list-column to capture this information. Take a look at how [tidyverse/googledrive](https://github.com/tidyverse/googledrive) stores file metadata in the `dribble()` class for an example.

Finally, the `add()` function would need to have a 'bb_' prefix added to ahere to function naming guidelines.
