### 2017-08-24 #139 @raymondben
Thanks @MilesMcBain - a quick question on the config arrangement (and, yes, it is unfortunately easy to unknowingly lose the attributes). The reason for doing it that way is because those attributes apply to the whole configuration (and thus should only be stored once, not in every row) whereas the rows hold data-source-specific information that is not necessarily common across rows. An alternative structure that we considered would be to create a list object that has something like a `data_sources` tibble inside it (i.e. the existing bb_config rows) and a `config` tibble (holding what are now stored as attributes). This is pretty close to what `opq` does - would that seem better to you? I can't remember now why we didn't go that way. Probably some arbitrary decision of mine!
