+++ tests/testthat/test_wget.R
@@ -11,9 +11,8 @@ test_that("wget_install works and wget_exe() finds something", {
 })
 
 test_that("wget help works", {
-    blah <- wget("--help",stdout=TRUE)
-    expect_gt(length(blah),10)
-    blah <- paste(blah,collapse="\n")
+    blah <- wget("--help")
+    blah <- rawToChar(blah$stdout)
     ## should see the phrase "Mail bug reports and suggestions to <bug-wget@gnu.org>" in the output
     expect_true(grepl("bug reports and suggestions",blah,ignore.case=TRUE))
-})    
+})
