### 2017-08-24 6968a77
pass return value when catching sync errors
