### 2017-08-24 d52aa5d
changes to wget for sys compatibility [ci skip]
