### 2017-08-24 fa16482
only split args if provided as string
