### 2017-08-24 dc00a29
tidier verbose output
