### 2017-08-25 086f417
more consistent handling of return value from bb_wget
