### 2017-08-25 627eeba
wget flags as character vector not string
