### 2017-08-25 4ca8dd9
test output captured by sink
