### 2017-08-25 53e740a
fix bug in wget, small doc changes
