### 2017-08-25 97b44fe
ensure return status are of type logical
