### 2017-08-25 dfff637
catch status returned by sync handler
