### 2017-08-25 4b1395a
reenable oceandata test
