### 2017-08-25 bcf1922
deal with flags as char vector in bb_wget [ci skip]
