+++ R/wget.R
@@ -1,3 +1,29 @@
+## internal: take flags and convert to character vector
+## flags can either be a string (will be split on white spaces)
+## or list (expect single element, which is a character vector) - as would get from tibble list column
+## or character vector (convert to empty character vector if null, NA, or empty string, otherwise return as is)
+flags_to_charvec <- function(fl) {
+    if (is.list(fl)) {
+        fl[[1]]
+    } else if (is.null(fl) || length(fl)<1) {
+        character()
+    } else if (is.character(fl)) {
+        if (is.string(fl)) {
+            if (is.null(fl) || is.na(fl) || !nzchar(fl)) {
+                character()
+            } else {
+                ## split on whitespaces
+                strsplit(fl,"[[:space:]]+")[[1]]
+            }
+        } else {
+            fl
+        }
+    } else {
+        stop("expecting flags as character vector or list")
+    }
+}
+
+
 #' Mirror an external data source using the wget utility
 #'
 #' @param cfrow data.frame: a single row from a bowerbird configuration (as returned by \code{bb_config})
@@ -18,9 +44,17 @@ bb_wget <- function(cfrow,verbose=FALSE,local_dir_only=FALSE) {
     if (local_dir_only)
         return(file.path(bb_attributes(cfrow)$local_file_root,directory_from_url(cfrow$source_url)))
 
-    this_flags <- if (is.na(cfrow$method_flags)) cfrow$wget_default_flags else cfrow$method_flags
+    this_flags <- flags_to_charvec(cfrow$method_flags)
+    if (length(this_flags)<1) {
+        this_flags <- flags_to_charvec(cfrow$wget_default_flags)
+    }
+    ##this_flags <- if (is.na(cfrow$method_flags)) cfrow$wget_default_flags else cfrow$method_flags
     ## add wget_global_flags
-    if (!is.null(cfrow$wget_global_flags)) this_flags <- paste(this_flags,cfrow$wget_global_flags,sep=" ")
+    ##if (!is.null(cfrow$wget_global_flags)) this_flags <- paste(this_flags,cfrow$wget_global_flags,sep=" ")
+    gflags <- flags_to_charvec(cfrow$wget_global_flags)
+    if (length(gflags)>0)
+        this_flags <- c(this_flags,gflags)
+
     ## proxy-user and proxy-password flags
     ## this needs to decide whether it should use http_proxy_user or ftp_proxy_user info - exclude for now
     #if (!grepl("proxy-user=",cfrow$wget_flags)) {
@@ -38,17 +72,20 @@ bb_wget <- function(cfrow,verbose=FALSE,local_dir_only=FALSE) {
         }
     }
     ## add user, password flags
-    if (!is.na(cfrow$user) && nchar(cfrow$user)>0) this_flags <- paste0(this_flags," --user='",cfrow$user,"'")
-    if (!is.na(cfrow$password) && nchar(cfrow$password)>0) this_flags <- paste0(this_flags," --password='",cfrow$password,"'")
+    ##if (!is.na(cfrow$user) && nchar(cfrow$user)>0) this_flags <- paste0(this_flags," --user='",cfrow$user,"'")
+    if (!is.na(cfrow$user) && nchar(cfrow$user)>0) this_flags <- c(this_flags,"--user",cfrow$user)
+    ##if (!is.na(cfrow$password) && nchar(cfrow$password)>0) this_flags <- paste0(this_flags," --password='",cfrow$password,"'")
+    if (!is.na(cfrow$password) && nchar(cfrow$password)>0) this_flags <- c(this_flags,"--password",cfrow$password)
     ##if (cfrow$wait>0) this_flags <- paste0(this_flags," --wait=",cfrow$wait)
 
[…]
