### 2017-08-25 d05ccad
earthdata handler pass flags as list to bb_wget [ci skip]
