### 2017-08-25 ab59339
fiddle with flag ordering in earthdata_get [ci skip]
