### 2017-08-25 eb4e2fc
wget_install to return path of installed exe
