### 2017-08-25 09152f4
fix 'quiet' test to cope with flags as char vector not string
