### 2017-08-25 22e19dc
provide flags as char vector to bb_source(), not as string
