+++ R/example_sources.R
@@ -32,7 +32,7 @@ bb_example_sources <- function() {
             source_url="ftp://ftp.cdc.noaa.gov/Datasets/noaa.oisst.v2/*",
             license="Please cite",
             method=quote(bb_wget),
-            method_flags="--recursive --level=1 --no-parent",
+            method_flags=c("--recursive","--level=1","--no-parent"),
             postprocess=NULL,
             access_function="readsst",
             collection_size=0.9,
@@ -46,7 +46,7 @@ bb_example_sources <- function() {
             source_url=c("ftp://ftp.sltac.cls.fr/Core/SEALEVEL_GLO_PHY_L4_REP_OBSERVATIONS_008_047/dataset-duacs-rep-global-merged-allsat-phy-l4-v3/"),
             license="See http://marine.copernicus.eu/services-portfolio/service-commitments-and-licence/",
             method=quote(bb_wget),
-            method_flags="--recursive --level=3 --follow-ftp --no-parent",
+            method_flags=c("--recursive","--level=3","--follow-ftp","--no-parent"),
             postprocess=quote(pp_gunzip),
             authentication_note="Copernicus Marine login required, see http://marine.copernicus.eu/services-portfolio/register-now/",
             user="",
@@ -62,7 +62,7 @@ bb_example_sources <- function() {
             citation="See https://oceancolor.gsfc.nasa.gov/citations",
             license="Please cite",
             method=quote(oceandata_get),
-            method_flags="search=S*L3m_MO_CHL_chlor_a_9km.nc",
+            method_flags=c("search=S*L3m_MO_CHL_chlor_a_9km.nc"),
             postprocess=NULL,
             collection_size=7.2,
             data_group="Ocean colour"),
@@ -76,7 +76,7 @@ bb_example_sources <- function() {
             license="Please cite, see http://nsidc.org/about/use_copyright.html",
             authentication_note="Requires Earthdata login, see https://urs.earthdata.nasa.gov/",
             method=quote(earthdata_get),
-            method_flags="--recursive --level=2 --no-parent --accept-regex=\"/NmIcEdg2.001/\"",
+            method_flags=c("--recursive","--level=2","--no-parent","--accept-regex=/NmIcEdg2.001/"),
             user="",
             password="",
             postprocess=NULL,
@@ -91,7 +91,7 @@ bb_example_sources <- function() {
             citation="Publisher: DOC/NOAA/NESDIS/NCEI > National Centers for Environmental Information, NESDIS, NOAA, U.S. Department of Commerce",
             license="https://www.ngdc.noaa.gov/ngdcinfo/privacy.html#copyright",
             method=quote(bb_wget),
-            method_flags="--recursive --level=2 --accept-regex=\"/netcdf/\" --reject=\"index.html*\"",
+            method_flags=c("--recursive","--level=2","--accept-regex=/netcdf/","--reject=index.html*"),
             comment="Only the netcdf format data are retrieved here - adjust the accept parameter in the method_flags to get other formats",
             postprocess=quote(pp_gunzip),
             collection_size=0.03,
