### 2017-08-25 b5c40e0
method flags in tests as char vectors
