### 2017-08-25 91b4416
fix att-to-column conversion
