### 2017-08-25 cb15755
specify empty flags as character() not c()
