+++ tests/testthat/test_sources.R
@@ -57,7 +57,7 @@ test_that("authentication checks work",{
         license="blah",
         authentication_note="auth note",
         method=bb_wget,
-        method_flags=c(),
+        method_flags=character(),
         postprocess=NULL,
         data_group="blah"))
 
@@ -72,7 +72,7 @@ test_that("authentication checks work",{
         authentication_note="auth note",
         user="",
         method=bb_wget,
-        method_flags=c(),
+        method_flags=character(),
         postprocess=NULL,
         data_group="blah"))
 
@@ -87,7 +87,7 @@ test_that("authentication checks work",{
         authentication_note="auth note",
         password="",
         method=bb_wget,
-        method_flags=c(),
+        method_flags=character(),
         postprocess=NULL,
         data_group="blah"))
 
@@ -104,7 +104,7 @@ test_that("authentication checks work",{
         user="user",
         password="password",
         method=bb_wget,
-        method_flags=c(),
+        method_flags=character(),
         postprocess=NULL,
         data_group="blah")
 })
