### 2017-08-25 bb34277
if empty flags provided, coerce to character()
