### 2017-08-26 82a1d3e
change config object structure [ci skip]
