+++ R/source.R
@@ -82,7 +82,7 @@ bb_source <- function(id,name,description=NA_character_,reference,source_url,cit
         license=if (assert_that(is.string(license))) license,
         comment=if (assert_that(is.string(comment))) comment,
         method=list(method),
-        method_flags=if (assert_that(is.character(method_flags))) list(method_flags),
+        method_flags=if (assert_that(is.character(method_flags))) list(str_trim(method_flags)),
         postprocess=list(postprocess),
         authentication_note=if (assert_that(is.string(authentication_note))) authentication_note,
         user=if (assert_that(is.string(user))) user,
