+++ R/config.R
@@ -14,7 +14,7 @@ bb_global_atts <- function() c("wget_default_flags","wget_global_flags","http_pr
 #' @param ftp_proxy string: URL of FTP proxy to use e.g. 'http://your.proxy:21' (NULL for no proxy)
 #' @param clobber numeric: 0=do not overwrite existing files, 1=overwrite if the remote file is newer than the local copy, 2=always overwrite existing files. For data sources that use method 'wget', an appropriate flag will be added to the wget call according to the clobber setting ("--no-clobber" to not overwrite existing files, "--timestamping" to overwrite if the remote file is newer than the local copy)
 #' @param skip_downloads logical: if TRUE, \code{bb_sync} will do a dry run of the synchronisation process but without actually downloading files. For data sources using method bb_wget, this means that the wget calls will not be executed, so e.g. any recursion handled by wget itself will not be simulated
-#' @return configuration tibble
+#' @return configuration object
 #'
 #' @seealso \code{\link{bb_source}}
 #'
@@ -36,61 +36,32 @@ bb_config <- function(local_file_root,wget_default_flags=character(),wget_global
     assert_that(is.flag(skip_downloads))
     assert_that(is.character(wget_default_flags))
     assert_that(is.character(wget_global_flags))
-    cf <- tibble()
-    attr(cf,"wget_default_flags") <- wget_default_flags
-    attr(cf,"wget_global_flags") <- wget_global_flags
-    attr(cf,"http_proxy") <- http_proxy
-    attr(cf,"ftp_proxy") <- ftp_proxy
-    attr(cf,"local_file_root") <- local_file_root
-    attr(cf,"clobber") <- clobber
-    attr(cf,"skip_downloads") <- skip_downloads
-    class(cf) <- c("bb_cf",class(cf))
+    structure(
+        list(data_sources=tibble(),
+             settings=list(
+                 wget_default_flags=str_trim(wget_default_flags),
+                 wget_global_flags=str_trim(wget_global_flags),
+                 http_proxy=http_proxy,
+                 ftp_proxy=ftp_proxy,
+                 local_file_root=local_file_root,
+                 clobber=clobber,
+                 skip_downloads=skip_downloads)),
+        class="bb_config")
+}
+
+## internal helper function, return the config but with only
+## one or more rows of its data_sources
+cf_subset <- function(cf,idx) {
+    cf$data_sources <- cf$data_sources[idx,]
     cf
 }
 
-#' Select a subset of data sources within a bowerbird configuration
-#'
-#' @param .data list: configuration, as returned by \code{bb_config}
-#' @param x list: configuration, as returned by \code{bb_config}
-#' @param i numeric: integer row values
-#' @param j numeric: ignored
-#' @param ... : integer row values
-#'
-#' @return configuration
-#'
-#' @examples
-#' \dontrun{
-#'   cf <- bb_config("/my/file/root") %>%
-#'     add(bb_sources()) %>%
-#'     slice(1:5)
-#' }
-#'
-#' @method slice bb_cf
-#' @export
-slice.bb_cf <- function(.data,...) {
-    ##re_add_class <- function(z) {class(z) <- c("bb_cf",class(z)); z}
-    ##class(.data) <- setdiff(class(.data),"bb_cf")
-    ##re_add_class(copy_bb_attributes(do.call(slice_,list(.data,...)),.data))
-
-    ## hmm, that doesn't work, so do this for now
-    .data[...,]
-}
-
-#' @rdname slice.bb_cf
-#' @method [ bb_cf
-#' @export
[…]
