### 2017-08-26 a2ea223
namespacing on tests of internal functions
