### 2017-08-26 a193997
rename cf_subset to bb_subset, and export
