+++ man/bb_subset.Rd
@@ -1,26 +1,27 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/config.R
-\name{add}
-\alias{add}
-\title{Add new data sources to a bowerbird configuration}
+\name{bb_subset}
+\alias{bb_subset}
+\title{Keep only selected data_sources in a bowerbird configuration}
 \usage{
-add(config, source)
+bb_subset(config, idx)
 }
 \arguments{
 \item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config})}
 
-\item{source}{data.frame: one or more data source definitions, as returned by \code{bb_source}, to add to the configuration}
+\item{idx}{logical or numeric: index vector of data_source rows to retain}
 }
 \value{
 configuration object
 }
 \description{
-Add new data sources to a bowerbird configuration
+Keep only selected data_sources in a bowerbird configuration
 }
 \examples{
 \dontrun{
   cf <- bb_config("/my/file/root") \%>\%
-    add(bb_example_sources())
+    add(bb_example_sources()) \%>\%
+    bb_subset(1:2)
 }
 }
 \seealso{
