+++ NAMESPACE
@@ -7,6 +7,7 @@ export(bb_config)
 export(bb_example_sources)
 export(bb_fingerprint)
 export(bb_source)
+export(bb_subset)
 export(bb_summary)
 export(bb_sync)
 export(bb_validate)
