### 2017-08-26 1a79ff9
rename bb_attributes to bb_settings
