+++ R/provenance.R
@@ -36,7 +36,7 @@ do_fingerprint <- function(this_dataset,hash,verbose,settings) {
     on.exit({ restore_settings(settings) })
     if (nrow(this_dataset$data_sources)!=1) stop("expecting single-row data set")
     ## copy bb attrs into this_dataset
-    this_dataset <- bb_attributes_to_cols(this_dataset)
+    this_dataset <- bb_settings_to_cols(this_dataset)
     ## check that the root directory exists
     if (!dir_exists(this_dataset$local_file_root)) {
         ## no, it does not exist
