+++ man/bb_settings.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/config.R
-\name{bb_attributes}
-\alias{bb_attributes}
+\name{bb_settings}
+\alias{bb_settings}
 \title{Returns a bowerbird configuration object's settings}
 \usage{
-bb_attributes(config)
+bb_settings(config)
 }
 \arguments{
 \item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config})}
@@ -17,7 +17,7 @@ These are repository-wide settings that are applied to all data sources added to
 }
 \examples{
 cf <- bb_config(local_file_root="/your/data/directory")
-bb_attributes(cf)
+bb_settings(cf)
 
 }
 \seealso{
