### 2017-08-26 6f2f02b
documentation tidyup
