+++ R/config.R
@@ -16,7 +16,7 @@
 #' @examples
 #' \dontrun{
 #'   cf <- bb_config("/my/file/root") %>%
-#'     add(bb_sources("NSIDC SMMR-SSM/I Nasateam sea ice concentration"))
+#'     add(bb_example_sources())
 #'
 #'   ## save to file
 #'   saveRDS(cf,file="my_config.rds")
@@ -124,7 +124,7 @@ bb_settings_to_cols <- function(obj) {
 
 #' Produce summary of bowerbird configuration
 #'
-#' @param cf data.frame: configuration, as returned by \code{bb_config}
+#' @param config bb_config: a bowerbird configuration (as returned by \code{bb_config})
 #' @param file string: path to file to write summary to. A temporary file is used by default
 #' @param format string: produce HTML ("html") or Rmarkdown ("Rmd") file?
 #' @param inc_license logical: include each source's license and citation details?
@@ -138,13 +138,13 @@ bb_settings_to_cols <- function(obj) {
 #' @examples
 #' \dontrun{
 #'   cf <- bb_config("/my/file/root") %>%
-#'     add(bb_sources("NSIDC SMMR-SSM/I Nasateam sea ice concentration"))
+#'     add(bb_example_sources())
 #'   browseURL(bb_summary(cf))
 #' }
 #'
 #' @export
-bb_summary <- function(cf,file=tempfile(fileext=".html"),format="html",inc_license=TRUE,inc_auth=TRUE,inc_size=TRUE,inc_access_function=TRUE,inc_path=TRUE) {
-    assert_that(is(cf,"bb_config"))
+bb_summary <- function(config,file=tempfile(fileext=".html"),format="html",inc_license=TRUE,inc_auth=TRUE,inc_size=TRUE,inc_access_function=TRUE,inc_path=TRUE) {
+    assert_that(is(config,"bb_config"))
     assert_that(is.string(file))
     assert_that(is.string(format))
     assert_that(is.flag(inc_license))
@@ -161,45 +161,45 @@ bb_summary <- function(cf,file=tempfile(fileext=".html"),format="html",inc_licen
     cat("Summary of bowerbird configuration\n========\n",file=rmd_file,append=TRUE)
     cat("\nLast updated: ",format(Sys.time()),"\n",file=rmd_file,append=TRUE)
 
-    cf <- bb_settings_to_cols(cf)
+    config <- bb_settings_to_cols(config)
 
-    cf <- cf %>% group_by_(~data_group,~name) %>% mutate_(source_urls=~paste(file.path(local_file_root,directory_from_url(source_url)),collapse=", ")) %>% ungroup() %>% select_(~-source_url,~-method,~-method_flags,~-postprocess) %>% unique()
+    config <- config %>% group_by_(~data_group,~name) %>% mutate_(source_urls=~paste(file.path(local_file_root,directory_from_url(source_url)),collapse=", ")) %>% ungroup() %>% select_(~-source_url,~-method,~-method_flags,~-postprocess) %>% unique()
 
-    cf$data_group[cf$data_group==""] <- NA ## so that arrange puts them last
-    cf <- cf[order(cf$data_group), ]
-    cf$data_group[is.na(cf$data_group)] <- ""
+    config$data_group[config$data_group==""] <- NA ## so that arrange puts them last
+    config <- config[order(config$data_group), ]
+    config$data_group[is.na(config$data_group)] <- ""
 
     last_group <- "blah"
-    for (k in seq_len(nrow(cf))) {
-        if (last_group!=cf$data_group[k]) {
-            cat("\n## Data group: ",cf$data_group[k],"\n",file=rmd_file,append=TRUE)
+    for (k in seq_len(nrow(config))) {
[…]
