+++ R/utils.R
@@ -95,7 +95,7 @@ dir_exists <- function(z) file.exists(dirname(z)) && !(!file.info(z)$isdir || is
 #' @examples
 #' \dontrun{
 #'   cf <- bb_config("/my/file/root") %>%
-#'     add(bb_sources(data_group="Sea surface temperature"))
+#'     add(bb_example_sources())
 #'   data_source_dir(cf)
 #' }
 #'
