+++ man/data_source_dir.Rd
@@ -18,7 +18,7 @@ Files from each data source are stored locally in the associated directory.
 \examples{
 \dontrun{
   cf <- bb_config("/my/file/root") \%>\%
-    add(bb_sources(data_group="Sea surface temperature"))
+    add(bb_example_sources())
   data_source_dir(cf)
 }
 
