+++ man/bb_config.Rd
@@ -33,7 +33,7 @@ The parameters provided here are repository-wide settings, and will be applied t
 \examples{
 \dontrun{
   cf <- bb_config("/my/file/root") \%>\%
-    add(bb_sources("NSIDC SMMR-SSM/I Nasateam sea ice concentration"))
+    add(bb_example_sources())
 
   ## save to file
   saveRDS(cf,file="my_config.rds")
