### 2017-08-26 28e883e
adjust tests for new config structure
