+++ R/provenance.R
@@ -8,7 +8,7 @@
 #' @examples
 #' \dontrun{
 #'   cf <- bb_config("/my/file/root") %>%
-#'     add(bb_example_sources())
+#'     bb_add(bb_example_sources())
 #'   bb_fingerprint(cf)
 #' }
 #'
