+++ man/bb_add.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/config.R
-\name{add}
-\alias{add}
+\name{bb_add}
+\alias{bb_add}
 \title{Add new data sources to a bowerbird configuration}
 \usage{
-add(config, source)
+bb_add(config, source)
 }
 \arguments{
 \item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config})}
@@ -20,7 +20,7 @@ Add new data sources to a bowerbird configuration
 \examples{
 \dontrun{
   cf <- bb_config("/my/file/root") \%>\%
-    add(bb_example_sources())
+    bb_add(bb_example_sources())
 }
 }
 \seealso{
