+++ man/bb_data_source_dir.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
-\name{data_source_dir}
-\alias{data_source_dir}
+% Please edit documentation in R/config.R
+\name{bb_data_source_dir}
+\alias{bb_data_source_dir}
 \title{Return the local directory of each data source in a configuration}
 \usage{
-data_source_dir(config)
+bb_data_source_dir(config)
 }
 \arguments{
 \item{config}{bb_config: configuration as returned by \code{\link{bb_config}}}
@@ -19,7 +19,7 @@ Files from each data source are stored locally in the associated directory.
 \dontrun{
   cf <- bb_config("/my/file/root") \%>\%
     bb_add(bb_example_sources())
-  data_source_dir(cf)
+  bb_data_source_dir(cf)
 }
 
 }
