+++ NAMESPACE
@@ -3,6 +3,7 @@
 export("%>%")
 export(bb_add)
 export(bb_config)
+export(bb_data_source_dir)
 export(bb_example_sources)
 export(bb_fingerprint)
 export(bb_settings)
@@ -12,7 +13,6 @@ export(bb_summary)
 export(bb_sync)
 export(bb_validate)
 export(bb_wget)
-export(data_source_dir)
 export(earthdata_get)
 export(install_wget)
 export(oceandata_get)
