+++ R/example_sources.R
@@ -47,7 +47,7 @@ bb_example_sources <- function() {
             license="See http://marine.copernicus.eu/services-portfolio/service-commitments-and-licence/",
             method=quote(bb_wget),
             method_flags=c("--recursive","--level=3","--follow-ftp","--no-parent"),
-            postprocess=quote(pp_gunzip),
+            postprocess=quote(bb_gunzip),
             authentication_note="Copernicus Marine login required, see http://marine.copernicus.eu/services-portfolio/register-now/",
             user="",
             password="",
@@ -93,7 +93,7 @@ bb_example_sources <- function() {
             method=quote(bb_wget),
             method_flags=c("--recursive","--level=2","--accept-regex=/netcdf/","--reject=index.html*"),
             comment="Only the netcdf format data are retrieved here - adjust the accept parameter in the method_flags to get other formats",
-            postprocess=quote(pp_gunzip),
+            postprocess=quote(bb_gunzip),
             collection_size=0.03,
             data_group="Topography")
     )
