+++ man/bb_cleanup.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/postprocess.R
-\name{pp_cleanup}
-\alias{pp_cleanup}
+\name{bb_cleanup}
+\alias{bb_cleanup}
 \title{Postprocessing: remove unwanted files}
 \usage{
-pp_cleanup(config, pattern, recursive = FALSE, ignore_case = FALSE, ...)
+bb_cleanup(config, pattern, recursive = FALSE, ignore_case = FALSE, ...)
 }
 \arguments{
 \item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source}
@@ -26,10 +26,10 @@ This function is not intended to be called directly, but instead can be specifie
 \examples{
 \dontrun{
   ## remove .asc files after synchronisation
-  my_source <- bb_source(...,postprocess=quote(pp_cleanup,pattern="\\\\.asc$"))
+  my_source <- bb_source(...,postprocess=quote(bb_cleanup,pattern="\\\\.asc$"))
 }
 
 }
 \seealso{
-\code{\link{bb_source}} \code{\link{bb_config}} \code{\link{pp_decompress}}
+\code{\link{bb_source}} \code{\link{bb_config}} \code{\link{bb_decompress}}
 }

[…]
