### 2017-08-26 67abe84
change add() to bb_add() in tests
