### 2017-08-26 65b630d
rename install_wget() to bb_install_wget()
