+++ tests/testthat.R
@@ -6,7 +6,7 @@ if (.Platform$OS.type=="windows") {
     tryCatch(wget_exe(),
              error=function(e) {
                  warning("At init of testing: could not find wget executable, installing.\n")
-                 install_wget()
+                 bb_install_wget()
              })
 }
 test_check("bowerbird")
