+++ man/wget.Rd
@@ -1,33 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/wget.R
-\name{wget}
-\alias{wget}
-\title{Make a wget call}
-\usage{
-wget(url, flags = character(), verbose = FALSE, stop_on_error = FALSE)
-}
-\arguments{
-\item{url}{string: the URL to retrieve}
-
-\item{flags}{character: character vector of command-line flags to pass to wget}
-
-\item{verbose}{logical: print trace output?}
-
-\item{stop_on_error}{logical: throw an error if the exit status is non-zero?}
-}
-\value{
-the result of the system call
-}
-\description{
-The wget system call is made using the \code{exec_internal} function from the sys package.
-}
-\examples{
-\dontrun{
-## get help about wget command line parameters
-wget("--help")
-}
-
-}
-\seealso{
-\code{\link{install_wget}}
-}
