+++ man/bb_install_wget.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/wget.R
-\name{install_wget}
-\alias{install_wget}
+\name{bb_install_wget}
+\alias{bb_install_wget}
 \title{Helper function to install wget}
 \usage{
-install_wget()
+bb_install_wget()
 }
 \value{
 the path to the installed executable
@@ -14,7 +14,7 @@ Currently only works on Windows platforms. The wget.exe executable will be downl
 }
 \examples{
 \dontrun{
-  install_wget()
+  bb_install_wget()
 }
 
 }
