+++ NAMESPACE
@@ -10,6 +10,7 @@ export(bb_decompress)
 export(bb_example_sources)
 export(bb_fingerprint)
 export(bb_gunzip)
+export(bb_install_wget)
 export(bb_settings)
 export(bb_source)
 export(bb_subset)
@@ -20,7 +21,6 @@ export(bb_unzip)
 export(bb_validate)
 export(bb_wget)
 export(earthdata_get)
-export(install_wget)
 export(oceandata_get)
 export(oceandata_parameter_map)
 export(oceandata_parameters)
