### 2017-08-26 65d9a7d
more wget flags to shorten
