### 2017-08-27 0da78a3
more robust checking for '-help' parm
