### 2017-08-27 5728d3b
rename bb_wget() to bb_handler_wget()
