### 2017-08-27 a11c52b
fix handling of --help
