+++ R/wget.R
@@ -139,7 +139,8 @@ wget <- function(url,flags=character(),verbose=FALSE,stop_on_error=FALSE) {
     assert_that(is.flag(verbose))
     assert_that(is.flag(stop_on_error))
     if (tolower(url) %in% c("-h","--help") || identical(tolower(flags),"-h") || identical(tolower(flags),"--help")) {
-        sys::exec_internal(wget_exe(),args="--help",error=stop_on_error)
+        out <- sys::exec_internal(wget_exe(),args="--help",error=TRUE)
+        message(rawToChar(out$stdout))
     } else {
         ## sys expects flags as a char vector, not a string
         ##if (is.string(flags))
