### 2017-08-27 ab404ae
rename wget() to bb_wget() and export
