+++ R/oceandata.R
@@ -50,11 +50,11 @@ oceandata_get <- function(config,verbose=FALSE,local_dir_only=FALSE) {
         ## sometimes this takes a couple of attempts!
         if (FALSE) {
             ## system2 code
-            myfiles <- wget("https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi",paste0("-q --post-data=\"cksum=1&",method_flags,"\" -O -"),stdout=TRUE)
+            myfiles <- bb_wget("https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi",paste0("-q --post-data=\"cksum=1&",method_flags,"\" -O -"),stdout=TRUE)
             if (is.null(attr(myfiles,"status")) || length(myfiles)>0) break
         } else {
             ## sys code, nb don't quote args, will break on unix
-            myfiles <- wget("https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi",c("-q",paste0("--post-data=cksum=1&",method_flags),"-O","-"))
+            myfiles <- bb_wget("https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi",c("-q",paste0("--post-data=cksum=1&",method_flags),"-O","-"))
             if (myfiles$status==0) break
         }
         tries <- tries+1
