+++ NAMESPACE
@@ -20,6 +20,7 @@ export(bb_sync)
 export(bb_uncompress)
 export(bb_unzip)
 export(bb_validate)
+export(bb_wget)
 export(earthdata_get)
 export(oceandata_get)
 export(oceandata_parameter_map)
