### 2017-08-27 86b95a2
update docs for --help
