### 2017-08-27 094be85
adjust test for new --help behaviour
