### 2017-08-27 7fda247
don't export or document internal oceandata functions
