+++ tests/testthat/test_oceandata.R
@@ -31,20 +31,20 @@ test_that("oceandata_get works",{
 })
 
 test_that("url mapper works", {
-    expect_identical(oceandata_url_mapper("https://oceandata.sci.gsfc.nasa.gov/cgi/getfile/A20021852007192.L3m_WC_SST_9.bz2",path_only=TRUE),"oceandata.sci.gsfc.nasa.gov/MODISA/Mapped/8D_Climatology/9km/SST/")
-    expect_identical(oceandata_url_mapper("https://oceandata.sci.gsfc.nasa.gov/cgi/getfile/A20030012003008.L3m_8D_KD490_Kd_490_9km.bz2",path_only=TRUE),"oceandata.sci.gsfc.nasa.gov/MODISA/Mapped/8Day/9km/Kd/2003/")
-    expect_identical(oceandata_url_mapper("https://oceandata.sci.gsfc.nasa.gov/cgi/getfile/A20090012009365.L3m_YR_FLH_ipar_9km.bz2",path_only=TRUE),"oceandata.sci.gsfc.nasa.gov/MODISA/Mapped/Annual/9km/ipar/")
-    expect_identical(oceandata_url_mapper("https://oceandata.sci.gsfc.nasa.gov/cgi/getfile/A20021852010120.L3m_CU_CDOM_cdom_index_9km.bz2",path_only=TRUE),"oceandata.sci.gsfc.nasa.gov/MODISA/Mapped/Cumulative/9km/cdom/")
-    expect_identical(oceandata_url_mapper("https://oceandata.sci.gsfc.nasa.gov/cgi/getfile/A2007005.L3m_DAY_FLH_nflh_9km.bz2",path_only=TRUE),"oceandata.sci.gsfc.nasa.gov/MODISA/Mapped/Daily/9km/nflh/2007/")
-    expect_identical(oceandata_url_mapper("https://oceandata.sci.gsfc.nasa.gov/cgi/getfile/A20023052002334.L3m_MO_FLH_nflh_9km.bz2",path_only=TRUE),"oceandata.sci.gsfc.nasa.gov/MODISA/Mapped/Monthly/9km/nflh/")
-    expect_identical(oceandata_url_mapper("https://oceandata.sci.gsfc.nasa.gov/cgi/getfile/A20021822006212.L3m_MC_SST4_9.bz2",path_only=TRUE),"oceandata.sci.gsfc.nasa.gov/MODISA/Mapped/Monthly_Climatology/9km/SST4/")
-    expect_identical(oceandata_url_mapper("https://oceandata.sci.gsfc.nasa.gov/cgi/getfile/A20110012011032.L3m_R32_RRS_aot_869_9km.bz2",path_only=TRUE),"oceandata.sci.gsfc.nasa.gov/MODISA/Mapped/Rolling_32_Day/9km/aot/2011/")
-    expect_identical(oceandata_url_mapper("https://oceandata.sci.gsfc.nasa.gov/cgi/getfile/A20040812004172.L3m_SNSP_PIC_pic_9km.bz2",path_only=TRUE),"oceandata.sci.gsfc.nasa.gov/MODISA/Mapped/Seasonal/9km/pic/")
-    expect_identical(oceandata_url_mapper("https://oceandata.sci.gsfc.nasa.gov/cgi/getfile/A20030802011171.L3m_SCSP_RRS_angstrom_9km.bz2",path_only=TRUE),"oceandata.sci.gsfc.nasa.gov/MODISA/Mapped/Seasonal_Climatology/9km/angstrom/")
[…]
