+++ NAMESPACE
@@ -23,11 +23,6 @@ export(bb_validate)
 export(bb_wget)
 export(earthdata_get)
 export(oceandata_get)
-export(oceandata_parameter_map)
-export(oceandata_parameters)
-export(oceandata_platform_map)
-export(oceandata_timeperiod_map)
-export(oceandata_url_mapper)
 importFrom(R.utils,bunzip2)
 importFrom(R.utils,gunzip)
 importFrom(assertthat,assert_that)

[…]
