+++ tests/testthat/test_sources.R
@@ -18,7 +18,7 @@ test_that("bb_source works with multiple postprocess actions", {
         source_url="",
         license="Please cite",
         comment="",
-        method=oceandata_get,
+        method=bb_handler_oceandata,
         method_flags=c("search=T20000322000060.L3m_MO_SST_sst_9km.nc"),
         postprocess=list(quote(bb_unzip(delete=TRUE)),bb_gunzip),
         access_function="",
@@ -33,7 +33,7 @@ test_that("bb_source works with multiple postprocess actions", {
         source_url="",
         license="Please cite",
         comment="",
-        method=oceandata_get,
+        method=bb_handler_oceandata,
         method_flags=c("search=T20000322000060.L3m_MO_SST_sst_9km.nc"),
         postprocess=list(bb_unzip,bb_gunzip),
         access_function="",
