+++ man/bb_handler_oceandata.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/earthdata_handler.R
-\name{earthdata_get}
-\alias{earthdata_get}
-\title{Handler for files files served up through Earthdata providers}
+% Please edit documentation in R/oceandata.R
+\name{bb_handler_oceandata}
+\alias{bb_handler_oceandata}
+\title{Handler for oceandata data sources}
 \usage{
-earthdata_get(config, verbose = FALSE, local_dir_only = FALSE)
+bb_handler_oceandata(config, verbose = FALSE, local_dir_only = FALSE)
 }
 \arguments{
 \item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source}
@@ -17,8 +17,8 @@ earthdata_get(config, verbose = FALSE, local_dir_only = FALSE)
 the directory if local_dir_only is TRUE, otherwise TRUE on success
 }
 \description{
-Handler for files files served up through Earthdata providers
+Handler for oceandata data sources
 }
 \references{
-https://wiki.earthdata.nasa.gov/display/EL/How+To+Register+With+Earthdata+Login
+https://oceandata.sci.gsfc.nasa.gov/
 }
