+++ man/bb_handler_earthdata.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/oceandata.R
-\name{oceandata_get}
-\alias{oceandata_get}
-\title{Handler for oceandata data sources}
+% Please edit documentation in R/earthdata_handler.R
+\name{bb_handler_earthdata}
+\alias{bb_handler_earthdata}
+\title{Handler for files files served up through Earthdata providers}
 \usage{
-oceandata_get(config, verbose = FALSE, local_dir_only = FALSE)
+bb_handler_earthdata(config, verbose = FALSE, local_dir_only = FALSE)
 }
 \arguments{
 \item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source}
@@ -17,8 +17,8 @@ oceandata_get(config, verbose = FALSE, local_dir_only = FALSE)
 the directory if local_dir_only is TRUE, otherwise TRUE on success
 }
 \description{
-Handler for oceandata data sources
+Handler for files files served up through Earthdata providers
 }
 \references{
-https://oceandata.sci.gsfc.nasa.gov/
+https://wiki.earthdata.nasa.gov/display/EL/How+To+Register+With+Earthdata+Login
 }
