### 2017-08-27 c18aba3
drop wget_default_flags entirely
