+++ R/config.R
@@ -3,8 +3,7 @@
 #' The parameters provided here are repository-wide settings, and will be applied to all data sources that are subsequently added to the configuration.
 #'
 #' @param local_file_root string: location of data repository on local file system
-#' @param wget_default_flags character vector: default flags to be passed to wget. These are overridden on a per-data source basis if the data source defines its own wget_flags
-#' @param wget_global_flags character vector: wget flags that will be applied to all data sources. These will be appended to either the data source wget flags (if specified), or the wget_default_flags
+#' @param wget_global_flags character vector: wget flags that will be applied to all data sources that call \code{bb_wget}. These will be appended to the data-source-specific wget flags provided via the source's method_flags argument
 #' @param http_proxy string: URL of HTTP proxy to use e.g. 'http://your.proxy:8080' (NULL for no proxy)
 #' @param ftp_proxy string: URL of FTP proxy to use e.g. 'http://your.proxy:21' (NULL for no proxy)
 #' @param clobber numeric: 0=do not overwrite existing files, 1=overwrite if the remote file is newer than the local copy, 2=always overwrite existing files. For data sources that use method 'wget', an appropriate flag will be added to the wget call according to the clobber setting ("--no-clobber" to not overwrite existing files, "--timestamping" to overwrite if the remote file is newer than the local copy)
@@ -25,16 +24,14 @@
 #' }
 #'
 #' @export
-bb_config <- function(local_file_root,wget_default_flags=character(),wget_global_flags=c("--restrict-file-names=windows","--progress=dot:giga"),http_proxy=NULL,ftp_proxy=NULL,clobber=1,skip_downloads=FALSE) {
+bb_config <- function(local_file_root,wget_global_flags=c("--restrict-file-names=windows","--progress=dot:giga"),http_proxy=NULL,ftp_proxy=NULL,clobber=1,skip_downloads=FALSE) {
     assert_that(is.string(local_file_root))
     assert_that(clobber %in% c(0,1,2))
     assert_that(is.flag(skip_downloads))
-    assert_that(is.character(wget_default_flags))
     assert_that(is.character(wget_global_flags))
     structure(
         list(data_sources=tibble(),
              settings=list(
-                 wget_default_flags=str_trim(wget_default_flags),
                  wget_global_flags=str_trim(wget_global_flags),
                  http_proxy=http_proxy,
                  ftp_proxy=ftp_proxy,
@@ -113,8 +110,7 @@ bb_settings <- function(config) {
 bb_settings_to_cols <- function(obj) {
     ## flags handled as lists
     obj$data_sources$wget_global_flags <- rep(list(obj$settings$wget_global_flags),nrow(obj$data_sources))
-    obj$data_sources$wget_default_flags <- rep(list(obj$settings$wget_default_flags),nrow(obj$data_sources))
-    for (nm in setdiff(names(obj$settings),c("wget_default_flags","wget_global_flags"))) {
+    for (nm in setdiff(names(obj$settings),c("wget_global_flags"))) {
         thisatt <- obj$settings[[nm]]
         if (!is.null(thisatt))
             obj$data_sources[,nm] <- thisatt
