+++ man/bb_config.Rd
@@ -4,7 +4,7 @@
 \alias{bb_config}
 \title{Initialize a bowerbird configuration}
 \usage{
-bb_config(local_file_root, wget_default_flags = character(),
+bb_config(local_file_root,
   wget_global_flags = c("--restrict-file-names=windows",
   "--progress=dot:giga"), http_proxy = NULL, ftp_proxy = NULL,
   clobber = 1, skip_downloads = FALSE)
@@ -12,9 +12,7 @@ bb_config(local_file_root, wget_default_flags = character(),
 \arguments{
 \item{local_file_root}{string: location of data repository on local file system}
 
-\item{wget_default_flags}{character vector: default flags to be passed to wget. These are overridden on a per-data source basis if the data source defines its own wget_flags}
-
-\item{wget_global_flags}{character vector: wget flags that will be applied to all data sources. These will be appended to either the data source wget flags (if specified), or the wget_default_flags}
+\item{wget_global_flags}{character vector: wget flags that will be applied to all data sources that call \code{bb_wget}. These will be appended to the data-source-specific wget flags provided via the source's method_flags argument}
 
 \item{http_proxy}{string: URL of HTTP proxy to use e.g. 'http://your.proxy:8080' (NULL for no proxy)}
 
