### 2017-08-27 d1160c3
tidy up commented-out code
