### 2017-08-27 1586992
added bb_find_wget function
