+++ man/bb_find_wget.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/wget.R
+\name{bb_find_wget}
+\alias{bb_find_wget}
+\title{Find the wget executable, and optionally install it if it is not found}
+\usage{
+bb_find_wget(install = FALSE)
+}
+\arguments{
+\item{install}{logical: install the executable if it is not found? (Windows only)}
+}
+\value{
+the path to the wget executable, or NULL if it was not found
+}
+\description{
+Installation (if required) currently only works on Windows platforms. The wget.exe executable will be downloaded from https://eternallybored.org/misc/wget/current/wget.exe and installed into your appdata directory (typically something like C:/Users/username/AppData/Roaming/)
+}
+\examples{
+\dontrun{
+  wget_path <- bb_find_wget()
+  wget_path <- bb_find_wget(install=TRUE) ## install (on windows) if needed
+}
+
+}
+\references{
+https://eternallybored.org/misc/wget/current/wget.exe
+}
+\seealso{
+\code{\link{bb_install_wget}}
+}
