+++ NAMESPACE
@@ -8,6 +8,7 @@ export(bb_config)
 export(bb_data_source_dir)
 export(bb_decompress)
 export(bb_example_sources)
+export(bb_find_wget)
 export(bb_fingerprint)
 export(bb_gunzip)
 export(bb_handler_earthdata)
