+++ README.md
@@ -18,7 +18,7 @@ library(devtools)
 install_github("AustralianAntarcticDivision/bowerbird",build_vignettes=TRUE)
 ```
 
-Bowerbird uses the third-party utility `wget` to do the heavy lifting of recursively downloading files from data providers. `wget` is typically installed by default on Linux. On Windows you can use the `install_wget()` function to install it. Otherwise download `wget` yourself (e.g. from <https://eternallybored.org/misc/wget/current/wget.exe>) and make sure it is on your path.
+Bowerbird uses the third-party utility `wget` to do the heavy lifting of recursively downloading files from data providers. `wget` is typically installed by default on Linux. On Windows you can use the `bb_install_wget()` function to install it. Otherwise download `wget` yourself (e.g. from <https://eternallybored.org/misc/wget/current/wget.exe>) and make sure it is on your path.
 
 Usage
 -----
@@ -42,14 +42,14 @@ my_source <- bb_source(
     citation="Spinoccia, M., 2012. XYZ multibeam bathymetric grids of the Macquarie Ridge. Geoscience Australia, Canberra.",
     source_url="http://www.ga.gov.au/corporate_data/73697/Macquarie_ESRI_Raster.zip",
     license="CC-BY 4.0",
-    method=quote(bb_wget),
+    method=quote(bb_handler_wget),
     method_flags="--recursive",
-    postprocess=quote(pp_unzip),
+    postprocess=quote(bb_unzip),
     collection_size=0.4,
     data_group="Topography")
 
 cf <- bb_config(local_file_root="/temp/data/bbtest") %>%
-    add(my_source)
+    bb_add(my_source)
 ```
 
 A few example data source definitions are provided as part of the bowerbird package --- see the list at the bottom of this document. Other packages (e.g. [blueant](https://github.com/AustralianAntarcticDivision/blueant)) provide themed sets of data sources that can be used with bowerbird.
@@ -88,9 +88,9 @@ my_source <- bb_source(
     citation="Spinoccia, M., 2012. XYZ multibeam bathymetric grids of the Macquarie Ridge. Geoscience Australia, Canberra.",
     source_url="http://www.ga.gov.au/corporate_data/73697/Macquarie_ESRI_Raster.zip",
     license="CC-BY 4.0",
-    method=quote(bb_wget),
+    method=quote(bb_handler_wget),
     method_flags="--recursive",
-    postprocess=quote(pp_unzip),
+    postprocess=quote(bb_unzip),
     collection_size=0.4,
     data_group="Topography")
 ```
@@ -101,19 +101,19 @@ Some particularly important components of this definition are:
 
 2.  The `license` and `citation` are important so that users know what conditions govern the usage of the data, and the appropriate citation to use to acknowledge the data providers. The `reference` entry should refer to a metadata or documentation page that describes the data in detail.
 
-3.  The `method`, `method_flags`, and `source_url` define how this data will be retrieved. Most sources will use the `bb_wget` method, which is a wrapper around the wget utility. If you are unfamiliar with wget, consult one of the many online tutorials. You can also see the in-built wget help by running `wget("--help")`.
+3.  The `method`, `method_flags`, and `source_url` define how this data will be retrieved. Most sources will use the `bb_handler_wget` method, which is a wrapper around the wget utility. If you are unfamiliar with wget, consult one of the many online tutorials. You can also see the in-built wget help by running `bb_wget("--help")`.
 
 Some subtleties to bear in mind:
 
[…]
