+++ R/sync.R
@@ -129,17 +129,17 @@ do_sync_repo <- function(this_dataset,create_root,verbose,settings) {
             if (is.function(qq)) {
                 ## passed as function
                 ## evaluate with extra args
-                do.call(qq,list(cfrow=this_dataset,file_list_before=file_list_before,file_list_after=file_list_after))
+                do.call(qq,list(config=this_dataset,file_list_before=file_list_before,file_list_after=file_list_after))
             } else if (is.symbol(qq)) {
                 ## passed as symbol
-                do.call(eval(qq),list(cfrow=this_dataset,file_list_before=file_list_before,file_list_after=file_list_after))
+                do.call(eval(qq),list(config=this_dataset,file_list_before=file_list_before,file_list_after=file_list_after))
             } else if (is.call(qq)) {
                 if (all.names(qq)[1]=="quote") {
                     ## call was constructed as e.g. enquote(fun)
-                    do.call(eval(qq),list(cfrow=this_dataset,file_list_before=file_list_before,file_list_after=file_list_after))
+                    do.call(eval(qq),list(config=this_dataset,file_list_before=file_list_before,file_list_after=file_list_after))
                 } else {
                     ## call was constructed as e.g. quote(fun()) or quote(fun(var=arg))
-                    thisargs <- inject_args(qq,list(cfrow=this_dataset,file_list_before=file_list_before,file_list_after=file_list_after))
+                    thisargs <- inject_args(qq,list(config=this_dataset,file_list_before=file_list_before,file_list_after=file_list_after))
                     do.call(all.names(qq)[1],thisargs)
                 }
             }
