### 2017-08-27 8aaf1f1
test decompress postprocess
