### 2017-08-28 adebf2f
temporarily skip wget flags test
