### 2017-08-28 4158b1a
set travis and appveyor as not cran so that extra tests are run
