### 2017-08-28 1ece8ad
skip postprocess tests on appveyor
