### 2017-08-28 d42fc9f
appveyor skip not_cran tests again
