### 2017-08-29 9a31881
remove stop_on_error parm from bb_wget, add capture_stdout
