+++ R/bowerbird.R
@@ -11,6 +11,6 @@
 #' @importFrom R.utils bunzip2 gunzip
 #' @importFrom rmarkdown render
 #' @importFrom stats na.omit
-#' @importFrom sys exec_wait
+#' @importFrom sys exec_wait exec_internal
 #' @importFrom utils download.file read.table str unzip
 NULL
