+++ R/oceandata.R
@@ -50,7 +50,7 @@ bb_handler_oceandata <- function(config,verbose=FALSE,local_dir_only=FALSE) {
         ## sometimes this takes a couple of attempts!
         if (FALSE) {
             ## system2 code
-            myfiles <- bb_wget("https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi",paste0("-q --post-data=\"cksum=1&",method_flags,"\" -O -"),stdout=TRUE)
+            myfiles <- bb_wget("https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi",paste0("-q --post-data=\"cksum=1&",method_flags,"\" -O -"),capture_stdout=TRUE)
             if (is.null(attr(myfiles,"status")) || length(myfiles)>0) break
         } else {
             ## sys code, nb don't quote args, will break on unix
