+++ NAMESPACE
@@ -50,6 +50,7 @@ importFrom(stringr,str_detect)
 importFrom(stringr,str_match)
 importFrom(stringr,str_split)
 importFrom(stringr,str_trim)
+importFrom(sys,exec_internal)
 importFrom(sys,exec_wait)
 importFrom(utils,download.file)
 importFrom(utils,read.table)
