### 2017-08-29 2529143
change example dataset, other tweaks
