+++ README.Rmd
@@ -22,6 +22,12 @@ knitr::opts_chunk$set(
 
 Often it's desirable to have local copies of third-party data sets. Fetching data on the fly from remote sources can be a great strategy, but for speed or other reasons it may be necessary to have local copies. This is particularly common in ecology and the environmental sciences. Bowerbird is an R package for maintaining a local collection of data sets from a range of data providers.
 
+Bowerbird might be used in several different modes:
+
+- interactively, from the R console, to download or update data files on an as-needed basis
+- from the command line, perhaps as a regular scheduled task
+- from within other R packages.
+
 ## Installing
 
 ```{r eval=FALSE}
@@ -43,24 +49,23 @@ Build up a configuration by first defining global options such as the destinatio
 cf <- bb_config(local_file_root="~/your/data/directory")
 ```
 
-Bowerbird must then be told which data sources to synchronize. Use the `bb_source()` function to define a data source:
+Bowerbird must then be told which data sources to synchronize. Use the `bb_source()` function to define a data source. Let's use a dataset of bike paths in Geelong as an example:
 
 ```{r eval=FALSE}
 my_source <- bb_source(
-    name="Geoscience Australia multibeam bathymetric grids of the Macquarie Ridge",
-    id="10.4225/25/53D9B12E0F96E",
-    description="This is a compilation of all the processed multibeam bathymetry data that are publicly available in Geoscience Australia's data holding for the Macquarie Ridge.",
-    reference="http://www.ga.gov.au/metadata-gateway/metadata/record/gcat_b9224f95-a416-07f8-e044-00144fdd4fa6/XYZ+multibeam+bathymetric+grids+of+the+Macquarie+Ridge",
-    citation="Spinoccia, M., 2012. XYZ multibeam bathymetric grids of the Macquarie Ridge. Geoscience Australia, Canberra.",
-    source_url="http://www.ga.gov.au/corporate_data/73697/Macquarie_ESRI_Raster.zip",
-    license="CC-BY 4.0",
+    name="Bike Paths - Greater Geelong",
+    id="http://data.gov.au/dataset/7af9cf59-a4ea-47b2-8652-5e5eeed19611",
+    description="Polyline data of bike path locations for the City of Greater Geelong.",
+    reference="https://data.gov.au/dataset/geelong-bike-paths",
+    citation="",
+    source_url="https://data.gov.au/dataset/7af9cf59-a4ea-47b2-8652-5e5eeed19611",
+    license="CC-BY",
     method=quote(bb_handler_wget),
-    method_flags="--recursive",
+    method_flags=c("--recursive","--level",1,"--accept-regex=download","--adjust-extension"),
     postprocess=quote(bb_unzip),
-    collection_size=0.4,
-    data_group="Topography")
+    collection_size=0.002)
 
-cf <- bb_config(local_file_root="/temp/data/bbtest") %>%
+cf <- bb_config(local_file_root="~/temp/data/bbtest") %>%
     bb_add(my_source)
 ```
 
@@ -76,9 +81,20 @@ bb_sync(cf)
 
 Congratulations! You now have your own local copy of your chosen data sets.
 
+
 ### Ongoing updates
 
 At a later time you can re-run this synchronization process. If the remote files have not changed, and assuming that your configuration has the `clobber` parameter set to 0 (do not overwrite existing files) or 1 (overwrite only if the remote file is newer than the local copy) then the sync process will run quite quickly because it will not need to re-download any data files.
+
+### Use of bowerbird within other packages
+
[…]
