### 2017-08-30 0f705e8
fix Z-decompression, add test
