+++ tests/testthat/test_postprocess.R
@@ -78,5 +78,30 @@ test_that("decompressing bzipped files works",{
     expect_named(x,c("a","b","c"))
 })
 
+test_that("decompressing Z-compressed files works",{
+    skip_on_cran()
+    skip_on_windows() ## requires archive, which isn't yet on cran and requires compilation from source
+    my_source <- bb_source(
+        name="Bowerbird test data",
+        id="bbtest-v0.1",
+        description="These are just some trivial test files provided with the bowerbird package.",
+        reference="https://github.com/AustralianAntarcticDivision/bowerbird",
+        citation="No citation needed.",
+        source_url="https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/20170822.nc.Z",
+        license="MIT",
+        method=quote(bb_handler_wget),
+        method_flags=c("--recursive","--level=1","-e","robots=off"),
+        postprocess=quote(bb_uncompress))
+
+    temp_root <- tempdir()
+    cf <- bb_add(bb_config(local_file_root=temp_root,clobber=2),my_source)
+    bb_sync(cf)
+
+    fp <- bb_data_source_dir(cf)
+    expect_true(file.exists(file.path(fp,"20170822.nc.Z")))
+    expect_true(file.exists(file.path(fp,"20170822.nc")))
+    expect_true(file.info(file.path(fp,"20170822.nc"))$size>1e4)
+})
+
 
 
