### 2017-08-30 3ee9cea
fix os check, skip_on_os('windows') not skip_on_windows()!
