### 2017-08-30 8d02369
partial reenable resolve_wget_clobber_flags test
