### 2017-08-30 52199f4
try again with appveyor running not-cran tests
