### 2017-08-31 ee8dbf2
add --no-check-certificate to wget calls to github to avoid cert failures on appveyor
