### 2017-08-31 b1e05ed
change test URLs from github/blah/raw/ to raw.githubusercontent.com to avoid 302 redirects
