+++ tests/testthat/test_postprocess.R
@@ -9,7 +9,7 @@ test_that("decompressing zip files works",{
         description="These are just some trivial test files provided with the bowerbird package.",
         reference="https://github.com/AustralianAntarcticDivision/bowerbird",
         citation="No citation needed.",
-        source_url="https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data.zip",
+        source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data.zip",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data.zip",
         license="MIT",
         method=quote(bb_handler_wget),
         method_flags=c("--recursive","--level=1","-e","robots=off"),
@@ -35,7 +35,7 @@ test_that("decompressing gzipped files works",{
         description="These are just some trivial test files provided with the bowerbird package.",
         reference="https://github.com/AustralianAntarcticDivision/bowerbird",
         citation="No citation needed.",
-        source_url="https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data_was_gzipped.csv.gz",
+        source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data_was_gzipped.csv.gz",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data_was_gzipped.csv.gz",
         license="MIT",
         method=quote(bb_handler_wget),
         method_flags=c("--recursive","--level=1","-e","robots=off"),
@@ -61,7 +61,7 @@ test_that("decompressing bzipped files works",{
         description="These are just some trivial test files provided with the bowerbird package.",
         reference="https://github.com/AustralianAntarcticDivision/bowerbird",
         citation="No citation needed.",
-        source_url="https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data_was_bzipped.csv.bz2",
+        source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data_was_bzipped.csv.bz2",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data_was_bzipped.csv.bz2",
         license="MIT",
         method=quote(bb_handler_wget),
         method_flags=c("--recursive","--level=1","-e","robots=off"),
@@ -87,7 +87,7 @@ test_that("decompressing Z-compressed files works",{
         description="These are just some trivial test files provided with the bowerbird package.",
         reference="https://github.com/AustralianAntarcticDivision/bowerbird",
         citation="No citation needed.",
-        source_url="https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/20170822.nc.Z",
+        source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/20170822.nc.Z",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/20170822.nc.Z",
         license="MIT",
         method=quote(bb_handler_wget),
         method_flags=c("--recursive","--level=1","-e","robots=off"),
