+++ tests/testthat/test_provenance.R
@@ -2,14 +2,14 @@ context("provenance functions")
 
 test_that("bb_fingerprint does something sensible",{
     skip_on_cran()
-    #skip_on_appveyor() ## appveyor does not recognize github's certificate
+    skip_on_appveyor() ## wget behaviour on appveyor
     my_source <- bb_source(
         name="Bowerbird test data",
         id="bbtest-v0.1",
         description="These are just some trivial test files provided with the bowerbird package.",
         reference="https://github.com/AustralianAntarcticDivision/bowerbird",
         citation="No citation needed.",
-        source_url="https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data_was_gzipped.csv.gz",
+        source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data_was_gzipped.csv.gz",##https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data_was_gzipped.csv.gz",
         license="MIT",
         method=quote(bb_handler_wget),
         method_flags=c("--recursive","--level=1","--no-check-certificate","-e","robots=off"),
