### 2017-08-31 fb97e46
skip oceandata tests on appveyor
