### 2017-09-26 a687911
minor readme text tweak
