### 2017-09-27 3d7996a
enforce sys version >= 1.4.9000 to avoid long line bug
