### 2017-09-30 #139 @maelle
@raymondben @mdsumner just checking in. 😸  Have you made enough progress on the splitting for @MilesMcBain & @lwasser to soon review the package? Thanks!
