### 2017-10-02 2421d77
don't shorten wget flags on windows now that sys bug is fixed
