### 2017-10-02 5748909
remove irrelevant comments
