### 2017-10-02 5a631dc
add test for missing auth info in earthdata handler
