### 2017-10-02 9c007a7
don't run postprocessing if download failed or was interrupted
