### 2017-10-02 41d70a1
update text in readme and vignettes
