+++ README.Rmd
@@ -20,15 +20,18 @@ knitr::opts_chunk$set(
 
 <img align="right" src="https://rawgit.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/bowerbird.svg" />
 
-Often it's desirable to have local copies of third-party data sets. Fetching data on the fly from remote sources can be a great strategy, but for speed or other reasons it may be necessary to have local copies. This is particularly common in ecology and the environmental sciences. Bowerbird is an R package for maintaining a local collection of data sets from a range of data providers.
+Often it's desirable to have local copies of third-party data sets. Fetching data on the fly from remote sources can be a great strategy, but for speed or other reasons it may be better to have local copies. This is particularly common in environmental and other sciences that deal with large data sets (e.g. satellite or global climate model products). Bowerbird is an R package for maintaining a local collection of data sets from a range of data providers.
 
-Bowerbird might be used in several different modes:
+Bowerbird can be used in several different modes:
 
-- interactively, from the R console, to download or update data files on an as-needed basis
+- interactively from the R console, to download or update data files on an as-needed basis
 - from the command line, perhaps as a regular scheduled task
-- from within other R packages.
+- programatically, including from within other R packages, scripts, or R markdown documents that require local copies of particular data files.
 
 ## Installing
+
+Temporary note: until `sys` version 1.5 hits CRAN, you will need to do `devtools::install_github("jeroen/sys")` to install from github, prior to installing bowerbird.
+
 
 ```{r eval=FALSE}
 install.packages("devtools")
@@ -49,7 +52,7 @@ Build up a configuration by first defining global options such as the destinatio
 cf <- bb_config(local_file_root="~/your/data/directory")
 ```
 
-Bowerbird must then be told which data sources to synchronize. Use the `bb_source()` function to define a data source. Let's use a dataset of bike paths in Geelong as an example:
+Bowerbird must then be told which data sources to synchronize. Use the `bb_source()` function to define a data source. Let's use a dataset of bike paths in Geelong, Australia as an example:
 
 ```{r eval=FALSE}
 my_source <- bb_source(
@@ -84,7 +87,7 @@ Congratulations! You now have your own local copy of your chosen data sets.
 
 ### Ongoing updates
 
-At a later time you can re-run this synchronization process. If the remote files have not changed, and assuming that your configuration has the `clobber` parameter set to 0 (do not overwrite existing files) or 1 (overwrite only if the remote file is newer than the local copy) then the sync process will run quite quickly because it will not need to re-download any data files.
+At a later time you can re-run this synchronization process. If the remote files have not changed, and assuming that your configuration has the `clobber` parameter set to 0 (do not overwrite existing files) or 1 (overwrite only if the remote file is newer than the local copy) then the sync process will run more quickly because it will not need to re-download any data files.
 
 ### Use of bowerbird within other packages
 
@@ -96,11 +99,16 @@ Bowerbird might be useful to developers of packages that download resources from
 - incrementally update files that you have previously downloaded. Bowerbird can be instructed not to re-download files that exist locally, unless they have changed on the remote server. Compressed files will also only be decompressed if changed.
 
 
+## Data source handlers
+
[…]
