### 2017-10-02 8bc84f7
change example data set in readme/vignette [ci skip]
