### 2017-10-03 d49dd1a
minor updates to readme/vignette text
