### 2017-10-03 0ef29ae
all source_url entries in c() just for consistency
