+++ R/example_sources.R
@@ -29,7 +29,7 @@ bb_example_sources <- function() {
             description="Weekly and monthly mean and long-term monthly mean SST data, 1-degree resolution, 1981 to present. Ice concentration data are also included, which are the ice concentration values input to the SST analysis",
             reference= "http://www.esrl.noaa.gov/psd/data/gridded/data.noaa.oisst.v2.html",
             citation="NOAA_OI_SST_V2 data provided by the NOAA/OAR/ESRL PSD, Boulder, Colorado, USA, from their web site at http://www.esrl.noaa.gov/psd/",
-            source_url="ftp://ftp.cdc.noaa.gov/Datasets/noaa.oisst.v2/*",
+            source_url=c("ftp://ftp.cdc.noaa.gov/Datasets/noaa.oisst.v2/*"),
             license="Please cite",
             method=quote(bb_handler_wget),
             method_flags=c("--recursive","--level=1","--no-parent"),
@@ -72,7 +72,7 @@ bb_example_sources <- function() {
             description="This data set (NmIcEdg2) estimates the location of the North and South Pole sea ice edges at various times during the mid to late 1960s, based on recovered Nimbus 1 (1964), Nimbus 2 (1966), and Nimbus 3 (1969) visible imagery.",
             reference="http://nsidc.org/data/nmicedg2/",
             citation="Gallaher, D. and G. Campbell. 2014. Nimbus Ice Edge Points from Nimbus Visible Imagery L2, CSV. [indicate subset used]. Boulder, Colorado USA: NASA National Snow and Ice Data Center Distributed Active Archive Center. http://dx.doi.org/10.5067/NIMBUS/NmIcEdg2",
-            source_url="https://n5eil01u.ecs.nsidc.org/NIMBUS/NmIcEdg2.001/",
+            source_url=c("https://n5eil01u.ecs.nsidc.org/NIMBUS/NmIcEdg2.001/"),
             license="Please cite, see http://nsidc.org/about/use_copyright.html",
             authentication_note="Requires Earthdata login, see https://urs.earthdata.nasa.gov/",
             method=quote(bb_handler_earthdata),
@@ -87,7 +87,7 @@ bb_example_sources <- function() {
             id="greatlakes-superior-bathymetry",
             description="A draft version of the Lake Superior Bathymetry was compiled as a component of a NOAA project to rescue Great Lakes lake floor geological and geophysical data, and make it more accessible to the public. No time frame has been set for completing bathymetric contours of Lake Superior, though a 3 arc-second (~90 meter cell size) grid is available.",
             reference="https://www.ngdc.noaa.gov/mgg/greatlakes/superior.html",
-            source_url="https://www.ngdc.noaa.gov/mgg/greatlakes/superior/data/",
+            source_url=c("https://www.ngdc.noaa.gov/mgg/greatlakes/superior/data/"),
             citation="Publisher: DOC/NOAA/NESDIS/NCEI > National Centers for Environmental Information, NESDIS, NOAA, U.S. Department of Commerce",
             license="https://www.ngdc.noaa.gov/ngdcinfo/privacy.html#copyright",
             method=quote(bb_handler_wget),
