### 2017-10-03 1427400
minor edit to doc for bb_source
