### 2017-10-03 6987a85
partial changes to make source_url in config a list column
