+++ R/source.R
@@ -77,7 +77,7 @@ bb_source <- function(id,name,description=NA_character_,reference,source_url,cit
         name=name,
         description=if (assert_that(is.string(description))) description,
         reference=if (assert_that(is.string(reference))) reference,
-        source_url=source_url,
+        source_url=list(source_url),
         citation=if (assert_that(is.string(citation))) citation,
         license=if (assert_that(is.string(license))) license,
         comment=if (assert_that(is.string(comment))) comment,
