### 2017-10-03 c704159
directory_from_url() can take char or list of char
