### 2017-10-04 e64b8a3
check that bb_source handles missing/NA/empty source_url correctly
