### 2017-10-04 29f037a
cope with source_url as list column
