+++ DESCRIPTION
@@ -27,5 +27,5 @@ Suggests:
     knitr,
     testthat,
     covr
-Remotes: jimhester/archive
+Remotes: jimhester/archive, jeroen/sys
 VignetteBuilder: knitr
