### 2017-10-04 98c8563
update earthdata handler for list-column source_url
