### 2017-10-04 d269cf6
alter comments on parallelization
