### 2017-10-05 c5c9b38
add get/set functions for config's data_sources and settings elements
