+++ NAMESPACE
@@ -1,11 +1,14 @@
 # Generated by roxygen2: do not edit by hand
 
 export("%>%")
+export("bb_data_sources<-")
+export("bb_settings<-")
 export(bb_add)
 export(bb_bunzip2)
 export(bb_cleanup)
 export(bb_config)
 export(bb_data_source_dir)
+export(bb_data_sources)
 export(bb_decompress)
 export(bb_example_sources)
 export(bb_find_wget)

[…]
