### 2017-10-05 0efe684
use bb_data_source_dir to work out destionation directory
