### 2017-10-06 d9bf61a
earthdata handler mods
