+++ R/example_sources.R
@@ -67,6 +67,22 @@ bb_example_sources <- function() {
             collection_size=7.2,
             data_group="Ocean colour"),
         bb_source(
+            name="Sea Ice Trends and Climatologies from SMMR and SSM/I-SSMIS, Version 2",
+            id="10.5067/EYICLBOAAJOU",
+            description="NSIDC provides this data set to aid in the investigations of the variability and trends of sea ice cover. Ice cover in these data are indicated by sea ice concentration: the percentage of the ocean surface covered by ice. The ice-covered area indicates how much ice is present; it is the total area of a pixel multiplied by the ice concentration in that pixel. Ice persistence is the percentage of months over the data set time period that ice existed at a location. The ice-extent indicates whether ice is present; here, ice is considered to exist in a pixel if the sea ice concentration exceeds 15 percent. This data set provides users with data about total ice-covered areas, sea ice extent, ice persistence, and monthly climatologies of sea ice concentrations.",
+            reference="https://nsidc.org/data/NSIDC-0192/versions/2",
+            citation="Stroeve, J. and W. Meier. 2017. Sea Ice Trends and Climatologies from SMMR and SSM/I-SSMIS, Version 2. [Indicate subset used]. Boulder, Colorado USA. NASA National Snow and Ice Data Center Distributed Active Archive Center. doi: http://dx.doi.org/10.5067/EYICLBOAAJOU. [Date Accessed].",
+            source_url=c("https://daacdata.apps.nsidc.org/pub/DATASETS/nsidc0192_seaice_trends_climo_v2/"),
+            license="Please cite, see http://nsidc.org/about/use_copyright.html",
+            authentication_note="Requires Earthdata login, see https://urs.earthdata.nasa.gov/. Note that you will need to authorize the application nsidc-daacdata",
+            method=quote(bb_handler_earthdata),
+            method_flags=c("--recursive","--level=3","--no-parent","--relative","--reject-regex=^/pub/DATASETS/"),
+            user="",
+            password="",
+            postprocess=NULL,
+            collection_size=NA,
+            data_group="Sea ice",warn_empty_auth=FALSE),
+        bb_source(
             name="Nimbus Ice Edge Points from Nimbus Visible Imagery",
             id="10.5067/NIMBUS/NmIcEdg2",
             description="This data set (NmIcEdg2) estimates the location of the North and South Pole sea ice edges at various times during the mid to late 1960s, based on recovered Nimbus 1 (1964), Nimbus 2 (1966), and Nimbus 3 (1969) visible imagery.",
@@ -74,9 +90,9 @@ bb_example_sources <- function() {
             citation="Gallaher, D. and G. Campbell. 2014. Nimbus Ice Edge Points from Nimbus Visible Imagery L2, CSV. [indicate subset used]. Boulder, Colorado USA: NASA National Snow and Ice Data Center Distributed Active Archive Center. http://dx.doi.org/10.5067/NIMBUS/NmIcEdg2",
             source_url=c("https://n5eil01u.ecs.nsidc.org/NIMBUS/NmIcEdg2.001/"),
             license="Please cite, see http://nsidc.org/about/use_copyright.html",
-            authentication_note="Requires Earthdata login, see https://urs.earthdata.nasa.gov/",
+            authentication_note="Requires Earthdata login, see https://urs.earthdata.nasa.gov/. Note that you will need to authorize the application NSIDC_DATAPOOL_OPS",
             method=quote(bb_handler_earthdata),
-            method_flags=c("--recursive","--level=2","--no-parent","--accept-regex=/NmIcEdg2.001/"),
+            method_flags=c("--recursive","--level=2","--no-parent"),##,"--accept-regex=/NmIcEdg2.001/"),
             user="",
             password="",
             postprocess=NULL,
