### 2017-10-06 1e9194e
mention wget's --debug flag
