### 2017-10-06 5f11136
clean up cookies temp file and index.html* files
