### 2017-10-06 a652955
drop ice edge example data set, slightly cleaner sea ice trends definition
