+++ R/example_sources.R
@@ -74,29 +74,13 @@ bb_example_sources <- function() {
             citation="Stroeve, J. and W. Meier. 2017. Sea Ice Trends and Climatologies from SMMR and SSM/I-SSMIS, Version 2. [Indicate subset used]. Boulder, Colorado USA. NASA National Snow and Ice Data Center Distributed Active Archive Center. doi: http://dx.doi.org/10.5067/EYICLBOAAJOU. [Date Accessed].",
             source_url=c("https://daacdata.apps.nsidc.org/pub/DATASETS/nsidc0192_seaice_trends_climo_v2/"),
             license="Please cite, see http://nsidc.org/about/use_copyright.html",
-            authentication_note="Requires Earthdata login, see https://urs.earthdata.nasa.gov/. Note that you will need to authorize the application nsidc-daacdata",
+            authentication_note="Requires Earthdata login, see https://urs.earthdata.nasa.gov/. Note that you will also need to authorize the application 'nsidc-daacdata' (see 'My Applications' at https://urs.earthdata.nasa.gov/profile)",
             method=quote(bb_handler_earthdata),
-            method_flags=c("--recursive","--level=3","--no-parent","--relative","--reject-regex=^/pub/DATASETS/"),
+            method_flags=c("--recursive","--level=4","--no-parent","--relative"),
             user="",
             password="",
             postprocess=NULL,
-            collection_size=NA,
-            data_group="Sea ice",warn_empty_auth=FALSE),
-        bb_source(
-            name="Nimbus Ice Edge Points from Nimbus Visible Imagery",
-            id="10.5067/NIMBUS/NmIcEdg2",
-            description="This data set (NmIcEdg2) estimates the location of the North and South Pole sea ice edges at various times during the mid to late 1960s, based on recovered Nimbus 1 (1964), Nimbus 2 (1966), and Nimbus 3 (1969) visible imagery.",
-            reference="http://nsidc.org/data/nmicedg2/",
-            citation="Gallaher, D. and G. Campbell. 2014. Nimbus Ice Edge Points from Nimbus Visible Imagery L2, CSV. [indicate subset used]. Boulder, Colorado USA: NASA National Snow and Ice Data Center Distributed Active Archive Center. http://dx.doi.org/10.5067/NIMBUS/NmIcEdg2",
-            source_url=c("https://n5eil01u.ecs.nsidc.org/NIMBUS/NmIcEdg2.001/"),
-            license="Please cite, see http://nsidc.org/about/use_copyright.html",
-            authentication_note="Requires Earthdata login, see https://urs.earthdata.nasa.gov/. Note that you will need to authorize the application NSIDC_DATAPOOL_OPS",
-            method=quote(bb_handler_earthdata),
-            method_flags=c("--recursive","--level=2","--no-parent"),##,"--accept-regex=/NmIcEdg2.001/"),
-            user="",
-            password="",
-            postprocess=NULL,
-            collection_size=0.1,
+            collection_size=0.02,
             data_group="Sea ice",warn_empty_auth=FALSE),
         bb_source(
             name="Bathymetry of Lake Superior",
