### 2017-10-06 403c9ab
switch test to new earthdata example source
