### 2017-10-09 7d52e89
update readme and vignette
