+++ R/example_sources.R
@@ -10,7 +10,7 @@
 #' }
 #' @references See the \code{reference} and \code{citation} field in each row of the returned tibble for references associated with these particular data sources
 #'
-#' @return data.frame
+#' @return tibble
 #'
 #' @seealso \code{\link{bb_config}} \code{\link{bb_handler_wget}} \code{\link{bb_handler_oceandata}} \code{\link{bb_handler_earthdata}}
 #'
