+++ R/source.R
@@ -19,7 +19,7 @@
 #' @param collection_size numeric: approximate disk space (in GB) used by the data collection, if known. If the data are supplied as compressed files, this size should reflect the disk space used after decompression. If the data_source definition contains multiple source_url entries, this size should reflect the overall disk space used by all combined
 #' @param warn_empty_auth logical: if TRUE, issue a warning if the data source requires authentication (authentication_note is not NA) but user and password have not been provided
 #'
-#' @return data.frame
+#' @return tibble
 #'
 #' @seealso \code{\link{bb_config}}
 #'
