### 2017-10-09 #139 @raymondben
@MilesMcBain @lwasser @maelle

Sorry for the hiatus, think we're OK to resume the review process now. The major change is that the data source definitions have been removed, except for a few representative examples. (The full set of definitions has moved to https://github.com/AustralianAntarcticDivision/blueant. This is a separate package and not part of this ropensci submission, but it's there should you want to get some idea of how a themed-data-package might make use of bowerbird.)

Please note that until `sys` version 1.5 hits CRAN, you'll need to install this from github using `devtools::install_github("jeroen/sys")` prior to installing bowerbird. Bowerbird requires v1.5 in order to avoid a Windows bug in 1.4. Also note that check is currently failing on both Travis and AppVeyor, but because of some Ubuntu-package issue on Travis (that I haven't yet figured out) and on AppVeyor because of the `sys` issue. But bowerbird is passing check locally on both Windows and Ubuntu, so I think it's OK.

Some replies to Miles' partial review:
> Not a great first user experience

Yah, fair enough. The example is now a small one, and in addition we've switched from using `system2` to `sys`, which means that the user can interrupt the external process without needing to kill their session.

> If this package is intended to be used exclusively from the command line ...

No, it's not exclusively this. See updated comments in README. Also, we can't give progress indicators (also discussed in README).

>  is there a way to check sync status in interactive mode without fetching these large repos?

I'm not sure what you mean by this, sorry.

> `bb_config` seems unusual in that returns its configuration as attributes of an empty tibble ...

That structure has now changed and doesn't rely on  attributes to carry the repo-wide settings. `bb_config` now returns a list with two elements: `data_sources` (a tibble) and `settings` (a list of repo-wide settings).

All exported function names are now prepended with `bb_`.
