### 2017-10-09 3ccf095
comment out apt-get update to see if it avoids current travis issue
