+++ .travis.yml
@@ -5,7 +5,7 @@ sudo: required
 cache: packages
 
 before_install:
-  - sudo apt-get -qq update
+#  - sudo apt-get -qq update
   - sudo apt-get install -y libarchive-dev libapparmor-dev
 
 # set NOT_CRAN to true to ensure that skip_on_cran tests are run here
