### 2017-10-09 2630bcd
fix bug in bb_validate, add test
