### 2017-10-12 2c54f8b
bb_sync now returns a tibble not a vector
