### 2017-10-12 5b596e6
remove jeroen/sys remote now that sys v1.5 is on CRAN
