+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: bowerbird
 Type: Package
 Title: Keep a Collection of Sparkly Data Resources
-Version: 0.5.1
+Version: 0.6.0
 Authors@R: c(person("Ben","Raymond",email="ben.raymond@aad.gov.au",
   	   role=c("aut","cre")),
 	   person("Michael","Sumner",role="aut"))
@@ -27,5 +27,5 @@ Suggests:
     knitr,
     testthat,
     covr
-Remotes: jimhester/archive, jeroen/sys
+Remotes: jimhester/archive
 VignetteBuilder: knitr
