### 2017-10-12 8f4a371
remove temporary note about sys package [ci skip]
