### 2017-10-24 #139 @MilesMcBain
Thanks for the update @raymondben. I look forward to reviewing the new version.

Regarding this comment:
> is there a way to check sync status in interactive mode without fetching these large repos?

Since the package is intended to be used interactively I thought it would be helpful if I can check I am on the latest version without initiating a large download. If it turns out I need to sync, I'd probably like to tee it up in separate session so that it doesn't block my main work environment. It might be this is already possible and I did not see the function.
