### 2017-10-24 #139 @raymondben
@MilesMcBain re:

> is there a way to check sync status in interactive mode without fetching these large repos?

Oh, right, gotcha. The answer is sort of "yes, but no". There is `bb_config(...,skip_downloads=TRUE)` which will cause the subsequent sync process to do a dry run without downloading any data files. But in reality, this is less useful than I'd hoped. It suppresses any actual calls to `wget`, and so any recursive `wget` action (descending from the top-level data set URL) doesn't get replicated in the dry run. So quite possibly you won't get any indication of whether or not you have particular data files within the repo. For that reason, I'm debating whether the `skip_downloads` parameter is even worth keeping. And as far as I know, there's no way to instruct wget to report on files that it would download, but not actually download them. On first glance the wget `--spider` flag *sounds* like it might be appropriate, but I've discovered that in this mode wget will delete existing files, which is less than helpful.
