### 2017-11-07 #139 @lwasser
HI guys -- where is the best place to get feedback on issues i'm running into ? for example i tried to run the vignette on the readme and got this error

```
sh: wget: command not found
Error in bb_sync(cf) :
  could not find the wget executable. Try the bb_install_wget() function, or install it yourself and ensure that it is on the path
```
i'd like to be able to work through the vignettes!
