### 2017-11-07 #139 @raymondben
Hi @lwasser, feel free to raise issues here. I don't have/use a Mac, but I if you use Homebrew I *think* that you can just do `brew install wget`  to install wget. I'll find someone here with a Mac and see if that works for me ...
