### 2017-11-07 #139 @lwasser
Thank you @raymondben  ! I did try brew. it wants me to update ruby however and i've got a few ruby scrips that i'm afraid of breaking... so i'm running those now to get my data and then i can risk it breaking tonight :) . if there is an alternative install option please say the word. i'm looking around now.

i have another question for you guys. Who is your intended user for this package? And do you have use cases? Do you want feedback related to the ease of use and expected knowledge someone who need to have to get this working?
