### 2017-11-07 #139 @raymondben
Nicely done. Clearly, better Mac support is something to put on our to-do list!

Re - questions:
1. we don't have a particular view on intended users (the README/vignette touches on this a little). It could be users who need to keep local copies of data files and just want a convenient mechanism to do so. It could be a super-user who does the same on behalf of a number of local users (this is our use case; see below). It could be useful to package authors who need to retrieve data but don't want to reinvent code for mirroring, incremental updates, etc. It could be useful as a means of distributing example code (e.g. I want to send someone some demo code of some sort, but it needs 100GB of data files to run, so rather than send the data files directly or have to explain to the other person where to get them, I can just send them a bowerbird snippet).
2. Uses cases - one is definitely the maintenance of a local library of data files. We use it to maintain a ~5TB or so library of satellite and other environmental data for use by a number of people in our local research community. See the blueant package mentioned above for our current list of ~50 data sources that we mirror.
3. Yes please. I will not be surprised, however, if you were to say that it's all a bit opaque and difficult to get working! We've tried to pave the way, but I am sure it can be improved and that's something I'm hoping the ropensci onboarding can help with.
