### 2017-11-07 #139 @lwasser
wonderful. thanks @raymondben  Ok well i did focus a lot on #3 above. And if my feedback is helpful -- i'm happy to continue to test things as the package developed :) .  My apologies that this took longer than I know you all wanted it to.
****
Hi @mdsumner  @raymondben  @sckott  @maelle

I want to first and foremost thank the developers for putting so much time and
effort into this package. And I want to thank ROpensci for asking me to do a review.
My reviewer perspective is from that of a regular user.
Specifically because I spend so much of my time teaching scientists, post docs,
grads, undergrads, etc etc, i think a lot about what a typical scientists knows
about data and APIs. Please take my review as a set of suggestions as I greatly
appreciate the effort here! I did not run all of the spell check, etc tests - yet including spelling and such. Instead I spent a good bit of time downloading data and thinking about pain points for users of various skill and knowledge levels.

- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README **I think i generally understand the need but the target audience isn't clear to me (specifically the knowedge / skill of the target audience)**
- [ ] **Installation instructions:** for the development version of package and any non-standard dependencies in README **see notes below**
- [ ] **Vignette(s)** demonstrating major functionality that runs successfully locally **see notes below**
- [ ] **Function Documentation:** for all exported functions in R help **see below**
- [ ] **Examples** for all exported functions in R Help that run successfully locally **see notes below**
- [ ] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).  **i may have missed this - is it there?**

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

**is this being submitted to JOSS? How do I know that?**

Estimated hours spent reviewing: so far i think i've spent about 5 hours reviewing with a few more playing with it when there was talk of it being split

---

### Overall Concept

In general I think this effort is a great one. I like the idea of a data download
utility that can help a user manage and maintain metadata around their downloaded
data. It also supports increased usability and repeatability of a project on the
"left side" of the spectrum - ie getting the data initially. I support and applaud
this effort!

However I have a few general concerns that I think could be addressed. I'd like
to better understand the use cases for this effort and the users technical
knowledge. In general, most
people don't fully understand how large some of these data are. And often the
starting place is to "download everything" without recognizing that for instance
their hard drive is quickly filling up.

Also these days things are moving away from local processing and storage and towards
cloud based activities. Thus, I think this package could provide some guidance
for users BEFORE they download and attempt to archive everything.

For instance,

* Perhaps there are suggestions that ask a user to consider their hard
drive space and storage and computation options before they download things over
a certain size.
* Perhaps there are some guidelines and warning associated with downloading data and the space availability on the drive were the data are being downloaded.
* Perhaps there is some guidance associated with the types of data that might be appropriate to use with this package

I think some of these learning points are important to a successful package because
so many users could quickly make huge API calls and fill up their hard drives unknowingly.

### Use Cases

I'd like to see several clear uses cases for this package that include the user personas
who are accessing the data - ie how much does the audience know for each use case?
particularly as it relates to remote sensing data. I see many potential
challenges for users who are less well versed in working with these data - particularly
in the data size realm and dealing with API calls and associated arguments.

I'd also like to see a bit more documentation and warnings associated with potentially
large requests and associated resource availability on a users machine.

### Package Documentation

In general, the readme and introductory package text is written for someone with
more technical experience and who has probably used API's before even though the
audience as described is "researchers". If the audience is anyone who is looking to access
data and build metadata around it, I think more clear documentation is required
up front about what the package does and how to use it. I provide some specifics
in this regard below!

I also think that all of the function documentation needs to be developed further.

## Setup Feedback

First I installed the package from github. No issues with the installation.

- [ ] Minor but i do suggest adding the `library(bowerbird)` call to the readme vignette

```{r pressure, echo=FALSE}
install.packages("devtools")
library(devtools)
# install instructions missing space
install_github("AustralianAntarcticDivision/bowerbird", build_vignettes = TRUE)

# add this to the vignette just for clarity sake!
library(bowerbird)
```

also minor:

this URL doesn't work for me: http://oceancolor.gsfc.nasa.gov/

### WGET Issues

When I ran through the vignettes, I immediately ran into issues with `wget()`.

```{r eval=FALSE}
sh: wget: command not found
Error in bb_sync(cf) :
  could not find the wget executable. Try the bb_install_wget() function, or install it yourself and ensure that it is on the path
```

This error seems to be focused on Windows users whereas I am on a MAC.
After shedding a few tears, I tried to use brew.

`brew install wget`

This approach tried to force me to update ruby which I wasn't prepared to do given
other running tasks (twitter API extravaganza is underway over here!!). The code below, however worked. I actually don't know why. It must have been the work of a helpful magical make-things-install-right gnome.

`brew install wget --with-libressl`

Notes:

To install wget requires admin access to your laptop. If you are on a
govt or university issued computer that is domain controlled, you may or may not
have access to install wget. I have super user powers on my university issued MAC
ONLY BECAUSE I BEGGED AND PLEADED and downright insisted. And because our
IT  person is awesome and gets it. So I was able to use to to install.
I think wget could be a pain point for many
users who are not advanced. If you want this to be widely used, i'd spend just a bit of
time beefing up the documentation about how to install this on Windows and Mac. And
let people know they may need admin rights...

Errors also should be specific to windows or mac or else you'll get Mac users unknowingly
trying to install windows things. I've seen this happen countless times.

## Suggested Actions

### Style

1. Please follow the tidy styleguide: http://style.tidyverse.org/ All vignettes,
documentation, etc. are missing spaces around function argument assignment
operators and after commas.
  * Some code and comments are well beyond 80 character suggested width for readability.
  * Could a lint test be included in the TravisCI setup to ensure style consistency?

### SETUP

1. Add `library(bowerbird)` to the readme
2. Add more explicit instructions and discussion about `wget()`. Add instructions
for both Mac and Windows and be explicitly about which install path is for which
OS. If you expect to only support advanced users, then they may be able to figure it out.
  *  Adjust error messages surrounding `wget()` to be windows vs mac specific to not mislead users - like me :) .

## Specific Recommendations

### Vignettes

`bb_config()` setups up the data "warehouse" directory where data will be kept.
The `bb_source()` function fails if a directory doesn't exist.

Recommendation: Test whether the input directory exists in `bb_config()`

# setup - a check on the directory here? !dir.exists(~/Documents/data/bowerbird/) or assert...
cf <- bb_config(local_file_root="~/Documents/data/bowerbird/")
```

Next, I tried to define a data source.

`bb_source()`

This function is dense with a lot of powerful arguments. However it is so dense,
that a user without some API experience could get lost and overwhelmed. It would
be very helpful to break down this function into key parts such as:

1. Metadata arguments that support the "database" created by bowerbird
2. API arguments that are required to download the data

Please note that API arguments are not straight forward for most users. If your audience
is really researchers and scientists rather than power users of APIs and technology
experts in general, you will need more documentation.

Notes:

* When I first look at the example below, if I am a less experienced user, I may think that the reference is actually the `source_url`.
* I also may be overwhelmed by number of arguments that I need to parse through.

Here are some suggestions to consider:

1. Provide the most bare bones of bare bones vignettes that only use required arguments. Currently the vignettes use both required and non required arguments.
2. Consider breaking up this mega function into two parts. Maybe one is a list or vector of metadata items and two is a list of api items? Or perhaps there is another clever
way to help a user generate this function call?
3. At a minimum break up the function documentation to have two sections of arguments to make the API vs the metadata explicit
4. I think there is a function passed to an argument here: `method=quote(bb_handler_wget)`. This will confuse a user because when you look up the documentation for this, you see additional arguments.
5. Rework documentation: for instance the earthdata function documentation page, only provides a link to how to register for the website. I am left uncertain about how to actually use the API to get earth data - data. And i've used that website a good bit - granted last year's version of it anyway. :)
5. Vignettes that explain how to identify the `source_url` string would be helpful. It is not clear to me how i would setup access to an earth data download. For instance - what if i wanted to get snow cover from NSIDC? https://n5eil01u.ecs.nsidc.org/VIIRS/VNP10.001/ Or what if i wanted landsat vegetation indices? Vignettes that explain how to access specific datasets (perhaps ones identified in your use cases) would help significantly. On the earth data website it actually took me a while to find any URL's that lead to data. This isn't Bowerbirds problem - it is Nasa's. BUT could this package guide a user through how to find an appropriate source_url for a dataset? Just a could examples...
6. I didnt notice this in the vignettes and perhaps i missed it in the code. is there a summary function that gives you an overview table of all of the data in your "warehouse"?  if there is that would be useful and please add that to the vignette
7. Why is this called bowerbird? The name doesn't fully seem to relate to the library function BUT maybe i need to read more about bowerbirds?!

## Provenance / Flexibility / Adaptability

Next, I tried to get some data from nasa earth data. I have used this site a good
bit to get RS data and noticed that the entire site has changed since last year. Of course it all happened right before my fall class that i use it with!

Here are some questions for the developers:

1. How does the package handle data updates? For instance, last year USGS reprocessed
all of the Landsat 8 data with the create of collections. Thus, all of the data that
I downloaded last year, are different now. Different naming conventions, some changes in files such as the cloud cover layers, etc. Will Bowerbird help a user identify this type of change and guide them through how they want to handle the change (ie re-download data, start a new library with the newly processed data)? How will it do that?
2. The nasa earth data site has also changed considerably since last year. Again, how will bowerbird handle / warn users of changes? We hope that DOI's will persist but with if the new available data have a different naming / metadata structure ... what happens?
3. Many of these datasets are extremely large. Does bowerbird help a user track resource availability on their computers - especially as it's downloading zips and unzipping? Bowerbird could be a beautiful helper package to manage data. It also could allow for less experiences users to make big mistakes in terms of trying to download data that are too large for their computing resources.
  *  Can bowerbird help a user not make huge requests or provide a warning? The beauty of this package is the ease in which you can download and maintain data. The danger is the same - its very easy to kick off some huge calls that may be fill a users hard drive.
4. In general I have a question that i'm sure everyone involved has grappled with: Do we want to encourage people to store data locally? Generally remote sensing data are so large that we may not even want to process it locally. Things are moving towards the cloud and we do want to avoid people starting to build their own mini DAACs on their machines. This is a general comment to consider. This package makes me think a lot about the behaviors i've seen with my students and colleagues at workshops accessing data.

****
Truly great effort and tremendous work has obviously been put into this package! Please take this feedback as an attempt to help with the eye of a regular (non power) user. :)

Also please let me know if I missed some key parts of this review. It is my first for ROpensci!! :)
--Leah
