### 2017-11-07 #139 @maelle
Many thanks for your review @lwasser! 😀 Great feedback! Could you just go over the checklist at the beginning of the refiew template and tell me what was not checked because you didn't have time to do the check vs there was an issue?

@milesmcbain today is the deadline 😉
