### 2017-11-07 #139 @lwasser
oh so sorry @maelle i didn't realize that was a check for me! I was thinking the ✔️ was me signing off on the fact everything was there in the `R` package 🙈 ! :) my mistake. I did go through the list. I'll check it off now!!
