### 2017-11-07 #139 @maelle
No you did it right, checking is when things are as good asyou want them to be but stupidly I had missed the comments in the list, all my fault !
