### 2017-11-07 #139 @raymondben
Thanks @lwasser - plenty of good points in there!
