### 2017-11-08 #139 @MilesMcBain
>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

---

Thanks for your hard work @mdsumner and @raymondben! I applaud the changes you have made since my initial feedback. My first experience with downloading the electoral data could not have been a bigger contrast to last time. It was one of those: "By what magic is this made possible?" moments. I now see `bowerbird` as a powerful tool for managing sources and syncing data from the web. You obviously get great mileage out of it yourselves and I think that is a solid foundation for a great package. The main question mark I have after my review is about how accessible this package will be to the wider community.

This concern arises from a number of areas: I find the documentation is too minimal for many functions. There are still some unconventional R styles used that hindered my ability to understand the package. A level of comfort with `wget` seems assumed in the package that did not hold for me and I am not sure holds widely.

In my comments below I detail my concerns. Please take them for what they are: my attempts to improve the wider appeal of your valuable work.

In the self-titled vignette the use of the `postprocess` argument is only mentioned in passing with respect to decompression. I'd argue it is major functionality with enough usecases that deserves its own subheading.

As I mention in the function documentation section. This package could benefit from a greater breadth of examples with commentary that explains how features of the datasource result in each choice of method flag and postprocessing function.

#### Improve function documentation

In general the function documentation is minimal to the point of being somewhat unhelpful. I learned the most from the examples and source code, but this involved jumping back and forth between many functions. Here are some comments by function:

* `bb_settings` - Documentation does not explain the usecase for this function. It is not clear why it would be needed in addition to `bb_config`. The `oceandata` example made this clearer.
* `bb_validate` - Documentation does not explain why this function does or is used for.
* `bb_cleanup` - The use of the `bb_config` argument is mystifying. The function is specified as part of the config?
* `bb_decompress` - As above.
* `bb_fingerprint()` - Unclear how this function is expected to be used. It gets a mentioned in the 'Data provenance' vignette but that does not explain how the output should be used to detect changes.
* `bb_handler_wget` - Unclear on the effect or valid usecase for `local_dir_only` arg. Did not appear to affect the operation of `bb_sync`.
* `bb_source` - The documentation for this key function is not sufficient for me to be able to use it confidently. `access_function` sounds like it is important but does not appear in any examples and the explanation is vague. The example uses some method flags that are not explained. I think this needs more examples and more commentary of said examples. Explain the nature of the web page and data and how that resulted in the choices of each `method_flag`, and `postprocess` step. Users would benefit from multiple examples of increasing complexity so they can get a sense of the usespace of `bowerbird`.

#### Explain `config` first argument

Either the help or vignettes or both need to clarify the mysterious first argument of `bb_handler_wget`, the `bb_decompress` family, and `bb_cleanup` - the vignette examples and the built-in example data sources use these functions without supplying this first argument, even though it has no default. It was only after digging into the code of `bb_sync` that I saw why this argument exists and how it is populated.

To make this clear you could mention in the help for the arg that it will typically be populated by `bb_sync` and does not need to be specified in `postprocessing` or `method`.

On a related but more general note, you might also cover in a vignette that anyone needing to implement their own handler needs to write a function that must take the special arguments `config`, `verbose`, and `local_dir_only` in addition to any context specific args.

#### Ditch `dplyr` or update to tidyeval
`dplyr` is used very little within this package. So I question its worth as a dependency, given the amount of installation time and number of recursive dependencies it adds. I can imagine `bowebird` being used with automated reproducible pipelines where this may be an issue.

Another issue is that the deprecated 'underscore' verbs (`mutate_`, `group_by_`, `select_`, `rename_`) are sprinkled around this package. For the sake of future package maintenance, these would need to be updated to to the new 'tidyeval' style or removed. Given their very light usage I would recommend just replacing with `base`/`purrr` with a view to removing `dplyr` altogether.

In the case of `bind_rows`, `reduce(., rbind)` can be used guilt-free for the small dataframes you are manipulating here.

#### Unusual usage of `quote`
At first I was confused as to the vignette examples `quote` the functions supplied to the `method` and `postprocess`.

After viewing the examples in various function help files, my understanding now is that you intend `quote` to be the mechanism with which users pass arguments with their specified functions. Is this correct? For me, this is quite unusual and not in keeping with the conventions I am familiar with in both base R or the `tidyverse`. I found a typo in the `bb_cleanup` example that supplied the arguments to `postprocess` in the style of *apply family, I find this telling about how unexpected this pattern is.

I am far from an expert on R internals, but I am not sure if `quote` is safe way to do this since it may produce strange results when variable names are passed instead of literals due to lexical scoping.
I think this needs mechanism needs to be reconsidered. If it is not changed to be more in line with conventional approaches, then it at least needs to be prominently flagged in the documentation.

#### Make some `wget` method flags into arguments

In `bb_config` you did a really nice thing where you created a local function argument (`clobber`) to provide a nicer API to the `wget` clobber arguments. The help for this argument is also quite detailed giving me confidence I could use it correctly if need be. I think this pattern is the bar that the rest of the package should aspire to.

In the self-titled vignette you state:

> You will probably want to specify `--recursive` as part of the `method_flags`, even if the data source doesn't require a recursive download.

before going on to explain the technical reason for this. So now `--recursive` is effectively boilerplate I will dutifully replicate in all my data sources until I get an error - because you said I should. This it not how you create a user friendly API that inspires confidence. I would argue that a much clearer thing to do would be to create an argument for `bb_source`: `recursive = TRUE` with an explanation of the rare case I need to think about changing it in the function documentation.

Another candidate would be `--level=1` which seems to pop up frequently in the examples. If you made it a first class argument of some `bowerbird` function you have prime opportunity to explain it in the documentation. Yet another candidate would be `--accept` - a VERY useful argument.

I am not saying you need to provide a mapping to all `wget` args. From what I gather there are probably less than a half-dozen that will cover > 80% of usecase for your users. You could still have a `more-args` for additional arguments to `wget` if you want. But I think making some commonly used flags into first-class arguments would dramatically improve the appeal of this package. YOU are the experts on `wget` and what you would be saying to us novices is: "Don't worry, I've understood all the wget arguments so you don't have to. These are the ones you care about and this is how you use them". You'd have a much more welcoming package that doesn't have to direct its users to a GNU manpage in the opening paras of its vignette!

The function for these args to live in is probably `bb_handler_wget()`. This might be a bit safer if you move away from the use of `quote` as I suggest above.

As things are there is also some weirdness with the fact that some `wget` flags are handled automatically by the handler eg: `-no-if-modified-since --restrict-file-names=windows --progress=dot:giga` while others are expected to be supplied. What happens if they clash?

#### Misc Comments
* I did not find top-level documentation using `?bowerbird` as per rOpenSci packaging guidelines.
* Widespread usage of `cat(sprintf(...))` for errors/warnings etc. Should replace with `message()/stop()` as per guidelines.
* Test coverage is good overall but sticks mostly to 'happy path'. Untested error paths appear in `sync.R`, `utils.R`, `wget.R`, `postprocess.R` etc. I feel like a jerk for bringing this up. I know how unfun it is to write test cases for these.
