### 2017-11-08 #139 @mdsumner
Yes and thanks!  There's a lot of really helpful feedback here, it's very good.
