### 2017-11-09 87f5eb8
initial bb_wget implementation with explicit parms for some flags
