### 2017-11-09 c440de5
use is_nna_flag rather than just is.flag
