+++ R/wget.R
@@ -12,8 +12,8 @@
 bb_handler_wget <- function(config,verbose=FALSE,local_dir_only=FALSE) {
     assert_that(is(config,"bb_config"))
     assert_that(nrow(bb_data_sources(config))==1)
-    assert_that(is.flag(verbose))
-    assert_that(is.flag(local_dir_only))
+    assert_that(is_nna_flag(verbose))
+    assert_that(is_nna_flag(local_dir_only))
 
     if (local_dir_only)
         return(file.path(bb_settings(config)$local_file_root,directory_from_url(bb_data_sources(config)$source_url)))
@@ -101,8 +101,8 @@ bb_handler_wget <- function(config,verbose=FALSE,local_dir_only=FALSE) {
 bb_wget <- function(url,flags=character(),verbose=FALSE,capture_stdout=FALSE) {
     assert_that(is.string(url))
     assert_that(is.character(flags))
-    assert_that(is.flag(verbose))
-    assert_that(is.flag(capture_stdout))
+    assert_that(is_nna_flag(verbose))
+    assert_that(is_nna_flag(capture_stdout))
     if (tolower(url) %in% c("-h","--help") || identical(tolower(flags),"-h") || identical(tolower(flags),"--help")) {
         out <- sys::exec_internal(bb_find_wget(),args="--help",error=TRUE)
         message(rawToChar(out$stdout))
@@ -123,7 +123,7 @@ bb_wget <- function(url,flags=character(),verbose=FALSE,capture_stdout=FALSE) {
 
 #' Make a wget call
 #'
-#' The wget system call is made using the \code{exec_wait} function from the sys package. Call \code{bb_wget2("--help")} to get a message giving information about wget's command line parameters
+#' The wget system call is made using the \code{exec_wait} function from the sys package. Call \code{bb_wget2("help")} to get a message giving information about wget's command line parameters
 #'
 #' @param url string: the URL to retrieve
 #' @param recursive logical: if true, turn on recursive retrieving
@@ -135,6 +135,8 @@ bb_wget <- function(url,flags=character(),verbose=FALSE,capture_stdout=FALSE) {
 #' @param reject_regex character: as for \code{accept_regex}, but specifying regular expressions to reject
 #' @param exclude_directories character: character vector with one or more entries. Each entry specifies a comma-separated list of directories you wish to exclude from download. Elements may contain wildcards
 #' @param execute character: a character vector with one or more entries. Each entry is an command to be executed. A common use for this is \code{execute=c("robots=off")}, in order to avoid wget's default behaviour of identifying as a robot. Some servers will exclude robots from certain parts of their sites. See \url(https://www.gnu.org/software/wget/manual/wget.html#Robot-Exclusion} for more information about robot exclusion, and \url{https://www.gnu.org/software/wget/manual/wget.html#Wgetrc-Commands} for a full list of commands that can be specified here
+#' @param restrict_file_names character: vector of one of more strings from the set "unix", "windows", "nocontrol", "ascii", "lowercase", and "uppercase". \code{restrict_file_names="windows"} is useful if you are downloading files to be used on both Windows and Unix file systems. See \url{https://www.gnu.org/software/wget/manual/wget.html#index-Windows-file-names} for more information on this parameter
[…]
