### 2017-11-09 17047db
no, don't use is_nna_flag, just use compound assert_that calls, gives better diagnostic output
