+++ R/sync.R
@@ -13,8 +13,8 @@
 bb_sync <- function(config,create_root=FALSE,verbose=TRUE,catch_errors=TRUE) {
     ## general synchronization handler
     assert_that(is(config,"bb_config"))
-    assert_that(is_nna_flag(create_root))
-    assert_that(is_nna_flag(verbose))
+    assert_that(is.flag(create_root),!is.na(create_root))
+    assert_that(is.flag(verbose),!is.na(verbose))
     if (nrow(bb_data_sources(config))<1) {
         warning("config has no data sources: nothing for bb_sync to do")
         return(invisible(NULL))
