### 2017-11-09 7a1c009
helper to detect OS type
