+++ R/utils.R
@@ -266,3 +266,23 @@ do_decompress_files <- function(method,files,overwrite=TRUE) {
     all_OK
 }
 
+
+## adapted from http://conjugateprior.org/2015/06/identifying-the-os-from-r/
+get_os <- function() {
+    if (.Platform$OS.type=="windows") return("windows")
+    sysinf <- Sys.info()
+    if (!is.null(sysinf)){
+        os <- sysinf["sysname"]
+        if (tolower(os)=="darwin")
+            os <- "osx"
+    } else {
+        os <- .Platform$OS.type
+        if (grepl("^darwin", R.version$os,ignore.case=TRUE))
+            os <- "osx"
+        if (grepl("linux-gnu", R.version$os,ignore.case=TRUE))
+            os <- "linux"
+    }
+    if (!os %in% c("windows","linux","unix","osx"))
+        stop("unknown operating system: ",os)
+    tolower(os)
+}
