### 2017-11-09 225e602
better error messages when wget is not found
