### 2017-11-10 007e2e0
more robust bb_find_wget: throw error on failure by default, better error messages, various other improvements
