+++ R/wget.R
@@ -216,12 +216,12 @@ bb_wget2 <- function(url,recursive=TRUE,level=1,wait=0,accept,reject,accept_rege
         extra_flags <- flags_to_charvec(extra_flags) ## will split string, or replace NA/"" with empty character vector
         flags <- c(flags,extra_flags)
         if (verbose) cat(sprintf(" executing wget %s %s\n",paste(flags,collapse=" "),url))
-        #if (capture_stdout) {
-        #    sys::exec_internal(bb_find_wget(),args=c(flags,url),error=FALSE)
-        #} else {
-        #    status <- sys::exec_wait(bb_find_wget(),args=c(flags,url),std_out=TRUE,std_err=TRUE)
-        #    list(status=status)
-        #}
+        if (capture_stdout) {
+            sys::exec_internal(bb_find_wget(),args=c(flags,url),error=FALSE)
+        } else {
+            status <- sys::exec_wait(bb_find_wget(),args=c(flags,url),std_out=TRUE,std_err=TRUE)
+            list(status=status)
+        }
     }
 }
 
@@ -243,7 +243,7 @@ bb_wget2 <- function(url,recursive=TRUE,level=1,wait=0,accept,reject,accept_rege
 #'
 #' @export
 bb_install_wget <- function() {
-    if (tolower(.Platform$OS.type)!="windows")
+    if (get_os()!="windows")
         stop("bb_install_wget only supports windows platforms")
     ## NOTE, could also use e.g. https://github.com/r-lib/rappdirs to find this directory
     path <- Sys.getenv("APPDATA")
@@ -270,9 +270,10 @@ bb_install_wget <- function() {
 #'
 #' @references https://eternallybored.org/misc/wget/current/wget.exe
 #'
-#' @param install logical: install the executable if it is not found? (Windows only)
+#' @param install logical: attempt to install the executable if it is not found? (Windows only)
+#' @param error logical: if wget is not found, raise an error if wget is not found. If FALSE, do not raise an error but return NULL
 #'
-#' @return the path to the wget executable, or NULL if it was not found
+#' @return the path to the wget executable, or (if error is FALSE) NULL if it was not found
 #'
 #' @examples
 #' \dontrun{
@@ -283,33 +284,57 @@ bb_install_wget <- function() {
 #' @seealso \code{\link{bb_install_wget}}
 #'
 #' @export
-bb_find_wget <- function(install=FALSE) {
+bb_find_wget <- function(install=FALSE,error=TRUE) {
     assert_that(is.flag(install),!is.na(install))
     bb_opts <- getOption("bowerbird")
     if (!is.null(bb_opts)) {
         if (!is.null(bb_opts$wget_exe)) {
-            ## already set, just return this
-            return(bb_opts$wget_exe)
+            ## already set
+            ## check that it actually exists
+            if (file.exists(bb_opts$wget_exe)) {
+                ## ok
+                return(bb_opts$wget_exe)
+            } else {
+                ## it's disappeared somehow
+                bb_opts$wget_exe <- NULL
+                options(bowerbird=bb_opts)
+                ## try again
+                return(bb_find_wget(install=install,error=error))
+            }
         }
     } else {
         bb_opts <- list()
     }
+    my_os <- get_os()
     if (wget_test("wget")) {
         myexe <- unname(Sys.which("wget"))
     } else {
-        if (.Platform$OS.type=="windows") {
+        if (my_os=="windows") {
             myexe <- file.path(Sys.getenv("APPDATA"),"bowerbird","wget.exe")
             if (!wget_test(myexe)) {
                 if (!install) {
+                    if (error) stop("could not find the wget executable.\n Try the bb_install_wget() function, or install wget yourself and ensure that it is on the system path")
[…]
