### 2017-11-10 fcd6326
remove wget checking, now done in bb_find_wget
