### 2017-11-10 778de40
updates for new bb_find_wget behaviour
