+++ tests/testthat.R
@@ -3,10 +3,10 @@ library(dplyr)
 library(bowerbird)
 
 if (.Platform$OS.type=="windows") {
-    if (is.null(bb_find_wget())) {
+    if (is.null(bb_find_wget(error=FALSE))) {
         warning("At init of testing: could not find wget executable, installing.\n")
         bb_install_wget()
-        if (is.null(bb_find_wget())) stop("could not install wget execitable")
+        tmp <- bb_find_wget() ## will throw error if wget not found
     }
 }
 test_check("bowerbird")
