### 2017-11-10 91dec7f
minor doc tweaks
