### 2017-11-10 1f1c4c0
minor doc tweaks
