### 2017-11-10 68a8a37
grr fix case sensitivity in os check
