+++ R/example_sources.R
@@ -58,7 +58,7 @@ bb_example_sources <- function() {
             name="Oceandata SeaWiFS Level-3 mapped monthly 9km chl-a",
             id="SeaWiFS_L3m_MO_CHL_chlor_a_9km",
             description="Monthly remote-sensing chlorophyll-a from the SeaWiFS satellite at 9km spatial resolution",
-            reference= "http://oceancolor.gsfc.nasa.gov/",
+            reference= "https://oceancolor.gsfc.nasa.gov/",
             citation="See https://oceancolor.gsfc.nasa.gov/citations",
             license="Please cite",
             method=quote(bb_handler_oceandata),
