### 2017-11-30 3dfcca6
half-implemented move to wget-with-flags-as-function-parameters
