+++ R/wget.R
@@ -35,7 +35,7 @@ bb_handler_wget <- function(config,verbose=FALSE,local_dir_only=FALSE) {
                                         #}
     ## add flags for clobber behaviour
     if (!is.null(cfrow$clobber) && !is.na(cfrow$clobber) && cfrow$clobber %in% c(0,1)) {
-        if (cfrow$clobber==0) {
+        if (cfrow$clobber<1) {
             this_flags <- resolve_wget_clobber_flags(this_flags,"--no-clobber")
         } else {
             this_flags <- resolve_wget_clobber_flags(this_flags,"--timestamping")
@@ -72,12 +72,11 @@ bb_handler_wget <- function(config,verbose=FALSE,local_dir_only=FALSE) {
             ## no object returned - this happens if process is interrupted by the user on unix
             ok <- as.logical(NA)
         } else {
-            ok <- syscall_obj$status==0
+            ok <- !syscall_obj$status ## status code of 0 means OK
         }
     }
     ok
 }
-
 
 #' Make a wget call
 #'
@@ -121,6 +120,91 @@ bb_wget <- function(url,flags=character(),verbose=FALSE,capture_stdout=FALSE) {
 }
 
 
+#' Mirror an external data source using the wget utility
+#'
+#' @param config bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source
+#' @param verbose logical: if TRUE, provide additional progress output
+#' @param local_dir_only logical: if TRUE, just return the local directory into which files from this data source would be saved
+#'
+#' @return the directory if local_dir_only is TRUE, otherwise TRUE on success
+#'
+#' @seealso \code{\link{bb_wget}}
+#'
+#' @export
+bb_handler_wget2 <- function(config,verbose=FALSE,local_dir_only=FALSE) {
+    assert_that(is(config,"bb_config"))
+    assert_that(nrow(bb_data_sources(config))==1)
+    assert_that(is.flag(verbose),!is.na(verbose))
+    assert_that(is.flag(local_dir_only),!is.na(local_dir_only))
+
+    if (local_dir_only)
+        return(file.path(bb_settings(config)$local_file_root,directory_from_url(bb_data_sources(config)$source_url)))
+
+    cfrow <- bb_settings_to_cols(config)
+    this_flags <- cfrow$method_flags[[1]]
+    ## add wget_global_flags
+    gflags <- cfrow$wget_global_flags[[1]]
+    if (length(gflags)>0)
+        this_flags <- c(this_flags,gflags)
+
+    ## proxy-user and proxy-password flags
+    ## this needs to decide whether it should use http_proxy_user or ftp_proxy_user info - exclude for now
+    #if (!grepl("proxy-user=",cfrow$wget_flags)) {
+    #    cfrow$wget_flags=paste(cfrow$wget_flags,paste0("--proxy-user=",cfrow$http_proxy_user),sep=" ")
+    #}
+    #if (!grepl("proxy-password=",cfrow$wget_flags)) {
+    #    cfrow$wget_flags=paste(cfrow$wget_flags,paste0("--proxy-password=",cfrow$http_proxy_password),sep=" ")
+                                        #}
+    ## add flags for clobber behaviour
+    if (!is.null(cfrow$clobber) && !is.na(cfrow$clobber) && cfrow$clobber %in% c(0,1)) {
+        if (cfrow$clobber<1) {
+            this_flags <- merge_wget_flags(this_flags,list(no_clobber=TRUE))
+        } else {
+            this_flags <- merge_wget_flags(this_flags,list(timestamping=TRUE))
+        }
+    }
+    ## add user, password flags
[…]
