+++ R/source.R
@@ -96,3 +96,99 @@ bb_source <- function(id,name,description=NA_character_,reference,source_url,cit
         collection_size=if (assert_that(is.numeric(collection_size) || is.na(collection_size))) collection_size)
 }
 
+
+#' Define a data source
+#'
+#' @param id string: (required) a unique identifier of the data source. If the data source has a DOI, use that. Otherwise, if the original data provider has an identifier for this dataset, that is probably a good choice here (include the data version number if there is one). The ID should be something that changes when the data set changes (is updated). A DOI is ideal for this
+#' @param name string: (required) a unique name for the data source. This should be a human-readable but still concise name
+#' @param description string: a description of the data source
+#' @param reference string: (required) URL to the metadata record or home page of the data source
+#' @param source_url character vector: one or more source URLs. Required for \code{bb_handler_wget}, although some \code{method} functions might not require one
+#' @param citation string: (required) details of the citation for the data source
+#' @param license string: (required) description of the license. For standard licenses (e.g. creative commons) include the license descriptor ("CC-BY", etc)
+#' @param comment string: comments about the data source. If only part of the original data collection is mirrored, mention that here
+#' @param method function, call, or symbol: (required) the function that handles the synchronisation process for this data source
+#' @param method_flags character vector: flags to pass to the method. If method is \code{bb_handler_wget}, these are wget flags. Run \code{bb_wget("--help")} to get help on these flags
+#' @param postprocess function, call, symbol, or list thereof: functions to apply after synchronisation has completed. If NULL or an empty list, no postprocessing will be applied
+#' @param authentication_note string: if authentication is required in order to access this data source, make a note of the process (include a URL to the registration page, if possible)
+#' @param user string: username, if required
+#' @param password string: password, if required
+#' @param access_function string: name of the R function that can be used to read these data
+#' @param data_group string: the name of the group to which this data source belongs. Useful for arranging sources in terms of thematic areas
+#' @param collection_size numeric: approximate disk space (in GB) used by the data collection, if known. If the data are supplied as compressed files, this size should reflect the disk space used after decompression. If the data_source definition contains multiple source_url entries, this size should reflect the overall disk space used by all combined
+#' @param warn_empty_auth logical: if TRUE, issue a warning if the data source requires authentication (authentication_note is not NA) but user and password have not been provided
+#'
+#' @return tibble
+#'
+#' @seealso \code{\link{bb_config}}
+#'
+#' @examples
+#'
+#' my_source <- bb_source(
+#'    id="gshhg_coastline",
+#'    name="GSHHG coastline data",
+#'    description="A Global Self-consistent, Hierarchical, High-resolution Geography Database",
+#'    reference= "http://www.soest.hawaii.edu/pwessel/gshhg",
+#'    citation="Wessel, P., and W. H. F. Smith, A Global Self-consistent, Hierarchical,
+#'      High-resolution Shoreline Database, J. Geophys. Res., 101, 8741-8743, 1996",
+#'    source_url="ftp://ftp.soest.hawaii.edu/gshhg/*",
+#'    license="",
+#'    comment="",
+#'    method=quote(bb_handler_wget),
+#'    method_flags=c("--recursive","--level=1","--accept=*bin*.zip,README.TXT"),
+#'    postprocess=quote(bb_unzip),
+#'    collection_size=0.6)
+#'
+#' cf <- bb_config("/my/repo/root")
+#' cf <- bb_add(cf,my_source)
+#'
+#' @export
[…]

[…]
