### 2017-11-30 e75a97d
use match.fun instead of get_function_from_method
