### 2017-11-30 e23e696
comment out earthdata example source until earthdata handler updated
