+++ R/example_sources.R
@@ -37,7 +37,7 @@ bb_example_sources <- function() {
             access_function="readsst",
             collection_size=0.9,
             data_group="Sea surface temperature"),
-        bb_source(
+        bb_source2(
             name="CMEMS global gridded SSH reprocessed (1993-ongoing)",
             id="SEALEVEL_GLO_PHY_L4_REP_OBSERVATIONS_008_047",
             description="For the Global Ocean - Multimission altimeter satellite gridded sea surface heights and derived variables computed with respect to a twenty-year mean. Previously distributed by Aviso+, no change in the scientific content. All the missions are homogenized with respect to a reference mission which is currently OSTM/Jason-2.\nVARIABLES\n- sea_surface_height_above_sea_level (SSH)\n- surface_geostrophic_eastward_sea_water_velocity_assuming_sea_level_for_geoid (UVG)\n- surface_geostrophic_northward_sea_water_velocity_assuming_sea_level_for_geoid (UVG)\n- sea_surface_height_above_geoid (SSH)\n- surface_geostrophic_eastward_sea_water_velocity (UVG)\n- surface_geostrophic_northward_sea_water_velocity (UVG)",
@@ -45,8 +45,8 @@ bb_example_sources <- function() {
             citation="In case of any publication, the Licensee will ensure credit the Copernicus Marine Service in the following manner: \"This study has been conducted using E.U. Copernicus Marine Service Information\"",
             source_url=c("ftp://ftp.sltac.cls.fr/Core/SEALEVEL_GLO_PHY_L4_REP_OBSERVATIONS_008_047/dataset-duacs-rep-global-merged-allsat-phy-l4-v3/"),
             license="See http://marine.copernicus.eu/services-portfolio/service-commitments-and-licence/",
-            method=quote(bb_handler_wget),
-            method_flags=c("--recursive","--level=3","--follow-ftp","--no-parent"),
+            method=quote(bb_handler_wget2),
+            method_flags=list(recursive=TRUE,level=3,no_parent=TRUE),
             postprocess=quote(bb_gunzip),
             authentication_note="Copernicus Marine login required, see http://marine.copernicus.eu/services-portfolio/register-now/",
             user="",
@@ -54,35 +54,35 @@ bb_example_sources <- function() {
             access_function="readssh",
             collection_size=310,
             data_group="Altimetry",warn_empty_auth=FALSE),
-        bb_source(
+        bb_source2(
             name="Oceandata SeaWiFS Level-3 mapped monthly 9km chl-a",
             id="SeaWiFS_L3m_MO_CHL_chlor_a_9km",
             description="Monthly remote-sensing chlorophyll-a from the SeaWiFS satellite at 9km spatial resolution",
             reference= "https://oceancolor.gsfc.nasa.gov/",
             citation="See https://oceancolor.gsfc.nasa.gov/citations",
             license="Please cite",
-            method=quote(bb_handler_oceandata),
-            method_flags=c("search=S*L3m_MO_CHL_chlor_a_9km.nc"),
+            method=quote(bb_handler_oceandata2),
+            method_flags=list(search="S*L3m_MO_CHL_chlor_a_9km.nc"),
             postprocess=NULL,
             collection_size=7.2,
             data_group="Ocean colour"),
-        bb_source(
-            name="Sea Ice Trends and Climatologies from SMMR and SSM/I-SSMIS, Version 2",
-            id="10.5067/EYICLBOAAJOU",
[…]
