### 2017-11-30 b352732
change handler syntax to take method-specific args as parameters instead of buried in config object
