+++ R/oceandata.R
@@ -110,6 +110,120 @@ bb_handler_oceandata <- function(config,verbose=FALSE,local_dir_only=FALSE) {
 }
 
 
+#' Handler for oceandata data sources
+#'
+#' @references https://oceandata.sci.gsfc.nasa.gov/
+#' @param config bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source
+#' @param search string: (required) the search string to pass to the oceancolor file searcher (https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi)
+#' @param dtype string: (optional) the data type (e.g. "L3m") to pass to the oceancolor file searcher (https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi)
+#' @param verbose logical: if TRUE, provide additional progress output
+#' @param local_dir_only logical: if TRUE, just return the local directory into which files from this data source would be saved
+#'
+#' @return the directory if local_dir_only is TRUE, otherwise TRUE on success
+#'
+#' @export
+bb_handler_oceandata2 <- function(config,verbose=FALSE,local_dir_only=FALSE,search,dtype) {
+    ## oceandata synchronisation handler
+
+    ## oceandata provides a file search interface, e.g.:
+    ## wget -q --post-data="cksum=1&search=A2002*DAY_CHL_chlor*9km*" -O - https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi
+    ## or
+    ## wget -q --post-data="dtype=L3b&cksum=1&search=A2014*DAY_CHL.*" -O - https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi
+    ## returns list of files and SHA1 checksum for each file
+    ## each file can be retrieved from https://oceandata.sci.gsfc.nasa.gov/cgi/getfile/filename
+
+    ## expect that config$data_sources$method_flags will contain the search and dtype components of the post string
+    ##  i.e. "search=...&dtype=..." in "dtype=L3m&addurl=1&results_as_file=1&search=A2002*DAY_CHL_chlor*9km*"
+    ##  or just include the data type in the search pattern e.g. "search=A2002*L3m_DAY_CHL_chlor*9km*
+
+    assert_that(is(config,"bb_config"))
+    assert_that(nrow(bb_data_sources(config))==1)
+    ##assert_that(is.list(bb_data_sources(config)$method_flags))
+    ##assert_that(is.character(bb_data_sources(config)$method_flags[[1]]))
+    assert_that(is.flag(verbose),!is.na(verbose))
+    assert_that(is.flag(local_dir_only),!is.na(local_dir_only))
+    assert_that(is.string(search),nzchar(search))
+    if (!missing(dtype)) assert_that(is.string(dtype),nzchar(dtype))
+
+    ##method_flags <- bb_data_sources(config)$method_flags[[1]]
+    ##assert_that(!is.null(method_flags$search),is.string(method_flags$search),nzchar(method_flags$search))
+    this_att <- bb_settings(config)
+    if (local_dir_only) {
+        ## highest-level dir
+        out <- "oceandata.sci.gsfc.nasa.gov"
+        ## refine by platform
+        this_search_spec <- search##sub("search=","",method_flags)
+        this_platform <- oceandata_platform_map(substr(this_search_spec,1,1))
+        if (nchar(this_platform)>0) out <- file.path(out,this_platform)
+        if (grepl("L3m",this_search_spec)) {
+            out <- file.path(out,"Mapped")
+        } else if (grepl("L3",this_search_spec)) {
+            out <- file.path(out,"L3BIN")
+        }
+        return(file.path(this_att$local_file_root,out))
+    }
+    tries <- 0
+    while (tries<3) {
+        ## sometimes this takes a couple of attempts!
[…]
