+++ man/bb_handler_oceandata2.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/oceandata.R
+\name{bb_handler_oceandata2}
+\alias{bb_handler_oceandata2}
+\title{Handler for oceandata data sources}
+\usage{
+bb_handler_oceandata2(config, verbose = FALSE, local_dir_only = FALSE,
+  search, dtype)
+}
+\arguments{
+\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source}
+
+\item{verbose}{logical: if TRUE, provide additional progress output}
+
+\item{local_dir_only}{logical: if TRUE, just return the local directory into which files from this data source would be saved}
+
+\item{search}{string: (required) the search string to pass to the oceancolor file searcher (https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi)}
+
+\item{dtype}{string: (optional) the data type (e.g. "L3m") to pass to the oceancolor file searcher (https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi)}
+}
+\value{
+the directory if local_dir_only is TRUE, otherwise TRUE on success
+}
+\description{
+Handler for oceandata data sources
+}
+\references{
+https://oceandata.sci.gsfc.nasa.gov/
+}
