+++ man/bb_handler_wget2.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/wget.R
+\name{bb_handler_wget2}
+\alias{bb_handler_wget2}
+\title{Mirror an external data source using the wget utility}
+\usage{
+bb_handler_wget2(config, verbose = FALSE, local_dir_only = FALSE, ...)
+}
+\arguments{
+\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source}
+
+\item{verbose}{logical: if TRUE, provide additional progress output}
+
+\item{local_dir_only}{logical: if TRUE, just return the local directory into which files from this data source would be saved}
+
+\item{...}{: extra arguments passed through to \code{bb_wget2}}
+}
+\value{
+the directory if local_dir_only is TRUE, otherwise TRUE on success
+}
+\description{
+Mirror an external data source using the wget utility
+}
+\seealso{
+\code{\link{bb_wget}}
+}

[…]
