### 2017-11-30 9f17b9f
reduce scope of test, now using match.fun
