+++ tests/testthat/test_utils.R
@@ -3,8 +3,8 @@ context("test bowerbird utilities")
 test_that("check_method_is works",{
     expect_true(check_method_is("bb_handler_wget",bb_handler_wget)) ## function name as string
     expect_true(check_method_is(bb_handler_wget,bb_handler_wget)) ## function
-    expect_true(check_method_is(enquote(bb_handler_wget),bb_handler_wget))
-    expect_true(check_method_is(quote(bb_handler_wget()),bb_handler_wget))
+    ##expect_true(check_method_is(enquote(bb_handler_wget),bb_handler_wget))
+    ##expect_true(check_method_is(quote(bb_handler_wget()),bb_handler_wget))
     expect_true(check_method_is(quote(bb_handler_wget),bb_handler_wget)) ## symbol
 })
 
