### 2017-11-30 bc296f4
temporarily re-add backwards-compatibility of bb_data_source_dir to work with old handlers
