### 2017-11-30 ce675cc
change data source def used for testing
