### 2017-11-30 125214a
update earthdata handler to new arg format
