+++ NAMESPACE
@@ -16,6 +16,7 @@ export(bb_find_wget)
 export(bb_fingerprint)
 export(bb_gunzip)
 export(bb_handler_earthdata)
+export(bb_handler_earthdata2)
 export(bb_handler_oceandata)
 export(bb_handler_oceandata2)
 export(bb_handler_wget)

[…]
