### 2017-11-30 13cd616
remove quotes from wget post call
