+++ R/example_sources.R
@@ -31,7 +31,7 @@ bb_example_sources <- function() {
             citation="NOAA_OI_SST_V2 data provided by the NOAA/OAR/ESRL PSD, Boulder, Colorado, USA, from their web site at http://www.esrl.noaa.gov/psd/",
             source_url=c("ftp://ftp.cdc.noaa.gov/Datasets/noaa.oisst.v2/*"),
             license="Please cite",
-            method=quote(bb_handler_wget2),
+            method="bb_handler_wget2",
             method_flags=list(recursive=TRUE,level=1,no_parent=TRUE),
             postprocess=NULL,
             access_function="readsst",
@@ -45,7 +45,7 @@ bb_example_sources <- function() {
             citation="In case of any publication, the Licensee will ensure credit the Copernicus Marine Service in the following manner: \"This study has been conducted using E.U. Copernicus Marine Service Information\"",
             source_url=c("ftp://ftp.sltac.cls.fr/Core/SEALEVEL_GLO_PHY_L4_REP_OBSERVATIONS_008_047/dataset-duacs-rep-global-merged-allsat-phy-l4-v3/"),
             license="See http://marine.copernicus.eu/services-portfolio/service-commitments-and-licence/",
-            method=quote(bb_handler_wget2),
+            method="bb_handler_wget2",
             method_flags=list(recursive=TRUE,level=3,no_parent=TRUE),
             postprocess=quote(bb_gunzip),
             authentication_note="Copernicus Marine login required, see http://marine.copernicus.eu/services-portfolio/register-now/",
@@ -61,7 +61,7 @@ bb_example_sources <- function() {
             reference= "https://oceancolor.gsfc.nasa.gov/",
             citation="See https://oceancolor.gsfc.nasa.gov/citations",
             license="Please cite",
-            method=quote(bb_handler_oceandata2),
+            method="bb_handler_oceandata2",
             method_flags=list(search="S*L3m_MO_CHL_chlor_a_9km.nc"),
             postprocess=NULL,
             collection_size=7.2,
@@ -75,7 +75,7 @@ bb_example_sources <- function() {
             source_url=c("https://daacdata.apps.nsidc.org/pub/DATASETS/nsidc0192_seaice_trends_climo_v2/"),
             license="Please cite, see http://nsidc.org/about/use_copyright.html",
             authentication_note="Requires Earthdata login, see https://urs.earthdata.nasa.gov/. Note that you will also need to authorize the application 'nsidc-daacdata' (see 'My Applications' at https://urs.earthdata.nasa.gov/profile)",
-            method=quote(bb_handler_earthdata2),
+            method="bb_handler_earthdata2",
             method_flags=list(recursive=TRUE,level=4,no_parent=TRUE,relative=TRUE),
             user="",
             password="",
@@ -90,7 +90,7 @@ bb_example_sources <- function() {
             source_url=c("https://www.ngdc.noaa.gov/mgg/greatlakes/superior/data/"),
             citation="Publisher: DOC/NOAA/NESDIS/NCEI > National Centers for Environmental Information, NESDIS, NOAA, U.S. Department of Commerce",
             license="https://www.ngdc.noaa.gov/ngdcinfo/privacy.html#copyright",
-            method=quote(bb_handler_wget2),
+            method="bb_handler_wget2",
             method_flags=list(recursive=TRUE,level=2,accept_regex="/netcdf/",reject="index.html*"),
             comment="Only the netcdf format data are retrieved here - adjust the accept parameter in the method_flags to get other formats",
             postprocess=quote(bb_gunzip),
@@ -104,7 +104,7 @@ bb_example_sources <- function() {
[…]
