### 2017-11-30 89650c4
change postprocess spec syntax to list
