### 2017-11-30 6be6c4d
extra bits for AEC data example source
