### 2017-11-30 004a391
clean up trace writes
