### 2017-11-30 b371a69
decode wget status codes
