### 2017-11-30 b596999
adjust postprocess tests for new syntax
