+++ tests/testthat/test_postprocess.R
@@ -3,7 +3,7 @@ context("postprocessing")
 test_that("decompressing zip files works",{
     skip_on_cran()
     skip_on_appveyor() ## fails for unknown reasons
-    my_source <- bb_source(
+    my_source <- bb_source2(
         name="Bowerbird test data",
         id="bbtest-v0.1",
         description="These are just some trivial test files provided with the bowerbird package.",
@@ -11,12 +11,12 @@ test_that("decompressing zip files works",{
         citation="No citation needed.",
         source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data.zip",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data.zip",
         license="MIT",
-        method=quote(bb_handler_wget),
-        method_flags=c("--recursive","--level=1","-e","robots=off"),
-        postprocess=quote(bb_unzip))
+        method="bb_handler_wget2",
+        method_flags=list(recursive=TRUE,level=1,execute=c("robots=off")),
+        postprocess=list("bb_unzip"))
 
     temp_root <- tempdir()
-    cf <- bb_add(bb_config(local_file_root=temp_root,clobber=2),my_source)
+    cf <- bb_add(bb_config2(local_file_root=temp_root,clobber=2),my_source)
     bb_sync(cf)
 
     fp <- bb_data_source_dir(cf)
@@ -29,7 +29,7 @@ test_that("decompressing zip files works",{
 test_that("decompressing gzipped files works",{
     skip_on_cran()
     skip_on_appveyor() ## fails for unknown reasons
-    my_source <- bb_source(
+    my_source <- bb_source2(
         name="Bowerbird test data",
         id="bbtest-v0.1",
         description="These are just some trivial test files provided with the bowerbird package.",
@@ -37,12 +37,12 @@ test_that("decompressing gzipped files works",{
         citation="No citation needed.",
         source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data_was_gzipped.csv.gz",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data_was_gzipped.csv.gz",
         license="MIT",
-        method=quote(bb_handler_wget),
-        method_flags=c("--recursive","--level=1","-e","robots=off"),
-        postprocess=quote(bb_gunzip))
+        method="bb_handler_wget2",
+        method_flags=list(recursive=TRUE,level=1,execute=c("robots=off")),
+        postprocess=list("bb_gunzip"))
 
     temp_root <- tempdir()
-    cf <- bb_add(bb_config(local_file_root=temp_root,clobber=2),my_source)
+    cf <- bb_add(bb_config2(local_file_root=temp_root,clobber=2),my_source)
     bb_sync(cf)
 
     fp <- bb_data_source_dir(cf)
@@ -55,7 +55,7 @@ test_that("decompressing gzipped files works",{
 test_that("decompressing bzipped files works",{
     skip_on_cran()
     skip_on_appveyor() ## fails for unknown reasons
-    my_source <- bb_source(
+    my_source <- bb_source2(
         name="Bowerbird test data",
         id="bbtest-v0.1",
         description="These are just some trivial test files provided with the bowerbird package.",
@@ -63,12 +63,12 @@ test_that("decompressing bzipped files works",{
         citation="No citation needed.",
[…]
