### 2017-12-01 e72b19c
move across to method=list(fun,args) syntax
