+++ R/sync.R
@@ -114,9 +114,10 @@ do_sync_repo <- function(this_dataset,create_root,verbose,settings) {
         if (verbose) cat(sprintf("done.\n"))
     }
     ## run the method
-    mth <- match.fun(bb_data_sources(this_dataset)$method[[1]])
+    mth <- tryCatch(match.fun(bb_data_sources(this_dataset)$method[[1]][[1]]),
+                    error=function(e) bb_data_sources(this_dataset)$method[[1]]) ## new and old options, todo: remove old option once transitioned
     if (check_method_is(mth,bb_handler_wget2) || check_method_is(mth,bb_handler_oceandata2) || check_method_is(mth,bb_handler_earthdata2)) {
-        ok <- do.call(mth,c(list(config=this_dataset,verbose=verbose),bb_data_sources(this_dataset)$method_flags[[1]]))
+        ok <- do.call(mth,c(list(config=this_dataset,verbose=verbose),bb_data_sources(this_dataset)$method[[1]][-1]))
     } else {
         ok <- do.call(mth,list(config=this_dataset,verbose=verbose))
     }
