+++ tests/testthat/test_postprocess.R
@@ -11,8 +11,7 @@ test_that("decompressing zip files works",{
         citation="No citation needed.",
         source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data.zip",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data.zip",
         license="MIT",
-        method="bb_handler_wget2",
-        method_flags=list(recursive=TRUE,level=1,execute=c("robots=off")),
+        method=list("bb_handler_wget2",recursive=TRUE,level=1,execute=c("robots=off")),
         postprocess=list("bb_unzip"))
 
     temp_root <- tempdir()
@@ -37,8 +36,7 @@ test_that("decompressing gzipped files works",{
         citation="No citation needed.",
         source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data_was_gzipped.csv.gz",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data_was_gzipped.csv.gz",
         license="MIT",
-        method="bb_handler_wget2",
-        method_flags=list(recursive=TRUE,level=1,execute=c("robots=off")),
+        method=list("bb_handler_wget2",recursive=TRUE,level=1,execute=c("robots=off")),
         postprocess=list("bb_gunzip"))
 
     temp_root <- tempdir()
@@ -63,8 +61,7 @@ test_that("decompressing bzipped files works",{
         citation="No citation needed.",
         source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data_was_bzipped.csv.bz2",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data_was_bzipped.csv.bz2",
         license="MIT",
-        method="bb_handler_wget2",
-        method_flags=list(recursive=TRUE,level=1,execute=c("robots=off")),
+        method=list("bb_handler_wget2",recursive=TRUE,level=1,execute=c("robots=off")),
         postprocess=list("bb_bunzip2"))
 
     temp_root <- tempdir()
@@ -89,8 +86,7 @@ test_that("decompressing Z-compressed files works",{
         citation="No citation needed.",
         source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/20170822.nc.Z",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/20170822.nc.Z",
         license="MIT",
-        method="bb_handler_wget2",
-        method_flags=list(recursive=TRUE,level=1,execute=c("robots=off")),
+        method=list("bb_handler_wget2",recursive=TRUE,level=1,execute=c("robots=off")),
         postprocess=list("bb_uncompress"))
 
     temp_root <- tempdir()
