+++ man/bb_decompress2.Rd
@@ -1,0 +1,44 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/postprocess.R
+\name{bb_decompress2}
+\alias{bb_decompress2}
+\title{Postprocessing: decompress zip, gz, bz2, Z files and optionally delete the compressed copy}
+\usage{
+bb_decompress2(config, file_list_before, file_list_after, verbose = TRUE,
+  method, delete = FALSE)
+}
+\arguments{
+\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source}
+
+\item{file_list_before}{data.frame: files present in the directory before synchronising, as returned by \code{file.info}}
+
+\item{file_list_after}{data.frame: files present in the directory after synchronising, as returned by \code{file.info}}
+
+\item{verbose}{logical: if TRUE, provide additional progress output}
+
+\item{method}{string: one of "unzip","gunzip","bunzip2","decompress"}
+
+\item{delete}{logical: delete the zip files after extracting their contents?}
+
+\item{...}{: parameters passed to bb_decompress2}
+}
+\value{
+TRUE on success
+}
+\description{
+This function is not intended to be called directly, but instead can be specified as a postprocessing step to apply to a data source. \code{bb_unzip}, \code{bb_gunzip}, \code{bb_bunzip2}, and \code{bb_uncompress} are convenience wrappers around \code{bb_decompress} that specify the method.
+The dots argument indicates additional arguments that are passed to \code{bb_decompress} when called by \code{bb_sync}. These include parameters named \code{file_list_before} and \code{file_list_after}, which are data.frames as returned by \code{file.info}, listing the files present in the target directory before and after synchronising. These are used if delete=TRUE.
+}
+\examples{
+\dontrun{
+  ## decompress .zip files after synchronisation but keep zip files intact
+  my_source <- bb_source(...,postprocess=list("bb_unzip"))
+
+  ## decompress .zip files after synchronisation and delete zip files
+  my_source <- bb_source(...,postprocess=list(list("bb_unzip",delete=TRUE)))
+}
+
+}
+\seealso{
+\code{\link{bb_source}} \code{\link{bb_config}} \code{\link{bb_cleanup}}
+}
