+++ NAMESPACE
@@ -6,11 +6,13 @@ export("bb_settings<-")
 export(bb_add)
 export(bb_bunzip2)
 export(bb_cleanup)
+export(bb_cleanup2)
 export(bb_config)
 export(bb_config2)
 export(bb_data_source_dir)
 export(bb_data_sources)
 export(bb_decompress)
+export(bb_decompress2)
 export(bb_example_sources)
 export(bb_find_wget)
 export(bb_fingerprint)

[…]
