### 2017-12-01 c0b0180
consolidate on new (xxx2) versions of functions
