+++ R/example_sources.R
@@ -23,7 +23,7 @@
 #' @export
 bb_example_sources <- function() {
     bind_rows(
-        bb_source2(
+        bb_source(
             name="NOAA OI SST V2",
             id="oisst.v2",
             description="Weekly and monthly mean and long-term monthly mean SST data, 1-degree resolution, 1981 to present. Ice concentration data are also included, which are the ice concentration values input to the SST analysis",
@@ -31,12 +31,12 @@ bb_example_sources <- function() {
             citation="NOAA_OI_SST_V2 data provided by the NOAA/OAR/ESRL PSD, Boulder, Colorado, USA, from their web site at http://www.esrl.noaa.gov/psd/",
             source_url=c("ftp://ftp.cdc.noaa.gov/Datasets/noaa.oisst.v2/*"),
             license="Please cite",
-            method=list("bb_handler_wget2",recursive=TRUE,level=1,no_parent=TRUE),
+            method=list("bb_handler_wget",recursive=TRUE,level=1,no_parent=TRUE),
             postprocess=NULL,
             access_function="readsst",
             collection_size=0.9,
             data_group="Sea surface temperature"),
-        bb_source2(
+        bb_source(
             name="CMEMS global gridded SSH reprocessed (1993-ongoing)",
             id="SEALEVEL_GLO_PHY_L4_REP_OBSERVATIONS_008_047",
             description="For the Global Ocean - Multimission altimeter satellite gridded sea surface heights and derived variables computed with respect to a twenty-year mean. Previously distributed by Aviso+, no change in the scientific content. All the missions are homogenized with respect to a reference mission which is currently OSTM/Jason-2.\nVARIABLES\n- sea_surface_height_above_sea_level (SSH)\n- surface_geostrophic_eastward_sea_water_velocity_assuming_sea_level_for_geoid (UVG)\n- surface_geostrophic_northward_sea_water_velocity_assuming_sea_level_for_geoid (UVG)\n- sea_surface_height_above_geoid (SSH)\n- surface_geostrophic_eastward_sea_water_velocity (UVG)\n- surface_geostrophic_northward_sea_water_velocity (UVG)",
@@ -44,7 +44,7 @@ bb_example_sources <- function() {
             citation="In case of any publication, the Licensee will ensure credit the Copernicus Marine Service in the following manner: \"This study has been conducted using E.U. Copernicus Marine Service Information\"",
             source_url=c("ftp://ftp.sltac.cls.fr/Core/SEALEVEL_GLO_PHY_L4_REP_OBSERVATIONS_008_047/dataset-duacs-rep-global-merged-allsat-phy-l4-v3/"),
             license="See http://marine.copernicus.eu/services-portfolio/service-commitments-and-licence/",
-            method=list("bb_handler_wget2",recursive=TRUE,level=3,no_parent=TRUE),
+            method=list("bb_handler_wget",recursive=TRUE,level=3,no_parent=TRUE),
             postprocess=list("bb_gunzip"),
             authentication_note="Copernicus Marine login required, see http://marine.copernicus.eu/services-portfolio/register-now/",
             user="",
@@ -52,18 +52,18 @@ bb_example_sources <- function() {
             access_function="readssh",
             collection_size=310,
             data_group="Altimetry",warn_empty_auth=FALSE),
-        bb_source2(
+        bb_source(
             name="Oceandata SeaWiFS Level-3 mapped monthly 9km chl-a",
             id="SeaWiFS_L3m_MO_CHL_chlor_a_9km",
             description="Monthly remote-sensing chlorophyll-a from the SeaWiFS satellite at 9km spatial resolution",
             reference= "https://oceancolor.gsfc.nasa.gov/",
[…]
