### 2017-12-01 c8c91f3
add debug parm to bb_wget
