+++ README.md
@@ -54,8 +54,7 @@ my_source <- bb_source(
     citation="Copyright Commonwealth of Australia 2017. As far as practicable, material for which the copyright is owned by a third party will be clearly labelled. The AEC has made all reasonable efforts to ensure that this material has been reproduced on this website with the full consent of the copyright owners.",
     source_url=c("http://results.aec.gov.au/20499/Website/HouseDownloadsMenu-20499-Csv.htm"),
     license="CC-BY",
-    method=quote(bb_handler_wget),
-    method_flags=c("--recursive","--level=1","--accept=csv","--no-if-modified-since"),
+    method=list("bb_handler_wget",recursive=TRUE,level=1,accept="csv",no_if_modified_since=TRUE,execute=c("robots=off"),reject_regex="Website/UserControls"),
     collection_size=0.01)
 
 cf <- bb_config(local_file_root="~/temp/data/bbtest") %>%
@@ -108,9 +107,8 @@ my_source <- bb_source(
     citation="Spinoccia, M., 2012. XYZ multibeam bathymetric grids of the Macquarie Ridge. Geoscience Australia, Canberra.",
     source_url="http://www.ga.gov.au/corporate_data/73697/Macquarie_ESRI_Raster.zip",
     license="CC-BY 4.0",
-    method=quote(bb_handler_wget),
-    method_flags="--recursive",
-    postprocess=quote(bb_unzip),
+    method=list("bb_handler_wget",recursive=TRUE),
+    postprocess=list("bb_unzip"),
     collection_size=0.4,
     data_group="Topography")
 ```
@@ -121,21 +119,21 @@ Some particularly important components of this definition are:
 
 2.  The `license` and `citation` are important so that users know what conditions govern the usage of the data, and the appropriate citation to use to acknowledge the data providers. The `reference` entry should refer to a metadata or documentation page that describes the data in detail.
 
-3.  The `method`, `method_flags`, and `source_url` define how this data will be retrieved. Most sources will use the `bb_handler_wget` method, which is a wrapper around the `wget` utility. If you are unfamiliar with wget, consult the [wget manual](https://www.gnu.org/software/wget/manual/wget.html) or one of the many online tutorials. You can also see the in-built wget help by running `bb_wget("--help")`.
+3.  The `method` and `source_url` define how this data will be retrieved. Most sources will use the `bb_handler_wget` method function, which is a wrapper around the `wget` utility. If you are unfamiliar with wget, consult the [wget manual](https://www.gnu.org/software/wget/manual/wget.html) or one of the many online tutorials. You can also see the in-built wget help by running `bb_wget("--help")`.
 
 Some subtleties to bear in mind:
 
 1.  If the data source delivers compressed files, you will most likely want to decompress them after downloading. The postprocess options `bb_decompress`, `bb_unzip`, etc will do this for you. By default, these *do not* delete the compressed files after decompressing. The reason for this is so that on the next synchronization run, the local (compressed) copy can be compared to the remote compressed copy, and the download can be skipped if nothing has changed. Deleting local compressed files will save space on your file system, but may result in every file being re-downloaded on every synchronization run.
 
[…]
