### 2017-12-01 5f2ce96
add regexp to expect_error and expect_warning
