### 2017-12-04 2cf394e
change 'reference' parm in data source to 'doc_url', expand docs for bb_source and bb_handler_wget
