+++ tests/testthat/test_sync.R
@@ -16,7 +16,7 @@ test_that("bb_sync is quiet when asked",{
         id="bilbobaggins",
         name="test",
         description="blah",
-        reference= "http://some.where.org/",
+        doc_url= "http://some.where.org/",
         citation="blah",
         license="blah",
         method=list("bb_handler_wget",recursive=TRUE,no_check_certificate=TRUE,level=1),
@@ -32,7 +32,7 @@ test_that("bb_sync works on oceandata",{
         id="bilbobaggins",
         name="Oceandata test",
         description="Monthly remote-sensing sea surface temperature from the MODIS Terra satellite at 9km spatial resolution",
-        reference= "http://oceancolor.gsfc.nasa.gov/",
+        doc_url= "http://oceancolor.gsfc.nasa.gov/",
         citation="See http://oceancolor.gsfc.nasa.gov/cms/citations",
         license="Please cite",
         comment="",
@@ -69,7 +69,7 @@ test_that("bb_sync works with a sink() call in place",{
         id="bilbobaggins",
         name="test",
         description="blah",
-        reference= "http://some.where.org/",
+        doc_url= "http://some.where.org/",
         citation="blah",
         license="blah",
         method=list("bb_handler_wget",recursive=TRUE,no_check_certificate=TRUE,level=1),
