+++ tests/testthat/test_postprocess.R
@@ -7,7 +7,7 @@ test_that("decompressing zip files works",{
         name="Bowerbird test data",
         id="bbtest-v0.1",
         description="These are just some trivial test files provided with the bowerbird package.",
-        reference="https://github.com/AustralianAntarcticDivision/bowerbird",
+        doc_url="https://github.com/AustralianAntarcticDivision/bowerbird",
         citation="No citation needed.",
         source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data.zip",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data.zip",
         license="MIT",
@@ -32,7 +32,7 @@ test_that("decompressing gzipped files works",{
         name="Bowerbird test data",
         id="bbtest-v0.1",
         description="These are just some trivial test files provided with the bowerbird package.",
-        reference="https://github.com/AustralianAntarcticDivision/bowerbird",
+        doc_url="https://github.com/AustralianAntarcticDivision/bowerbird",
         citation="No citation needed.",
         source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data_was_gzipped.csv.gz",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data_was_gzipped.csv.gz",
         license="MIT",
@@ -57,7 +57,7 @@ test_that("decompressing bzipped files works",{
         name="Bowerbird test data",
         id="bbtest-v0.1",
         description="These are just some trivial test files provided with the bowerbird package.",
-        reference="https://github.com/AustralianAntarcticDivision/bowerbird",
+        doc_url="https://github.com/AustralianAntarcticDivision/bowerbird",
         citation="No citation needed.",
         source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data_was_bzipped.csv.bz2",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data_was_bzipped.csv.bz2",
         license="MIT",
@@ -82,7 +82,7 @@ test_that("decompressing Z-compressed files works",{
         name="Bowerbird test data",
         id="bbtest-v0.1",
         description="These are just some trivial test files provided with the bowerbird package.",
-        reference="https://github.com/AustralianAntarcticDivision/bowerbird",
+        doc_url="https://github.com/AustralianAntarcticDivision/bowerbird",
         citation="No citation needed.",
         source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/20170822.nc.Z",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/20170822.nc.Z",
         license="MIT",
