### 2017-12-04 13b582a
bb_settings should only update values passed to it, not overwrite the entire $settings object
