+++ tests/testthat/test_config.R
@@ -65,3 +65,16 @@ test_that("config validation works",{
     cf <- bb_config("/some/local/path") %>% bb_add(src) %>% bb_add(src2)
     expect_true(bb_validate(cf))
 })
+
+test_that("bb_settings works",{
+    cf <- bb_config(local_file_root="/your/data/directory")
+    temp <- bb_settings(cf)
+    expect_true(setequal(names(temp),bowerbird:::allowed_settings()))
+    expect_null(bb_settings(cf)$http_proxy)
+    ## change a setting
+    bb_settings(cf) <- list(http_proxy="something")
+    expect_identical(bb_settings(cf)$http_proxy,"something")
+    ## try to add a non-recognized setting
+    expect_warning(bb_settings(cf) <- list(bilbobaggins=1),"is not a recognized bowerbird config setting")
+    expect_null(bb_settings(cf)$bilbobaggins)
+})
