+++ man/bb_validate.Rd
@@ -1,27 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/config.R
-\name{bb_validate}
-\alias{bb_validate}
-\title{Validate a bowerbird configuration}
-\usage{
-bb_validate(config)
-}
-\arguments{
-\item{config}{bb_config: a bowerbird configuration (as returned by \code{bb_config})}
-}
-\value{
-TRUE or throw error
-}
-\description{
-Runs some basic sanity checks on a bowerbird configuration.
-}
-\examples{
-\dontrun{
-  cf <- bb_config("/my/file/root") \%>\%
-    bb_add(bb_example_sources())
-  bb_validate() ## will complain about lacking authentication info
-}
-}
-\seealso{
-\code{\link{bb_config}}
-}
