+++ NAMESPACE
@@ -25,7 +25,6 @@ export(bb_summary)
 export(bb_sync)
 export(bb_uncompress)
 export(bb_unzip)
-export(bb_validate)
 importFrom(R.utils,bunzip2)
 importFrom(R.utils,gunzip)
 importFrom(assertthat,assert_that)
