### 2017-12-07 dfbc1ed
package-level doc
