+++ R/config.R
@@ -8,6 +8,8 @@
 #' @param ftp_proxy string: URL of FTP proxy to use e.g. 'http://your.proxy:21' (NULL for no proxy)
 #' @param clobber numeric: 0=do not overwrite existing files, 1=overwrite if the remote file is newer than the local copy, 2=always overwrite existing files. For data sources that use method 'wget', an appropriate flag will be added to the wget call according to the clobber setting ("--no-clobber" to not overwrite existing files, "--timestamping" to overwrite if the remote file is newer than the local copy)
 #' @param skip_downloads logical: if TRUE, \code{bb_sync} will do a dry run of the synchronisation process but without actually downloading files. For data sources using method bb_handler_wget, this means that the wget calls will not be executed, so e.g. any recursion handled by wget itself will not be simulated
+#' @param warn_large_downloads numeric: if non-negative, \code{bb_sync} will ask the user for confirmation to download any data source of size greater than this number (in GB). A value of zero will trigger confirmation on every data source. A negative or NULL value will not prompt for confirmation. Note that this only applies when R is being used interactively. The expected download size is taken from the \code{collection_size} parameter of the data source, and so its accuracy is dependent on the accuracy of the data source definition
+#'
 #' @return configuration object
 #'
 #' @seealso \code{\link{bb_source}}
@@ -24,11 +26,17 @@
 #' }
 #'
 #' @export
-bb_config <- function(local_file_root,wget_global_flags=list(restrict_file_names="windows",progress="dot:giga"),http_proxy=NULL,ftp_proxy=NULL,clobber=1,skip_downloads=FALSE) {
+bb_config <- function(local_file_root,wget_global_flags=list(restrict_file_names="windows",progress="dot:giga"),http_proxy=NULL,ftp_proxy=NULL,clobber=1,skip_downloads=FALSE,warn_large_downloads=0.1) {
     assert_that(is.string(local_file_root))
     assert_that(clobber %in% c(0,1,2))
     assert_that(is.flag(skip_downloads),!is.na(skip_downloads))
     assert_that(is.list(wget_global_flags))
+    if (!is.null(warn_large_downloads)) {
+        assert_that(is.numeric(warn_large_downloads),!is.na(warn_large_downloads))
+        if (warn_large_downloads<0) warn_large_downloads <- Inf
+    } else {
+        warn_large_downloads <- Inf
+    }
     structure(
         list(data_sources=tibble(),
              settings=list(
@@ -37,7 +45,8 @@ bb_config <- function(local_file_root,wget_global_flags=list(restrict_file_names
                  ftp_proxy=ftp_proxy,
                  local_file_root=local_file_root,
                  clobber=clobber,
-                 skip_downloads=skip_downloads)),
+                 skip_downloads=skip_downloads,
+                 warn_large_downloads=warn_large_downloads)),
         class="bb_config")
 }
 
@@ -134,7 +143,7 @@ bb_settings <- function(config) {
 }
 
 ## internal: the list of recognized config settings
-allowed_settings <- function() c("wget_global_flags","http_proxy","ftp_proxy","local_file_root","clobber","skip_downloads")
+allowed_settings <- function() c("wget_global_flags","http_proxy","ftp_proxy","local_file_root","clobber","skip_downloads","warn_large_downloads")
 
 
 #' Gets or sets a bowerbird configuration object's data sources
