### 2017-12-07 f437641
drop execute parm from wget, replace with robots_off
