### 2017-12-07 a296983
update tests
