+++ tests/testthat/test_provenance.R
@@ -11,7 +11,7 @@ test_that("bb_fingerprint does something sensible",{
         citation="No citation needed.",
         source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data_was_gzipped.csv.gz",##https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data_was_gzipped.csv.gz",
         license="MIT",
-        method=list("bb_handler_wget",recursive=TRUE,level=1,no_check_certificate=TRUE,execute=c("robots=off")),
+        method=list("bb_handler_wget",recursive=TRUE,level=1,no_check_certificate=TRUE,robots_off=TRUE),
         postprocess=list("bb_gunzip"))
 
     temp_root <- tempdir()
