+++ NAMESPACE
@@ -54,6 +54,7 @@ importFrom(stringr,str_trim)
 importFrom(sys,exec_internal)
 importFrom(sys,exec_wait)
 importFrom(utils,download.file)
+importFrom(utils,menu)
 importFrom(utils,read.table)
 importFrom(utils,str)
 importFrom(utils,unzip)
