### 2018-01-11 917e7d1
drop dplyr usage, drop verbose arg
