### 2018-01-11 8850d36
drop dplyr usage
