+++ R/bowerbird.R
@@ -7,7 +7,7 @@
 #' @docType package
 #' @references \url{http://data.aad.gov.au}
 #' @importFrom assertthat assert_that is.flag is.string
-#' @importFrom dplyr arrange_ bind_rows group_by_ mutate_ rename_ select_ slice tbl_df tibble ungroup %>%
+#' @importFrom magrittr %>%
 #' @importFrom methods is
 #' @importFrom openssl sha1 md5
 #' @importFrom stringr regex str_detect str_match str_split str_trim
@@ -15,5 +15,6 @@
 #' @importFrom rmarkdown render
 #' @importFrom stats na.omit
 #' @importFrom sys exec_wait exec_internal
+#' @importFrom tibble as_tibble tibble
 #' @importFrom utils download.file menu read.table str unzip
 NULL
