+++ R/re_exports.R
@@ -1,2 +1,2 @@
 #' @export
-dplyr::`%>%`
+magrittr::`%>%`
