+++ NAMESPACE
@@ -30,20 +30,11 @@ importFrom(R.utils,gunzip)
 importFrom(assertthat,assert_that)
 importFrom(assertthat,is.flag)
 importFrom(assertthat,is.string)
-importFrom(dplyr,"%>%")
-importFrom(dplyr,arrange_)
-importFrom(dplyr,bind_rows)
-importFrom(dplyr,group_by_)
-importFrom(dplyr,mutate_)
-importFrom(dplyr,rename_)
-importFrom(dplyr,select_)
-importFrom(dplyr,slice)
-importFrom(dplyr,tbl_df)
-importFrom(dplyr,tibble)
-importFrom(dplyr,ungroup)
+importFrom(magrittr,"%>%")
 importFrom(methods,is)
 importFrom(openssl,md5)
 importFrom(openssl,sha1)
+importFrom(rlang,.data)
 importFrom(rmarkdown,render)
 importFrom(stats,na.omit)
 importFrom(stringr,regex)
@@ -53,6 +44,8 @@ importFrom(stringr,str_split)
 importFrom(stringr,str_trim)
 importFrom(sys,exec_internal)
 importFrom(sys,exec_wait)
+importFrom(tibble,as_tibble)
+importFrom(tibble,tibble)
 importFrom(utils,download.file)
 importFrom(utils,menu)
 importFrom(utils,read.table)
