+++ NAMESPACE
@@ -34,7 +34,6 @@ importFrom(magrittr,"%>%")
 importFrom(methods,is)
 importFrom(openssl,md5)
 importFrom(openssl,sha1)
-importFrom(rlang,.data)
 importFrom(rmarkdown,render)
 importFrom(stats,na.omit)
 importFrom(stringr,regex)
