### 2018-01-12 34d8e42
updates to README
