+++ R/source.R
@@ -8,7 +8,7 @@
 #'
 #' @param id string: (required) a unique identifier of the data source. If the data source has a DOI, use that. Otherwise, if the original data provider has an identifier for this dataset, that is probably a good choice here (include the data version number if there is one). The ID should be something that changes when the data set changes (is updated). A DOI is ideal for this
 #' @param name string: (required) a unique name for the data source. This should be a human-readable but still concise name
-#' @param description string: a description of the data source
+#' @param description string: a plain-language description of the data source, providedso that users can get an idea of what the data source contains (for full details they can consult the \code{doc_url} link)
 #' @param doc_url string: (required) URL to the metadata record or other documentation of the data source
 #' @param source_url character vector: one or more source URLs. Required for \code{bb_handler_wget}, although some \code{method} functions might not require one
 #' @param citation string: (required) details of the citation for the data source
