+++ README.Rmd
@@ -18,7 +18,7 @@ knitr::opts_chunk$set(
 
 # Bowerbird
 
-<img align="right" src="https://rawgit.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/bowerbird.svg" />
+<!--img align="right" src="https://rawgit.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/bowerbird.svg" /-->
 
 Often it's desirable to have local copies of third-party data sets. Fetching data on the fly from remote sources can be a great strategy, but for speed or other reasons it may be better to have local copies. This is particularly common in environmental and other sciences that deal with large data sets (e.g. satellite or global climate model products). Bowerbird is an R package for maintaining a local collection of data sets from a range of data providers.
 
@@ -44,23 +44,143 @@ library(devtools)
 install_github("AustralianAntarcticDivision/bowerbird",build_vignettes=TRUE)
 ```
 
-Bowerbird uses the third-party utility `wget` to do the heavy lifting of recursively downloading files from data providers. `wget` is typically installed by default on Linux.
-On Windows you can use the `bb_install_wget()` function to install it. Otherwise download `wget` yourself (e.g. from https://eternallybored.org/misc/wget/current/wget.exe) and make sure it is on your path.
+You will also need to have the third-party utility `wget` available, because Bowerbird uses this to do the heavy lifting of recursively downloading files from data providers. Note that installing `wget` may require admin privileges on your local machine.
 
-## Usage
+#### Linux
+
+`wget` is typically installed by default on Linux. Otherwise use your package manager to install it, e.g. `sudo apt install wget` on Ubuntu/Debian or `sudo yum install wget` on Fedora/CentOS.
+
+#### Windows
+
+On Windows you can use the `bb_install_wget()` function to install it. Otherwise download `wget` yourself (e.g. from https://eternallybored.org/misc/wget/current/wget.exe) and make sure it is on your system path.
+
+#### Mac
+
+Use `brew install wget` or `brew install --with-libressl wget` if you get SSL-related errors. If you do not have brew installed, see https://brew.sh/.
+
+
+## Usage overview
 
 ### Configuration
 
 Build up a configuration by first defining global options such as the destination on your local file system:
 
 ```{r eval=FALSE}
+library(bowerbird)
 cf <- bb_config(local_file_root="~/your/data/directory")
 ```
 
-Bowerbird must then be told which data sources to synchronize. Use the `bb_source()` function to define a data source. Let's use data from the 2016 Australian federal election as an example:
+Bowerbird must then be told which data sources to synchronize. Let's use data from the Australian 2016 federal election, which is provided as one of the example data sources:
 
 ```{r eval=FALSE}
-my_source <- bb_source(
+my_source <- bb_example_sources() ## all the example data sources
+my_source <- my_source[my_source$id=="aus-election-house-2016"] ## just the 2016 election data example
+
+## add this data source to the configuration
+cf <- bb_add(cf,my_source)
+```
+
+Once the configuration has been defined and the data source added to it, run the sync process:
+
+```{r eval=FALSE}
+status <- bb_sync(cf)
+```
+
+Congratulations! You now have your own local copy of your chosen data set. This particular example is fairly small (about 10MB), so it should not take too long to download.
+
+At a later time you can re-run this synchronization process. If the remote files have not changed, and assuming that your configuration has the `clobber` parameter set to 0 ("do not overwrite existing files") or 1 ("overwrite only if the remote file is newer than the local copy") then the sync process will run more quickly because it will not need to re-download any data files.
+
+
+## Defining data sources
+
+### Prepackaged data source definitions
+
[…]

[…]
