### 2018-01-13 5e307e3
test for bb_cleanup added
