+++ tests/testthat/test_postprocess.R
@@ -99,5 +99,24 @@ test_that("decompressing Z-compressed files works",{
     expect_true(file.info(file.path(fp,"20170822.nc"))$size>1e4)
 })
 
+test_that("cleanup postprocessing works",{
+    skip_on_cran()
+    skip_on_appveyor() ## fails for unknown reasons
+    my_source <- bb_source(
+        name="Bowerbird test data",
+        id="bbtest-v0.1",
+        description="These are just some trivial test files provided with the bowerbird package.",
+        doc_url="https://github.com/AustralianAntarcticDivision/bowerbird",
+        citation="No citation needed.",
+        source_url="https://raw.githubusercontent.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/example_data.zip",##"https://github.com/AustralianAntarcticDivision/bowerbird/raw/master/inst/extdata/example_data.zip",
+        license="MIT",
+        method=list("bb_handler_wget",recursive=TRUE,level=1,robots_off=TRUE),
+        postprocess=list("bb_unzip",list("bb_cleanup",pattern="\\.csv$"))
+    )
+    temp_root <- tempdir()
+    cf <- bb_add(bb_config(local_file_root=temp_root,clobber=2),my_source)
+    bb_sync(cf)
 
-
+    fp <- bb_data_source_dir(cf)
+    expect_false(file.exists(file.path(fp,"example_data_was_zipped.csv")))
+})
