### 2018-01-13 48dda98
change to verbose=FALSE by default
