### 2018-01-13 f45128d
adjust some method parms
