### 2018-01-13 fb61191
break postprocess functions up so that the function definition exposed to the user is simplified
