### 2018-01-13 7f4f3a6
add retr_symlinks parm, words to doc
