### 2018-01-13 fe29db9
readme updates
