+++ README.Rmd
@@ -73,8 +73,7 @@ cf <- bb_config(local_file_root="~/your/data/directory")
 Bowerbird must then be told which data sources to synchronize. Let's use data from the Australian 2016 federal election, which is provided as one of the example data sources:
 
 ```{r eval=FALSE}
-my_source <- bb_example_sources() ## all the example data sources
-my_source <- my_source[my_source$id=="aus-election-house-2016"] ## just the 2016 election data example
+my_source <- subset(bb_example_sources(),id=="aus-election-house-2016")
 
 ## add this data source to the configuration
 cf <- bb_add(cf,my_source)
@@ -130,7 +129,7 @@ The parameters provided here are all mandatory:
 - `doc_url` is a link to a metadata record or documentation page that describes the data in detail
 - `license` is the license under which the data are being distributed, and is required so that users are aware of the conditions that govern the usage of the data
 - `citation` gives citation details for the data source. It's generally considered good practice to cite data providers, and indeed under some data licenses this is in fact mandatory
-- `method` defines the handler function that will be used to retrieve this data set (`bb_handler_wget`, in this case), along with any required parameters (none here)
+- the `method` parameter is specified as a list, where the first entry is the name of the handler function that will be used to retrieve this data set (`bb_handler_wget`, in this case)and the remaining entries are data-source-specific arguments to pass to that function (none here)
 - and finally the `source_url` tells bowerbird where to go to get the data.
 
 
@@ -238,7 +237,7 @@ src3 <- bb_source(
     source_url=c("https://daacdata.apps.nsidc.org/pub/DATASETS/nsidc0192_seaice_trends_climo_v2/"),
     license="Please cite, see http://nsidc.org/about/use_copyright.html",
     authentication_note="Requires Earthdata login, see https://wiki.earthdata.nasa.gov/display/EL/How+To+Register+With+Earthdata+Login . Note that you will also need to authorize the application 'nsidc-daacdata' (see 'My Applications' at https://urs.earthdata.nasa.gov/profile)",
-    method=list("bb_handler_earthdata",recursive=TRUE,level=4,no_parent=TRUE,relative=TRUE),
+    method=list("bb_handler_earthdata",recursive=TRUE,level=4,relative=TRUE),
     user="your_earthdata_username",
     password="your_earthdata_password",
     collection_size=0.02,
@@ -249,7 +248,6 @@ This is very similar to our previous examples, with these differences:
 
 - the `method` specifies `bb_handler_earthdata` (whereas previously we used `bb_handler_wget`). The `bb_handler_earthdata` is actually very similar to `bb_handler_wget`, but it takes care of some Earthdata-specific details like authentication using your Earthdata credentials
 - we want a `recursive=TRUE` download, because the data are arranged in subdirectories. Manual browsing of the data set indicates that we need four levels of recursion, hence `level=4`
-- `no_parent=TRUE` prevents `wget` from ascending to a parent directory during its recursion process, because if it did so it would download files that are not part of the data set that we want. Specifying `no_parent=TRUE` means that `wget` will not climb up out of the `https://daacdata.apps.nsidc.org/pub/DATASETS/nsidc0192_seaice_trends_climo_v2/` directory and into some other area
 - `relative=TRUE` means that `wget` will only follow relative links (i.e. links of the form `<a href="/some/directory/">...</a>`, which by definition must be on the same server as our `source_url`). Absolute links (i.e. links of the form `<a href="http://some.other.server/some/path">...</a>` will not be followed. This is another mechanism to prevent the recursive download from downloading stuff we don't want.
 
 
@@ -271,60 +269,79 @@ mysrc <- src3
[…]
