### 2018-01-13 7d1d062
readme updates
