+++ README.Rmd
@@ -18,7 +18,7 @@ knitr::opts_chunk$set(
 
 # Bowerbird
 
-<!--img align="right" src="https://rawgit.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/bowerbird.svg" /-->
+<img align="right" src="https://rawgit.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/bowerbird.svg" />
 
 Often it's desirable to have local copies of third-party data sets. Fetching data on the fly from remote sources can be a great strategy, but for speed or other reasons it may be better to have local copies. This is particularly common in environmental and other sciences that deal with large data sets (e.g. satellite or global climate model products). Bowerbird is an R package for maintaining a local collection of data sets from a range of data providers.
 
@@ -275,25 +275,30 @@ cf <- bb_add(cf,mysrc)
 
 ### Example 4: an Oceandata source
 
-NASA's [Oceandata](https://oceandata.sci.gsfc.nasa.gov/) system provides access to a range of satellite-derived marine data products. The `bb_oceandata_handler` can be used to download these data. It does a two-step process: first it makes a query to the Oceancolour data file search tool (https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi) to find files that match your specified criterion, and then downloads the matching files.
+NASA's [Oceandata](https://oceandata.sci.gsfc.nasa.gov/) system provides access to a range of satellite-derived marine data products. The `bb_oceandata_handler` can be used to download these data. It uses a two-step process: first it makes a query to the Oceancolour data file search tool (https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi) to find files that match your specified criterion, and then downloads the matching files.
 
+Oceandata uses standardized file naming conventions (see https://oceancolor.gsfc.nasa.gov/docs/format/), so once you know which products you want you can construct a suitable file name pattern to search for. For example, "S*L3m_MO_CHL_chlor_a_9km.nc" would match monthly level-3 mapped chlorophyll data from the SeaWiFS satellite at 9km resolution, in netcdf format. This pattern is passed as the `search` argument to the `bb_handler_oceandata` handler function. Note that the `bb_handler_oceandata` does not take need `source_url` to be specified.
 
+Here, for the sake of a small example, we'll limit ourselves to a single file ("T20000322000060.L3m_MO_SST_sst_9km.nc", which is sea surface temperature from the Terra satellite in February 2000):
 
-#' my_source <- bb_source(
-#'   name="Oceandata SeaWiFS Level-3 mapped monthly 9km chl-a",
-#'   id="SeaWiFS_L3m_MO_CHL_chlor_a_9km",
-#'   description="Monthly remote-sensing chlorophyll-a from the SeaWiFS satellite at
-#'     9km spatial resolution",
-#'   doc_url="https://oceancolor.gsfc.nasa.gov/",
-#'   citation="See https://oceancolor.gsfc.nasa.gov/citations",
-#'   license="Please cite",
-#'   method=list("bb_handler_oceandata",search="S*L3m_MO_CHL_chlor_a_9km.nc"),
-#'   postprocess=NULL,
-#'   collection_size=7.2,
-#'   data_group="Ocean colour")
+```{r eval=FALSE}
+src4 <- bb_source(
+    name="Oceandata test file",
+    id="oceandata-test",
+    description="Monthly, 9km remote-sensed sea surface temperature from the MODIS Terra satellite",
+    doc_url= "http://oceancolor.gsfc.nasa.gov/",
+    citation="See http://oceancolor.gsfc.nasa.gov/cms/citations",
+    license="Please cite",
+    method=list("bb_handler_oceandata",search="T20000322000060.L3m_MO_SST_sst_9km.nc"),
+    data_group="Sea surface temperature")
 
[…]
