### 2018-01-13 89f21f4
remove lost and lonely dplyr call
