### 2018-01-13 fae7b76
break handler functions up so that parameterization exposed to the user is simplified
