+++ R/earthdata_handler.R
@@ -1,14 +1,13 @@
-#' Handler for files files served up through Earthdata providers
+#' Handler for data sets from Earthdata providers
 #'
-#' This function is not intended to be called directly by the user, but rather will be called internally by the \code{bb_sync} function. The typical usage of \code{bb_handler_earthdata} is to specify it in the \code{method} parameter of a definition: see the example below.
+#' This is a handler function to be used with data sets from NASA's Earthdata system. This function is not intended to be called directly, but rather is specified as a \code{postprocess} option in \code{\link{bb_source}}.
+#'
+#' This function uses \code{\link{bb_wget}}, and so data sources using this function will need to provide appropriate \code{\link{bb_wget}} parameters.
 #'
 #' @references https://wiki.earthdata.nasa.gov/display/EL/How+To+Register+With+Earthdata+Login
-#' @param config bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source
-#' @param verbose logical: if TRUE, provide additional progress output
-#' @param local_dir_only logical: if TRUE, just return the local directory into which files from this data source would be saved
-#' @param ... : additional parameters passed through to bb_handler_wget
+#' @param ... : parameters passed to \code{\link{bb_wget}}
 #'
-#' @return the directory if local_dir_only is TRUE, otherwise TRUE on success
+#' @return TRUE on success
 #'
 #' @examples
 #' \dontrun{
@@ -33,7 +32,15 @@
 #' }
 #'
 #' @export
-bb_handler_earthdata <- function(config,verbose=FALSE,local_dir_only=FALSE,...) {
+bb_handler_earthdata <- function(...) {
+    do.call(bb_handler_earthdata_inner,list(...))
+}
+
+
+# @param config bb_config: a bowerbird configuration (as returned by \code{bb_config}) with a single data source
+# @param verbose logical: if TRUE, provide additional progress output
+# @param local_dir_only logical: if TRUE, just return the local directory into which files from this data source would be saved
+bb_handler_earthdata_inner <- function(config,verbose=FALSE,local_dir_only=FALSE,...) {
     assert_that(is(config,"bb_config"))
     assert_that(nrow(bb_data_sources(config))==1)
     assert_that(is.flag(verbose),!is.na(verbose))
@@ -57,13 +64,13 @@ bb_handler_earthdata <- function(config,verbose=FALSE,local_dir_only=FALSE,...) 
     on.exit(file.remove(cookies_file))
     ##mflags <- list(...)##flags_to_charvec(dummy$method_flags)
     ## must use --auth-no-challenge else the server redirects to the html login page, rather than accepting the provided credentials
-    dummy$method <- list("bb_handler_wget",...,extra_flags=c("--http-user",dummy$user,"--http-password",dummy$password,"--auth-no-challenge","--load-cookies",cookies_file,"--save-cookies",cookies_file,"--keep-session-cookies"),reject="index.html*",robots_off=TRUE)
+    dummy$method <- list(list("bb_handler_wget",...,extra_flags=c("--http-user",dummy$user,"--http-password",dummy$password,"--auth-no-challenge","--load-cookies",cookies_file,"--save-cookies",cookies_file,"--keep-session-cookies"),reject="index.html*",robots_off=TRUE))
     ##dummy$method_flags <- list(c(mflags,"--http-user",dummy$user,"--http-password",dummy$password,"--auth-no-challenge","--load-cookies",cookies_file,"--save-cookies",cookies_file,"--keep-session-cookies","--reject=index.html*","-e","robots=off"))
     dummy$user <- NA_character_
     dummy$password <- NA_character_
     bb_data_sources(config) <- dummy
[…]
