### 2018-01-13 c1a1a2f
doco tweaks
