+++ vignettes/bowerbird.Rmd
@@ -4,7 +4,7 @@ author: "Ben Raymond, Michael Sumner"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Bowerbird}
+  %\VignetteIndexEntry{bowerbird}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
