### 2018-01-14 bb88d93
fix odd bb_settings behaviour
