+++ man/bb_settings.Rd
@@ -18,7 +18,10 @@ bb_settings(config) <- value
 named list
 }
 \description{
-These are repository-wide settings that are applied to all data sources added to the configuration.
+Gets or sets a bowerbird configuration object's settings. These are repository-wide settings that are applied to all data sources added to the configuration.
+}
+\details{
+Note that an assignment along the lines of \code{bb_settings(cf) <- new_settings} replaces all of the settings in the configuration with the \code{new_settings}. The most common usage pattern is to read the existing settings, modify them as needed, and then rewrite the whole lot back into the configuration object (as per the examples here).
 }
 \examples{
 cf <- bb_config(local_file_root="/your/data/directory")
@@ -27,10 +30,14 @@ cf <- bb_config(local_file_root="/your/data/directory")
 bb_settings(cf)
 
 ## add an http proxy
-bb_settings(cf) <- list(http_proxy="http://my.proxy")
+sets <- bb_settings(cf)
+sets$http_proxy <- "http://my.proxy"
+bb_settings(cf) <- sets
 
 ## change the current local_file_root setting
-bb_settings(cf) <- list(local_file_root="/new/location")
+sets <- bb_settings(cf)
+sets$local_file_root <- "/new/location"
+bb_settings(cf) <- sets
 
 }
 \seealso{
