### 2018-01-14 d4b80ff
update docs, better messages around wget installation
