+++ man/bb_find_wget.Rd
@@ -2,20 +2,20 @@
 % Please edit documentation in R/wget.R
 \name{bb_find_wget}
 \alias{bb_find_wget}
-\title{Find the wget executable, and optionally install it if it is not found}
+\title{Find the wget executable}
 \usage{
 bb_find_wget(install = FALSE, error = TRUE)
 }
 \arguments{
 \item{install}{logical: attempt to install the executable if it is not found? (Windows only)}
 
-\item{error}{logical: if wget is not found, raise an error if wget is not found. If FALSE, do not raise an error but return NULL}
+\item{error}{logical: if wget is not found, raise an error. If \code{FALSE}, do not raise an error but return NULL}
 }
 \value{
-the path to the wget executable, or (if error is FALSE) NULL if it was not found
+the path to the wget executable, or (if error is \code{FALSE}) NULL if it was not found
 }
 \description{
-Installation (if required) currently only works on Windows platforms. The wget.exe executable will be downloaded from https://eternallybored.org/misc/wget/current/wget.exe and installed into your appdata directory (typically something like C:/Users/username/AppData/Roaming/)
+This function will return the path to the wget executable if it can be found on the local system, and optionally install it if it is not found. Installation (if required) currently only works on Windows platforms. The wget.exe executable will be downloaded from https://eternallybored.org/misc/wget/current/wget.exe and installed into your appdata directory (typically something like C:/Users/username/AppData/Roaming/)
 }
 \examples{
 \dontrun{
