+++ man/bb_data_source_dir.Rd
@@ -13,13 +13,11 @@ bb_data_source_dir(config)
 character vector of directories
 }
 \description{
-Files from each data source are stored locally in the associated directory.
+Return the local directory of each data source in a configuration. Files from each data source are stored locally in the associated directory.
 }
 \examples{
-\dontrun{
-  cf <- bb_config("/my/file/root") \%>\%
-    bb_add(bb_example_sources())
-  bb_data_source_dir(cf)
-}
+cf <- bb_config("/my/file/root") \%>\%
+  bb_add(bb_example_sources())
+bb_data_source_dir(cf)
 
 }
