### 2018-01-14 1e6dd30
fix wget install test
