### 2018-01-14 a28d99b
docs, vignettes
