+++ README.md
@@ -32,15 +32,15 @@ install_github("AustralianAntarcticDivision/bowerbird",build_vignettes=TRUE)
 
 You will also need to have the third-party utility `wget` available, because Bowerbird uses this to do the heavy lifting of recursively downloading files from data providers. Note that installing `wget` may require admin privileges on your local machine.
 
-#### Linux
+### Linux
 
 `wget` is typically installed by default on Linux. Otherwise use your package manager to install it, e.g. `sudo apt install wget` on Ubuntu/Debian or `sudo yum install wget` on Fedora/CentOS.
 
-#### Windows
+### Windows
 
 On Windows you can use the `bb_install_wget()` function to install it. Otherwise download `wget` yourself (e.g. from <https://eternallybored.org/misc/wget/current/wget.exe>) and make sure it is on your system path.
 
-#### Mac
+### Mac
 
 Use `brew install wget` or `brew install --with-libressl wget` if you get SSL-related errors. If you do not have brew installed, see <https://brew.sh/>.
 
@@ -82,8 +82,7 @@ Defining data sources
 
 A few example data source definitions are provided as part of the bowerbird package --- see `bb_example_sources()` (these are also listed at the bottom of this document). Other packages (e.g. [blueant](https://github.com/AustralianAntarcticDivision/blueant)) provide themed sets of data sources that can be used with bowerbird.
 
-Defining your own data sources
-------------------------------
+### Defining your own data sources
 
 The general bowerbird workflow is to build up a configuration with one or more data sources, and pass that configuration object to the `bb_sync` function to kick off the download process. Each data source contains the details required by bowerbird to be able to fetch it, including a *handler* function that bb\_sync will call to do the actual download.
 
@@ -91,9 +90,9 @@ The `bb_handler_wget` function is a generic handler function that will be suitab
 
 `bb_handler_wget` is a wrapper around the `wget` utility. The philosophy of bowerbird is to use `wget` as much as possible to handle web transactions. Using `wget` (and in particular its recursive download functionality) simplifies the process of writing and maintaining data source definitions. Typically, one only needs to define a data source in terms of its top-level URL and appropriate flags to pass to `wget`, along with some basic metadata (primarily intended to be read by the user).
 
-Specifying a data source is done by the `bb_source` function. This can seem a little daunting, so let's work through some examples. (Most of these examples are included in `bb_example_sources()`.)
+Specifying a data source is done by the `bb_source` function. This can seem a little daunting, so let's work through some examples. Most of these examples are included in `bb_example_sources()`.
 
-### Example 1: a single data file
+#### Example 1: a single data file
 
 Say we've found [this bathymetric data set](https://doi.org/10.4225/25/53D9B12E0F96E) and we want to define a data source for it. It's a single zip file that contains some ArcGIS binary grids. A minimal data source definition might look like this:
 
@@ -158,9 +157,11 @@ src1 <- bb_source(
     data_group="Topography")
 ```
 
-### Example 2: multiple files via recursive download
+#### Example 2: multiple files via recursive download
 
 This data source (Australian Election 2016 House of Representatives data) is provided as one of the example data sources in `bb_example_sources()`, but let's look in a little more detail here.
+
+The primary entry point to this data set is an HTML index page, which links to a number of data files. In principle we could generate a list of all of these data files and download them one by one, but that would be tedious and prone to breaking (if the data files changed our hard-coded list would no longer be correct). Instead we can start at the HTML index and recursively download linked data files.
 
 The definition for this data source is:
 
@@ -195,7 +196,7 @@ Once again the data have been saved into a subdirectory that reflects the URL st
 
[…]
