### 2018-01-14 6868fa4
docs, vignettes
