### 2018-01-15 6351e7a
move postprocess-related funcs from util to postprocess; make do_decompress honour the verbose setting
