+++ R/postprocess.R
@@ -54,18 +54,186 @@ bb_decompress_inner <- function(config,file_list_before,file_list_after,verbose=
                            )
     if (delete) {
         files_to_decompress <- list.files(directory_from_url(bb_data_sources(config)$source_url),pattern=file_pattern,recursive=TRUE,ignore.case=ignore_case)
-        do_decompress_files(paste0(method,"_delete"),files=files_to_decompress)
+        do_decompress_files(paste0(method,"_delete"),files=files_to_decompress,verbose=verbose)
     } else {
         ## decompress but retain compressed file. decompress only if .zip/.gz/.bz2/.Z file has changed
         files_to_decompress <- find_changed_files(file_list_before,file_list_after,file_pattern)
-        do_decompress_files(method,files=files_to_decompress)
+        do_decompress_files(method,files=files_to_decompress,verbose=verbose)
         ## also decompress if uncompressed file does not exist
         files_to_decompress <- setdiff(rownames(file_list_after),files_to_decompress) ## those that we haven't just dealt with
         files_to_decompress <- files_to_decompress[str_detect(files_to_decompress,regex(file_pattern,ignore_case=ignore_case))] ## only .zip/.gz/.bz2/.Z files
-        do_decompress_files(method,files=files_to_decompress,overwrite=FALSE)
+        do_decompress_files(method,files=files_to_decompress,overwrite=FALSE,verbose=verbose)
         ## nb this may be slow, so might be worth explicitly checking for the existence of uncompressed files
     }
 }
+
+## internal function
+do_decompress_files <- function(method,files,overwrite=TRUE,verbose=FALSE) {
+    ## decompress (unzip/gunzip) compressed files
+    ## this function overwrites existing decompressed files if overwrite is TRUE
+    assert_that(is.string(method))
+    method <- match.arg(method,c("unzip","unzip_delete","gunzip","gunzip_delete","bunzip2","bunzip2_delete","uncompress","uncompress_delete"))
+    if (grepl("uncompress",method)) {
+        ## uncompress uses archive::file_read, which is a suggested package
+        ## check that we have it
+        if (!requireNamespace("archive",quietly=TRUE))
+            stop("the archive package is needed for uncompress functionality")
+
+    }
+    ## unzip() issues warnings in some cases when operations have errors, and sometimes issues actual errors
+    warn <- getOption("warn") ## save current setting
+    options(warn=0) ## so that we can be sure that last.warning will be set
+    ## MDS: this should use tail(warnings(), 1) instead
+    last.warning <- NULL ## to avoid check note
+    all_OK <- TRUE
+    for (thisf in files) {
+        ## decompress, check for errors in doing so
+        if (verbose) cat(sprintf("  decompressing: %s ... ",thisf))
+        switch(method,
+               "unzip_delete"=,
+               "unzip"={
+                   was_ok <- FALSE
+                   suppressWarnings(warning("")) ## clear last.warning message
+                   ## unzip will put files in the current directory by default, so we need to extract the target directory for this file
+                   target_dir <- dirname(thisf)
+                   tryCatch({ unzipped_files <- unzip(thisf,list=TRUE) ## get list of files in archive
+                              files_to_extract <- unzipped_files$Name
+                              if (!overwrite) {
+                                  ## extract only files that don't exist
+                                  files_to_extract<-files_to_extract[!file.exists(file.path(target_dir,files_to_extract))]
+                              }
+                              if (length(files_to_extract)>0) {
+                                  if (verbose) cat(sprintf('extracting %d files into %s ... ',length(files_to_extract),target_dir))
+                                  unzip(thisf,files=files_to_extract,exdir=target_dir) ## now actually unzip them
+                                  was_ok <- is.null(last.warning[[1]]) && all(file.info(files_to_extract)$size>0)
+                              } else {
[…]
