+++ R/utils.R
@@ -103,179 +103,6 @@ directory_from_url <- function(this_url) {
     this_url ## returns char vector
 }
 
-find_changed_files <- function(file_list_before,file_list_after,filename_pattern=".*") {
-    ## expect both file_list_before and file_list_after to be a data.frame from file.info()
-    ## detect changes on basis of ctime and size attributes
-    ## returns names only
-    changed_files <- setdiff(rownames(file_list_after),rownames(file_list_before)) ## anything that has appeared afterwards
-    for (thisf in intersect(rownames(file_list_after),rownames(file_list_before))) {
-        ## files in both
-        thisfile_after <- file_list_after[rownames(file_list_after)==thisf,]
-        thisfile_before <- file_list_before[rownames(file_list_before)==thisf,]
-        if ((thisfile_after$ctime>thisfile_before$ctime) | (thisfile_after$size!=thisfile_before$size)) {
-            changed_files <- c(changed_files,thisf)
-        }
-    }
-    changed_files <- changed_files[str_detect(changed_files,filename_pattern)]
-    if (is.null(changed_files)) {
-        c()
-    } else {
-        changed_files
-    }
-}
-
-
-do_decompress_files <- function(method,files,overwrite=TRUE) {
-    ## decompress (unzip/gunzip) compressed files
-    ## this function overwrites existing decompressed files if overwrite is TRUE
-    assert_that(is.string(method))
-    method <- match.arg(method,c("unzip","unzip_delete","gunzip","gunzip_delete","bunzip2","bunzip2_delete","uncompress","uncompress_delete"))
-    if (grepl("uncompress",method)) {
-        ## uncompress uses archive::file_read, which is a suggested package
-        ## check that we have it
-        if (!requireNamespace("archive",quietly=TRUE))
-            stop("the archive package is needed for uncompress functionality")
-
-    }
-    ## unzip() issues warnings in some cases when operations have errors, and sometimes issues actual errors
-    warn <- getOption("warn") ## save current setting
-    options(warn=0) ## so that we can be sure that last.warning will be set
-    ## MDS: this should use tail(warnings(), 1) instead
-    last.warning <- NULL ## to avoid check note
-    all_OK <- TRUE
-    for (thisf in files) {
-        ## decompress, check for errors in doing so
-        cat(sprintf("  decompressing: %s ... ",thisf))
-        switch(method,
-               "unzip_delete"=,
-               "unzip"={
-                   was_ok <- FALSE
-                   suppressWarnings(warning("")) ## clear last.warning message
-                   ## unzip will put files in the current directory by default, so we need to extract the target directory for this file
-                   target_dir <- dirname(thisf)
-##cat(sprintf("\nWe are in %s\n",getwd()))
-##cat(sprintf("thisf is: %s\n",thisf))
-                   tryCatch({ unzipped_files <- unzip(thisf,list=TRUE) ## get list of files in archive
-##cat("unzipped files: ")
-##cat(str(unzipped_files),"\n")
-                              files_to_extract <- unzipped_files$Name
-                              if (!overwrite) {
-                                  ## extract only files that don't exist
-                                  files_to_extract<-files_to_extract[!file.exists(file.path(target_dir,files_to_extract))]
-                              }
-##cat("files to extract: ")
-##cat(str(files_to_extract),"\n")
-                              if (length(files_to_extract)>0) {
-                                  cat(sprintf('extracting %d files into %s ... ',length(files_to_extract),target_dir))
-                                  unzip(thisf,files=files_to_extract,exdir=target_dir) ## now actually unzip them
-                                  was_ok <- is.null(last.warning[[1]]) && all(file.info(files_to_extract)$size>0)
-                              } else {
-                                  cat(sprintf('no new files to extract (not overwriting existing files) ... '))
[…]
