### 2018-01-28 21d732a
adjust items in function doc to match function
