### 2018-01-28 c764c13
synchronis* spelling to synchroniz*
