### 2018-01-28 7c7c995
add dontrun example to bb_sync, update readme example
