### 2018-01-28 fc7b7ac
update function doc, return full file paths
