### 2018-01-28 c6b7bf5
update function docs
