### 2018-01-28 be82c50
update readme, vignette
