### 2018-01-29 4b22621
bb_install_wget to use temp dir by default, update link to wget binary
