### 2018-01-29 c810e2e
path backslashes in doc to forward slashes to avoid 'unknown macro' warnings
