### 2018-01-29 2a1f371
rename warn_large_downloads to confirm_downloads_larger_than, move it to from bb_config to bb_sync
