### 2018-01-29 5f332c0
stricter check on parms
