### 2018-01-29 f3098e4
rename skip_downloads to dry_run, move from config to sync
