### 2018-01-29 e9623a5
tidy up output on dry_run
