### 2018-01-29 2190217
add section on expected user knowledge [ci-skip]
