### 2018-01-29 60ff27a
tweak error message text
