### 2018-01-29 a17e993
change http://oceancolor urls to https
