+++ tests/testthat/test_sources.R
@@ -88,8 +88,8 @@ test_that("bb_source works with multiple postprocess actions", {
         id="bilbobaggins",
         name="Oceandata test",
         description="Monthly remote-sensing sea surface temperature from the MODIS Terra satellite at 9km spatial resolution",
-        doc_url= "http://oceancolor.gsfc.nasa.gov/",
-        citation="See http://oceancolor.gsfc.nasa.gov/cms/citations",
+        doc_url= "https://oceancolor.gsfc.nasa.gov/",
+        citation="See https://oceancolor.gsfc.nasa.gov/cms/citations",
         source_url="",
         license="Please cite",
         comment="",
@@ -102,8 +102,8 @@ test_that("bb_source works with multiple postprocess actions", {
         id="bilbobaggins",
         name="Oceandata test",
         description="Monthly remote-sensing sea surface temperature from the MODIS Terra satellite at 9km spatial resolution",
-        doc_url= "http://oceancolor.gsfc.nasa.gov/",
-        citation="See http://oceancolor.gsfc.nasa.gov/cms/citations",
+        doc_url= "https://oceancolor.gsfc.nasa.gov/",
+        citation="See https://oceancolor.gsfc.nasa.gov/cms/citations",
         source_url="",
         license="Please cite",
         comment="",
