### 2018-01-29 6a5859f
add dataset name to confirmation dialog
