+++ CONTRIBUTING.md
@@ -1,0 +1,10 @@
+# Contributions
+
+Suggestions, bug reports, and code are welcome. Please [fork this repository](https://help.github.com/articles/fork-a-repo/) and submit a pull request.
+
+Requests are also welcome: open an [issue](issues).
+
+# Code of Conduct
+
+Maintainers and contributors must follow this repository's [code of conduct](CONDUCT.md).
+
