+++ CONTRIBUTING.md
@@ -2,7 +2,7 @@
 
 Suggestions, bug reports, and code are welcome. Please [fork this repository](https://help.github.com/articles/fork-a-repo/) and submit a pull request.
 
-Requests are also welcome: open an [issue](issues).
+Requests are also welcome: open an [issue](https://github.com/AustralianAntarcticDivision/bowerbird/issues).
 
 # Code of Conduct
 
