### 2018-01-30 0e09027
remove redundant logic
