### 2018-01-30 AustralianAntarcticDivision/bowerbird#15 @raymondben
Add guidance on writing new handler functions
See placeholder section in README/vignette.
