### 2018-01-30 d8b8608
remove errant trace message
