+++ R/wget.R
@@ -234,9 +234,11 @@ bb_wget <- function(url,recursive=TRUE,level=1,wait=0,accept,reject,accept_regex
             out <- list(status=status)
         }
         out$status_text <- decode_wget_exit_status(out$status)
-        if ((is.na(out$status) || out$status>0) && verbose)
+        if ((is.na(out$status) || out$status>0) && verbose) {
             ## echo the error message but don't throw it as a full blown error
+            ## often these are unimportant (e.g. 404 response codes during recursion) so we don't want to halt the entire process
             cat(sprintf(" bb_wget exited with status code %d indicating an error (%s)\n",out$status,out$status_text))
+        }
         out
     }
 }
