### 2018-01-30 90dbc93
don't use message() for sync failures if running verbosely
