### 2018-01-30 #139 @maelle
Thanks! @lwasser @MilesMcBain could you please have a look at the answer?

Reg. the `tidyverse` style, adding white spaces will make the code much easier to read and using [`styler`](https://github.com/r-lib/styler) it should be a one function call (`styler::style_pkg()`). 😉
