### 2018-01-30 3a6397b
minor reword [ci skip]
