### 2018-01-31 #139 @maelle
Thanks a lot @lwasser! Yes `styler` looks super useful (I'm yet to actually use it!)

@raymondben I forgot to tell you to add the rOpenSci review badge to the README! 🙈
