### 2018-02-07 6af3d31
don't need library(magrittr) here
