+++ tests/testthat.R
@@ -1,5 +1,4 @@
 library(testthat)
-library(magrittr)
 library(bowerbird)
 
 if (.Platform$OS.type=="windows") {
