### 2018-02-07 8b77647
fix doco typo
