+++ R/earthdata_handler.R
@@ -1,6 +1,6 @@
 #' Handler for data sets from Earthdata providers
 #'
-#' This is a handler function to be used with data sets from NASA's Earthdata system. This function is not intended to be called directly, but rather is specified as a \code{postprocess} option in \code{\link{bb_source}}.
+#' This is a handler function to be used with data sets from NASA's Earthdata system. This function is not intended to be called directly, but rather is specified as a \code{method} option in \code{\link{bb_source}}.
 #'
 #' This function uses \code{\link{bb_wget}}, and so data sources using this function will need to provide appropriate \code{\link{bb_wget}} parameters.
 #'
@@ -33,7 +33,7 @@
 #'
 #' @export
 bb_handler_earthdata <- function(...) {
-    do.call(bb_handler_earthdata_inner,list(...))
+    bb_handler_earthdata_inner(...)
 }
 
 
