### 2018-02-08 ae3fa95
fix bug in uncompressing Z files
