### 2018-02-08 28c22bf
fix bug in uncompressing Z files
