### 2018-02-08 8cf571d
pare out NA/NULL rows when subsetting cf data sources
