+++ tests/testthat/test_config.R
@@ -86,3 +86,24 @@ test_that("bb_settings works",{
     ## because the new settins list contained only the bilbobaggins element
     expect_true(length(bb_settings(cf))<1)
 })
+
+test_that("bb_subset works",{
+    temp_root <- tempdir()
+    cf <- bb_add(bb_config(temp_root),bb_example_sources())
+    temp <- bb_subset(cf,1:2)
+    expect_equal(nrow(bb_data_sources(temp)),2)
+    idx <- rep(FALSE,nrow(bb_data_sources(cf)))
+    idx[1:2] <- TRUE
+    temp <- bb_subset(cf,idx)
+    expect_equal(nrow(bb_data_sources(temp)),2)
+    idx <- rep(FALSE,nrow(bb_data_sources(cf))+1)
+    idx[1:2] <- TRUE
+    expect_warning(temp <- bb_subset(cf,idx),"Length of logical index must be") ## logical index longer than number of rows
+    expect_equal(nrow(bb_data_sources(temp)),2) ## but still get two rows back
+    temp <- bb_subset(cf,c(1:2,100)) ## two legit indices plus one dud one
+    expect_equal(nrow(bb_data_sources(temp)),2) ## dud one should be ignored
+    temp <- bb_subset(cf,-1)
+    expect_equal(nrow(bb_data_sources(temp)),5)
+    temp <- bb_subset(cf,-100)
+    expect_identical(temp,cf) ## can't drop the 100th row, nothing to do
+})
