### 2018-02-09 e9bd2c9
source_url component of data source def must be a list
