### 2018-02-12 fd37ddb
additional check on response from oceancolour file search
