### 2018-02-14 cf729ab
remove trace writes
