+++ R/oceandata.R
@@ -90,8 +90,8 @@ bb_handler_oceandata_inner <- function(config,verbose=FALSE,local_dir_only=FALSE
             qry <- paste0(qry,"&dtype=",dtype)
         ##if (get_os()=="windows") qry <- paste0("\"",qry,"\"") ## not sure if need these on windows or not!!
         myfiles <- bb_wget("https://oceandata.sci.gsfc.nasa.gov/search/file_search.cgi",recursive=FALSE,extra_flags=c("-q",qry,"-O","-"),capture_stdout=TRUE,verbose=verbose)
-   cat("search status: ",myfiles$status,"\n")
-   cat("search stdout: ",rawToChar(myfiles$stdout),"\n")
+        ##cat("search status: ",myfiles$status,"\n")
+        ##cat("search stdout: ",rawToChar(myfiles$stdout),"\n")
 
         if (myfiles$status==0) break
         tries <- tries+1
