### 2018-02-19 4d53ea4
remove old, commented-out code
