+++ R/sync.R
@@ -84,7 +84,6 @@ bb_sync <- function(config,create_root=FALSE,verbose=FALSE,catch_errors=TRUE,con
     on.exit({ restore_settings(settings) })
     ## iterate through each dataset in turn
     ## first expand the source_url list-column, so that we have one row per source_url entry
-    ## tidyr::unnest does something like this, but is unhappy with the other list-columns in the data_sources tbl
     temp <- bb_data_sources(config)
     ns <- vapply(temp$source_url,length,FUN.VALUE=1) ## number of source_url entries per row
     bb_data_sources(config) <- do.call(rbind,lapply(seq_len(nrow(temp)),function(z){ out <- temp[rep(z,ns[z]),]; out$source_url <- temp$source_url[[z]]; out}))
