+++ .travis.yml
@@ -4,9 +4,25 @@ language: R
 sudo: required
 cache: packages
 
-before_install:
-#  - sudo apt-get -qq update
-  - sudo apt-get install -y libarchive-dev libapparmor-dev
+addons:
+    apt:
+        packages:
+            - libarchive-dev
+            - libapparmor-dev
+
+matrix:
+    include:
+        - os: linux
+          r: release
+          dist: trusty # 14.04
+          env:
+           - BUILD_NAME=trusty_release
+           - DETAILS="trusty release build"
+        - os: osx
+          r: release
+          env:
+           - BUILD_NAME=osx_release
+           - DETAILS="osx release build"
 
 # set NOT_CRAN to true to ensure that skip_on_cran tests are run here
 env:
