+++ .travis.yml
@@ -5,10 +5,10 @@ sudo: required
 cache: packages
 
 addons:
-    apt:
-        packages:
-            - libarchive-dev
-            - libapparmor-dev
+  apt:
+    packages:
+      - libarchive-dev
+      - libapparmor-dev
 
 matrix:
     include:
@@ -23,6 +23,7 @@ matrix:
           env:
            - BUILD_NAME=osx_release
            - DETAILS="osx release build"
+          brew_packages: libarchive
 
 # set NOT_CRAN to true to ensure that skip_on_cran tests are run here
 env:
