### 2018-02-19 5045ca5
readme and pkgdown docs
