+++ README.Rmd
@@ -35,6 +35,7 @@ knitr::opts_chunk$set(
 <li><a href="#developer-notes">Developer notes</a></li>
 <li><a href="#data-source-summary">Data source summary</a></li>
 </ul>
+&nbsp;
 
 ## Overview
 
