### 2018-02-21 39a8a85
shorten README, point to pkgdown docs
