+++ README.Rmd
@@ -18,41 +18,15 @@ knitr::opts_chunk$set(
 [![Build Status](http://badges.herokuapp.com/travis/AustralianAntarcticDivision/bowerbird?branch=master&env=BUILD_NAME=osx_release&label=osx)](https://travis-ci.org/AustralianAntarcticDivision/bowerbird) 
 [![AppVeyor Build status](https://ci.appveyor.com/api/projects/status/5idrimyx0uuv6liu?svg=true)](https://ci.appveyor.com/project/raymondben/bowerbird)
 [![codecov](https://codecov.io/gh/AustralianAntarcticDivision/bowerbird/branch/master/graph/badge.svg)](https://codecov.io/gh/AustralianAntarcticDivision/bowerbird)
+[![](https://badges.ropensci.org/139_status.svg)](https://github.com/ropensci/onboarding/issues/139)
 
 # Bowerbird
 
 <img align="right" src="https://rawgit.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/bowerbird.svg" />
 
-### Table of contents
-
-<ul>
-<li><a href="#overview">Overview</a></li>
-<li><a href="#installing">Installing</a></li>
-<li><a href="#usage-overview">Usage overview</a></li>
-<li><a href="#users-level-of-usage-and-expected-knowledge">Users: level of usage and expected knowledge</a></li>
-<li><a href="#defining-data-sources">Defining data sources</a></li>
-<li><a href="#nuances">Nuances</a></li>
-<li><a href="#developer-notes">Developer notes</a></li>
-<li><a href="#data-source-summary">Data source summary</a></li>
-</ul>
-&nbsp;
-
-## Overview
-
 Often it's desirable to have local copies of third-party data sets. Fetching data on the fly from remote sources can be a great strategy, but for speed or other reasons it may be better to have local copies. This is particularly common in environmental and other sciences that deal with large data sets (e.g. satellite or global climate model products). Bowerbird is an R package for maintaining a local collection of data sets from a range of data providers.
 
-Bowerbird can be used in several different modes:
-
-- interactively from the R console, to download or update data files on an as-needed basis
-- from the command line, perhaps as a regular scheduled task
-- programatically, including from within other R packages, scripts, or R markdown documents that require local copies of particular data files.
-
-When might you consider using bowerbird rather than, say, [curl](https://cran.r-project.org/package=curl) or [crul](https://cran.r-project.org/package=crul)? The principal advantage of bowerbird is that it can download files recursively. In many cases, it is only necessary to specify the top-level URL, and bowerbird can recursively download linked resources. Bowerbird can also:
-
-
-- decompress downloaded files (if the remote server provides them in, say, zipped or gzipped form).
-
-- incrementally update files that you have previously downloaded. Bowerbird can be instructed not to re-download files that exist locally, unless they have changed on the remote server. Compressed files will also only be decompressed if changed.
+A comprehensive introduction to bowerbird can be found at https://australianantarcticdivision.github.io/bowerbird/articles/bowerbird.html, along with full package documentation.
 
 ## Installing
 
@@ -118,362 +92,9 @@ list.files(bb_data_source_dir(cf),recursive=TRUE,full.names=TRUE)
 
 At a later time you can re-run this synchronization process. If the remote files have not changed, and assuming that your configuration has the `clobber` parameter set to 0 ("do not overwrite existing files") or 1 ("overwrite only if the remote file is newer than the local copy") then the sync process will run more quickly because it will not need to re-download any data files.
 
-## Users: level of usage and expected knowledge
 
-Users can interact with bowerbird at several levels, with increasing levels of complexity:
+## Data source definitions
 
[…]
