+++ README.md
@@ -1,58 +1,15 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-[![Travis-CI Build Status](http://badges.herokuapp.com/travis/AustralianAntarcticDivision/bowerbird?branch=master&env=BUILD_NAME=trusty_release&label=linux)](https://travis-ci.org/AustralianAntarcticDivision/bowerbird) [![Build Status](http://badges.herokuapp.com/travis/AustralianAntarcticDivision/bowerbird?branch=master&env=BUILD_NAME=osx_release&label=osx)](https://travis-ci.org/AustralianAntarcticDivision/bowerbird) [![AppVeyor Build status](https://ci.appveyor.com/api/projects/status/5idrimyx0uuv6liu?svg=true)](https://ci.appveyor.com/project/raymondben/bowerbird) [![codecov](https://codecov.io/gh/AustralianAntarcticDivision/bowerbird/branch/master/graph/badge.svg)](https://codecov.io/gh/AustralianAntarcticDivision/bowerbird)
+[![Travis-CI Build Status](http://badges.herokuapp.com/travis/AustralianAntarcticDivision/bowerbird?branch=master&env=BUILD_NAME=trusty_release&label=linux)](https://travis-ci.org/AustralianAntarcticDivision/bowerbird) [![Build Status](http://badges.herokuapp.com/travis/AustralianAntarcticDivision/bowerbird?branch=master&env=BUILD_NAME=osx_release&label=osx)](https://travis-ci.org/AustralianAntarcticDivision/bowerbird) [![AppVeyor Build status](https://ci.appveyor.com/api/projects/status/5idrimyx0uuv6liu?svg=true)](https://ci.appveyor.com/project/raymondben/bowerbird) [![codecov](https://codecov.io/gh/AustralianAntarcticDivision/bowerbird/branch/master/graph/badge.svg)](https://codecov.io/gh/AustralianAntarcticDivision/bowerbird) [![](https://badges.ropensci.org/139_status.svg)](https://github.com/ropensci/onboarding/issues/139)
 
 Bowerbird
 =========
 
 <img align="right" src="https://rawgit.com/AustralianAntarcticDivision/bowerbird/master/inst/extdata/bowerbird.svg" />
 
-### Table of contents
-
-<ul>
-<li>
-<a href="#overview">Overview</a>
-</li>
-<li>
-<a href="#installing">Installing</a>
-</li>
-<li>
-<a href="#usage-overview">Usage overview</a>
-</li>
-<li>
-<a href="#users-level-of-usage-and-expected-knowledge">Users: level of usage and expected knowledge</a>
-</li>
-<li>
-<a href="#defining-data-sources">Defining data sources</a>
-</li>
-<li>
-<a href="#nuances">Nuances</a>
-</li>
-<li>
-<a href="#developer-notes">Developer notes</a>
-</li>
-<li>
-<a href="#data-source-summary">Data source summary</a>
-</li>
-</ul>
- 
-
-Overview
---------
-
 Often it's desirable to have local copies of third-party data sets. Fetching data on the fly from remote sources can be a great strategy, but for speed or other reasons it may be better to have local copies. This is particularly common in environmental and other sciences that deal with large data sets (e.g. satellite or global climate model products). Bowerbird is an R package for maintaining a local collection of data sets from a range of data providers.
 
-Bowerbird can be used in several different modes:
-
--   interactively from the R console, to download or update data files on an as-needed basis
--   from the command line, perhaps as a regular scheduled task
--   programatically, including from within other R packages, scripts, or R markdown documents that require local copies of particular data files.
-
[…]
