### 2018-02-22 c7a6d4a
add all_files parm to cleanup
