### 2018-02-23 975f146
fix yearly subset example
