+++ vignettes/bowerbird.Rmd
@@ -425,12 +425,14 @@ cf <- cf %>% bb_add(mysrc)
 
 Sometimes you might only want part of a data collection. Perhaps you only want a few years from a long-term collection, or perhaps the data are provided in multiple formats and you only need one. If the data source uses the `bb_handler_wget` method, you can restrict what is downloaded by modifying the arguments passed through the data source's `method` parameter, particularly the `accept`, `reject`, `accept_regex`, and `reject_regex` options. If you are modifying an existing data source configuration, you most likely want to leave the original method flags intact and just add extra flags.
 
-Say a particular data provider arranges their files in yearly directories. It would be fairly easy to restrict ourselves to, say, only the 2017 data:
+Say a particular data provider uses predictable file name patterns that include the year information. It would be fairly easy to restrict ourselves to only the 2017 data:
 
 ```{r eval=FALSE}
 library(dplyr)
-mysrc <- mysrc %>%
-  mutate(method=c(method,list(accept_regex="/2017/")))
+mysrc <- bb_example_sources() %>%
+  filter(name=="CMEMS global gridded SSH reprocessed (1993-ongoing)") %>%
+  mutate(user="yourusername",password="yourpassword")
+mysrc$method[[1]]$accept <- "*l4_2017*"
 cf <- cf %>% bb_add(mysrc)
 ```
 
