### 2018-02-24 54dd496
cope with NULL passed for mandatory parameters
