### 2018-02-24 909566a
add bb_modify_source function
