+++ R/source.R
@@ -82,9 +82,9 @@ bb_source <- function(id,name,description=NA_character_,doc_url,source_url,citat
     if (warn_empty_auth && (!is.na(authentication_note) && (na_or_empty(user) || na_or_empty(password)))) {
         warning(sprintf("The data source \"%s\" requires authentication, but the user and/or password fields have not been set.\nThe authentication_note for this data source is:\n %s",name,authentication_note))
     }
-    if (missing(license) || !is_nonempty_string(license) || missing(citation) || !is_nonempty_string(citation))
+    if (missing(license) || is.null(license) || missing(citation) || is.null(citation))
         stop("Please provide license and citation information for the data source, so that users properly acknowledge it")
-    if (missing(doc_url) || !is_nonempty_string(doc_url))
+    if (missing(doc_url) || is.null(doc_url))
         stop("Please provide a doc_url (a URL to the data source's metadata record or home page")
 
     if (missing(source_url) || is.null(source_url)) source_url <- NA_character_
@@ -127,4 +127,119 @@ bb_source <- function(id,name,description=NA_character_,doc_url,source_url,citat
         access_function=if (assert_that(is.string(access_function))) access_function,
         data_group=if (assert_that(is.string(data_group))) data_group,
         collection_size=if (assert_that(is.numeric(collection_size) || is.na(collection_size))) collection_size)
+}
+
+
+#' Modify a data source
+#'
+#' This is a helper function designed to make it easier to modify an already-defined data source. Generally, parameters passed here will replace existing entries in \code{src} if they exist, or will be added if not. The \code{method} and \code{postprocess} parameters are slightly different: see Details, below.
+#'
+#' With the exception of the \code{method} and \code{postprocess} parameters, any parameter provided here will entirely replace its equivalent in the \code{src} object. Pass a new value of \code{NULL} to remove an existing parameter.
+#'
+#' The \code{method} and \code{postprocess} parameters are lists, and modification for these takes place at the list-element level: any element of the new list will replace its equivalent element in the list in src. If the src list does not contain that element, it will be added. To illustrate, say that we have created a data source with:
+#'
+#' \code{src <- bb_source(method=list("bb_handler_wget",parm1=value1,parm2=value2),...)}
+#'
+#' Calling
+#'
+#' \code{bb_modify_source(src,method=list(parm1=newvalue1))}
+#'
+#' will result in a new \code{method} value of \code{list("bb_handler_wget",parm1=newvalue1,parm2=value2)}
+#'
+#' Modifying \code{postprocess} elements is similar. Note that it is not currently possible to entirely remove a postprocess component using this function. If you need to do so, you'll need to do it manually.
+#'
+#' @param src data.frame or tibble: a single-row data source (as returned by \code{bb_source})
+#' @param ... : parameters as for \code{bb_source}
+#'
+#' @return as for \code{bb_source}: a tibble with columns as per the \code{bb_source} function arguments (excluding \code{warn_empty_auth})
+#'
+#' @seealso \code{\link{bb_source}}
+#'
+#' @examples
+#'
+#' ## this pre-defined source requires a username and password
+#' src <- subset(bb_example_sources(),id=="10.5067/EYICLBOAAJOU")
+#'
+#' ## add username and password
+#' src <- bb_modify_source(src,user="myusername",password="mypassword")
+#'
+#' ## or using the pipe operator
+#' src <- subset(bb_example_sources(),id=="10.5067/EYICLBOAAJOU") %>%
+#'   bb_modify_source(user="myusername",password="mypassword")
+#'
+#' ## remove the existing "data_group" component
[…]
