### 2018-02-28 1083920
force TLSv1.2 in test, does it help appveyor?
