+++ tests/testthat/test_sync.R
@@ -88,8 +88,8 @@ test_that("bb_sync works with a sink() call in place",{
         doc_url= "http://some.where.org/",
         citation="blah",
         license="blah",
-        method=list("bb_handler_wget",recursive=TRUE,level=1,extra_flags="--secure-protocol=TLSv1_2"),
-        source_url="https://github.com/AustralianAntarcticDivision/bowerbird/blob/master/README.Rmd") ## just some file to download
+        method=list("bb_handler_wget",recursive=TRUE,level=1),
+        source_url="http://webdav.data.aad.gov.au/data/environmental/derived/readme.txt") ## just some file to download
     temp_root <- tempdir()
     cf <- bb_add(bb_config(local_file_root=temp_root,clobber=2),myds)
     bb_sync(cf,verbose=TRUE,confirm_downloads_larger_than=NULL)
