### 2018-02-28 aa58fc9
change badge and github URLs to point to ropensci [ci-skip]
