+++ vignettes/bowerbird.Rmd
@@ -50,7 +50,7 @@ When might you consider using bowerbird rather than, say, [curl](https://cran.r-
 ```{r eval=FALSE}
 install.packages("devtools")
 library(devtools)
-install_github("AustralianAntarcticDivision/bowerbird",build_vignettes=TRUE)
+install_github("ropensci/bowerbird",build_vignettes=TRUE)
 ```
 
 You will also need to have the third-party utility `wget` available, because bowerbird uses this to do the heavy lifting of recursively downloading files from data providers. Note that installing `wget` may require admin privileges on your local machine.
