### 2018-02-28 d2f6ceb
change badge and github URLs to point to ropensci
