### 2018-03-03 #139 @maelle
Cool, thanks! Could you also add the review badge mentioned in the checklist?

I've activated the repo in Appveyor.
