### 2018-03-03 #139 @maelle
Note that the badges do not render now, but this issue will soon be fixed, so please add it to your README before I close this issue. :-)
