### 2018-03-03 #139 @raymondben
The review badge is in the readme, just not rendering. It was showing under our org, but not now.
